% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelematrix_search.R
\name{allelematrix_search}
\alias{allelematrix_search}
\title{allelematrix search}
\usage{
allelematrix_search(markerprofileDbId = "", markerDbId = "",
  expandHomozygotes = FALSE, unknownString = "-", sepPhased = "|",
  sepUnphased = "/", format = "json", page = 0, pageSize = 10000,
  rclass = "tibble")
}
\arguments{
\item{markerprofileDbId}{character vector; default ""}

\item{markerDbId}{character vector; default ""}

\item{expandHomozygotes}{logical; default false}

\item{unknownString}{chaaracter; default: '-'}

\item{sepPhased}{character; default: '|'}

\item{sepUnphased}{character; default: '/'}

\item{format}{character; default: json; other: csv, tsv}

\item{page}{integer; default: 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{character; default: tibble}
}
\value{
data.frame
}
\description{
Gets markers in matrix format. If the format parameter is set to either csv or tsv the returned object
is always a tibble object. If the format parameter is 'json' (default) the rclass parameter can be used
to as in other functions.
}
\note{
The handling of long-running responses via asynch status messages is not yet implemented.
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/allelematrix_search}
}
\seealso{
Other brapi_call: \code{\link{attributes_categories}},
  \code{\link{attributes}}, \code{\link{calls}},
  \code{\link{crops}}, \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{locations}}, \code{\link{maps_details}},
  \code{\link{maps_positions_range}},
  \code{\link{maps_positions}}, \code{\link{maps}},
  \code{\link{marker_search}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}},
  \code{\link{observationLevels}}, \code{\link{programs}},
  \code{\link{seasons}}, \code{\link{studyTypes}},
  \code{\link{trials}}

Other genotyping: \code{\link{attributes_categories}},
  \code{\link{attributes}},
  \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{maps_details}},
  \code{\link{maps_positions_range}},
  \code{\link{maps_positions}}, \code{\link{maps}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}}
}

