% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.brapi_db.R
\name{as.brapi_db}
\alias{as.brapi_db}
\title{as.brapi_db}
\usage{
as.brapi_db(secure = FALSE, protocol = "http://", db = "127.0.0.1",
  port = 80, apipath = NULL, multicrop = FALSE, crop = "", user = "",
  password = "", token = "", granttype = "password",
  clientid = "rbrapi", bms = FALSE)
}
\arguments{
\item{secure}{logical; default: false}

\item{protocol}{character}

\item{db}{character}

\item{port}{integer; default: 80}

\item{apipath}{character; default: ''}

\item{multicrop}{logical; default: false}

\item{crop}{character}

\item{user}{character; default: ''}

\item{password}{character; default: ''}

\item{token}{character; default: ''}

\item{granttype}{character; default: 'password'}

\item{clientid}{character; default: 'rbrapi'}

\item{bms}{logical; default: false}
}
\value{
brapi_db class
}
\description{
as.brapi_db
}

