% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{attributes}
\alias{attributes}
\title{attributes}
\usage{
attributes(attributeCategoryDbId = 0, rclass = "tibble")
}
\arguments{
\item{attributeCategoryDbId}{integer; default: 0}

\item{rclass}{character; default: tibble}
}
\value{
rclass as set by parameter
}
\description{
attibutes call.
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/list-attributes-available/attributes-by-attributecategorydbid}
}

