% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes_categories.R
\name{attributes_categories}
\alias{attributes_categories}
\title{attributes_categories}
\usage{
attributes_categories(page = 0, pageSize = 10, rclass = "tibble")
}
\arguments{
\item{page}{integer; default 0}

\item{pageSize}{integer; default 10}

\item{rclass}{character; default: tibble}
}
\value{
rclass as set by parameter
}
\description{
attibutes call.
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/list-attribute-categories}
}

