% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_germplasm.R
\name{averages_germplasm}
\alias{averages_germplasm}
\title{averages_germplasm}
\usage{
averages_germplasm(study, has_trait)
}
\arguments{
\item{study}{a dataframe}

\item{has_trait}{a trait}
}
\value{
averges
}
\description{
calculates the average across repetitions for a certain trait from a fieldbook.
}

