% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_authenticate.R
\name{ba_authenticate}
\alias{ba_authenticate}
\title{Authenticate against a brapi compatible database.}
\usage{
ba_authenticate(con)
}
\arguments{
\item{con}{A required object of class 'brapi_con'.}
}
\value{
The object of class 'brapi_con' with the authentication token filled.
}
\description{
Caches a session token and informs of success or failure.
}
\author{
Reinhard Simon, Maikel Verouden
}
