% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_calls.R
\name{ba_calls}
\alias{ba_calls}
\title{ba_calls}
\usage{
ba_calls(con = NULL, datatype = "all", pageSize = 1000, page = 0,
  rclass = "tibble")
}
\arguments{
\item{con}{list, brapi connection object}

\item{datatype}{character, filter implemented brapi calls by supported data
type, e.g. "csv", "tsv" or "xls"; default: "all"}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{page}{integer, the requested page to be returend; default: 0 (1st page)}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"data.frame"}
}
\value{
An object of class as defined by rclass containing all the implemented
        BrAPI calls.
}
\description{
Returns calls available on a brapi server.
}
\details{
TODO v1.2: deprecate this and add ba_commonCropNames. The new
         function won't have datatype parameter.
}
\note{
Tested against: sweetpotatobase, test-server

BrAPI Version: 1.0, 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  # Need to connect to a database with genetic data

  # con <- ba_db()$xxxxx

  # ba_calls(con = con, ...)
}
}
\references{
\href{https://github.com/plantbreeding/API/tree/master/Specification/Calls}{github}
}
\seealso{
Other brapicore: \code{\link{ba_commoncropnames}},
  \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon, Maikel Verouden
}
