% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_can_internet.R
\name{ba_can_internet}
\alias{ba_can_internet}
\title{ba_can_internet}
\usage{
ba_can_internet(url = "www.google.org")
}
\arguments{
\item{url}{a url to check (default \href{https://www.google.org}{www.google.org})}
}
\value{
NULL when internet access is available or ERROR when internet access is not available
}
\description{
Check for internet connection.
}
\examples{
if (interactive()) {
  library(brapi)

  # Checks against a well known and high uptime web site. Default: google

  ba_can_internet()

  # or use an alternative

  ba_can_internet("www.amazon.com")

}
}
\references{
\href{http://stackoverflow.com/questions/5076593/how-to-determine-if-you-have-an-internet-connection-in-r}{Stack Overflow} and  \code{\link[curl]{has_internet}}
}
\seealso{
Other brapiutils: \code{\link{as.ba_db}},
  \code{\link{ba_chart.ba_genomemaps_details}},
  \code{\link{ba_chart.ba_genomemaps}},
  \code{\link{ba_chart.ba_locations}},
  \code{\link{ba_chart}}, \code{\link{ba_check}},
  \code{\link{ba_connect}}, \code{\link{ba_db}},
  \code{\link{ba_describe.ba_locations}},
  \code{\link{ba_describe}}, \code{\link{ba_login}},
  \code{\link{ba_logout}}, \code{\link{is.ba_con}},
  \code{\link{print.ba_con}},
  \code{\link{print.ba_db_list}}
}

