% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_chart-methods.R
\name{ba_chart.ba_genomemaps}
\alias{ba_chart.ba_genomemaps}
\title{ba_chart.ba_genomemaps}
\usage{
\method{ba_chart}{ba_genomemaps}(x, ...)
}
\arguments{
\item{x}{a brapi_locations data.frame/tibble or a numeric vector of longitudes}

\item{...}{other plot parameters}
}
\description{
print method for an object of class brapi_brapi_genomemaps, which will only display the crop,  database address:port and user
}
\examples{
if (interactive()) {
  library(brapi)

  # Need to connect to a database with genetic data

  # con <- ba_db()$xxxxx

  # gm <- ba_genome_maps(con = con, ...)

  # chart(gm)
}
}
\seealso{
Other brapiutils: \code{\link{as.ba_db}},
  \code{\link{ba_can_internet}},
  \code{\link{ba_chart.ba_genomemaps_details}},
  \code{\link{ba_chart.ba_locations}},
  \code{\link{ba_chart}}, \code{\link{ba_check}},
  \code{\link{ba_connect}}, \code{\link{ba_db}},
  \code{\link{ba_describe.ba_locations}},
  \code{\link{ba_describe}}, \code{\link{ba_login}},
  \code{\link{ba_logout}}, \code{\link{is.ba_con}},
  \code{\link{print.ba_con}},
  \code{\link{print.ba_db_list}}
}
\author{
Reinhard Simon
}
