% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_check.R
\name{ba_check}
\alias{ba_check}
\title{ba_check}
\usage{
ba_check(con = NULL, verbose = TRUE, brapi_calls = "any")
}
\arguments{
\item{con}{brapi_connection object}

\item{verbose}{logical; default TRUE}

\item{brapi_calls}{character vector; default: any. Use to check if one or more calls are implemented by the server according to the calls url.}
}
\value{
logical
}
\description{
Checks if a BrAPI server can be reached given the connection details.
}
\details{
Raises errors.
}
\examples{
if (interactive()) {
  library(brapi)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  ba_check(con)
}
}
\author{
Reinhard Simon
}
