% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_commoncropnames.R
\name{ba_commoncropnames}
\alias{ba_commoncropnames}
\title{ba_commoncropnames}
\usage{
ba_commoncropnames(con = NULL, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{rclass}{character; default: "tibble" possible other values: "vector"/"json"/"list"/"data.frame"}
}
\value{
a vector of crop names or NULL
}
\description{
Lists supported crops in a database
}
\note{
Tested against: test-server

BrAPI Version: 1.2, 1.3

BrAPI Status:
}
\examples{
if (interactive()) {
  library(brapi)

  # Need to connect to a database with genetic data

  # con <- ba_db()$xxxxx

  # ba_crops(con = con, ...)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Crops/CommonCropNames_GET.md}{github}
}
\seealso{
Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon, Maikel Verouden
}
