% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_crops.R
\name{ba_crops-deprecated}
\alias{ba_crops-deprecated}
\alias{ba_crops}
\title{ba_crops}
\usage{
ba_crops(con = NULL, rclass = "tibble")
}
\arguments{
\item{con}{list, brapi connection object}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"vector"/"data.frame"}
}
\value{
an object of class as specified by rclass with crop names or NULL
}
\description{
List supported crops in a database.
}
\note{
Tested against: BMS, sweetpotatobase, test-server

BrAPI Version: 1.0, 1.1

BrAPI Status: deprecated
}
\examples{
if (interactive()) {
  library(brapi)

  con <-  ba_db()$testserver

  ba_crops(con = con)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Crops/ListCrops_Deprecated.md}{github}
}
\seealso{
\code{\link{brapi-deprecated}}

Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}},
  \code{\link{brapi-deprecated}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{brapicore}
\keyword{internal}
