% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_db.R
\name{ba_db}
\alias{ba_db}
\title{ba_db}
\usage{
ba_db()
}
\value{
list
}
\description{
A white list of known brapi databases.
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)

  # make sure brapiTS::mock_server() is running in a separate process

  bdb <- ba_db()

  print(bdb)

  bdb$mockbase \%>\% print

}
}
\author{
Reinhard Simon
}
