% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ba_describe}
\alias{ba_describe}
\title{ba_describe}
\usage{
ba_describe(x, ...)
}
\arguments{
\item{x}{an object to be described}

\item{...}{additional parameters}
}
\description{
A generic method to sum up brapi objects
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  lcs <- ba_locations(con)
  lcs \%>\% ba_describe
}
}
\seealso{
Other brapiutils: \code{\link{as.ba_db}},
  \code{\link{ba_can_internet}},
  \code{\link{ba_chart.ba_genomemaps_details}},
  \code{\link{ba_chart.ba_genomemaps}},
  \code{\link{ba_chart.ba_locations}},
  \code{\link{ba_chart}}, \code{\link{ba_check}},
  \code{\link{ba_connect}}, \code{\link{ba_db}},
  \code{\link{ba_describe.ba_locations}},
  \code{\link{ba_login}}, \code{\link{ba_logout}},
  \code{\link{is.ba_con}}, \code{\link{print.ba_con}},
  \code{\link{print.ba_db_list}}
}
\author{
Reinhard Simon
}
