% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_genomemaps.R
\name{ba_genomemaps}
\alias{ba_genomemaps}
\title{ba_genomemaps}
\usage{
ba_genomemaps(con = NULL, species = "all", type = "all", page = 0,
  pageSize = 30, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{species}{string; default:}

\item{type}{string, default: all}

\item{page}{integer; default 0}

\item{pageSize}{integer; default 30}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists genomemaps available on a brapi server
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  ba_genomemaps(con)

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/GenomeMaps/ListOfGenomeMaps.md}{github}
}
\seealso{
Other genomemaps: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_germplasmattributes}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search}}
}
\author{
Reinhard Simon
}
