% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_genomemaps_data_range.R
\name{ba_genomemaps_data_range}
\alias{ba_genomemaps_data_range}
\title{ba_genomemaps_data_range}
\usage{
ba_genomemaps_data_range(con = NULL, mapDbId = "",
  linkageGroupId = NULL, linkageGroupName = "", min = 0,
  max = 2000, pageSize = 1000, page = 0, rclass = c("tibble",
  "data.frame", "list", "json"))
}
\arguments{
\item{con}{brapi connection object}

\item{mapDbId}{character; default ''}

\item{linkageGroupId}{character; default NULL}

\item{linkageGroupName}{character; default ''}

\item{min}{character; default ''}

\item{max}{character; default ''}

\item{pageSize}{character; default 30}

\item{page}{integer; default 0}

\item{rclass}{character; default: tibble}
}
\value{
rclass as defined
}
\description{
Get map data by range on linkageGroup
}
\details{
markers ordered by linkageGroup and position
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_genomemaps_data_range(con, "gm1", linkageGroupName = "1")
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/GenomeMaps/GenomeMapDataByRangeOnLinkageGroup.md}{github}
}
\seealso{
Other genomemaps: \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}}

Other genotyping: \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_attributes}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search_post}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_attributes}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_markerprofiles_allelematrices_details}},
  \code{\link{ba_markerprofiles_allelematrices_search}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search_post}},
  \code{\link{ba_markers_search}}
}
\author{
Reinhard Simon
}
\concept{genomemaps}
\concept{genotyping}
