% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_genomemaps_details.R
\name{ba_genomemaps_details}
\alias{ba_genomemaps_details}
\title{ba_genomemaps_details}
\usage{
ba_genomemaps_details(con = NULL, mapDbId = "1", rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{mapDbId}{character; default 1}

\item{rclass}{character; default: tibble}
}
\value{
rclass as defined
}
\description{
lists genomemaps_details available on a brapi server
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_genomemaps_details(con, "gm1")
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/GenomeMaps/GenomeMapDetails.md}{github}
}
\seealso{
Other genomemaps: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_attributes}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_germplasmattributes}},
  \code{\link{ba_markerprofiles_allelematrices_search}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search}}
}
\author{
Reinhard Simon
}
