% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_germplasm_breeding_methods.R
\name{ba_germplasm_breedingmethods}
\alias{ba_germplasm_breedingmethods}
\title{ba_germplasm_breedingmethods}
\usage{
ba_germplasm_breedingmethods(con = NULL, pageSize = 1000, page = 0,
  rclass = "tibble")
}
\arguments{
\item{con}{list, brapi connection object}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"data.frame"}
}
\value{
An object of class as defined by rclass containing breedingmethods.
}
\description{
Lists breeding methods available on a brapi server.
}
\note{
Tested against: test-server

BrAPI Version: 1.2

BrAPI Status: active
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Germplasm/BreedingMethods_GET.yaml}{github}
}
\author{
Reinhard Simon, Maikel Verouden
}
