% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_germplasm_breedingmethods_details.R
\name{ba_germplasm_breedingmethods_details}
\alias{ba_germplasm_breedingmethods_details}
\title{ba_germplasm_breedingmethods_details}
\usage{
ba_germplasm_breedingmethods_details(con = NULL, breedingMethodDbId = "",
  rclass = "tibble")
}
\arguments{
\item{con}{list, brapi connection object}

\item{breedingMethodDbId}{character}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"data.frame"}
}
\value{
An object of class as defined by rclass containing breedingmethods.
}
\description{
Lists breeding methods available on a brapi server.
}
\note{
Tested against: test-server

BrAPI Version: 1.2

BrAPI Status: active
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Germplasm/BreedingMethods_BreedingMethodDbId_GET.md}{github}
}
\author{
Reinhard Simon, Maikel Verouden
}
