% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_germplasm_markerprofiles.R
\name{ba_germplasm_markerprofiles}
\alias{ba_germplasm_markerprofiles}
\title{ba_germplasm_markerprofiles}
\usage{
ba_germplasm_markerprofiles(con = NULL, germplasmDbId = "",
  rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{germplasmDbId}{character}

\item{rclass}{character, default: list; alternative: vector}
}
\value{
list of marker profile ids
}
\description{
Gets minimal marker profile data from database using database internal id
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_germplasm_markerprofiles(con, germplasmDbId = "1")

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Germplasm/GermplasmMarkerprofiles.md}{github}
}
\seealso{
Other germplasm: \code{\link{ba_germplasm_details_study-deprecated}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search}},
  \code{\link{brapi-deprecated}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_attributes}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_germplasmattributes}},
  \code{\link{ba_markerprofiles_allelematrices_search}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search}}
}
\author{
Reinhard Simon
}
\concept{genotyping}
\concept{germplasm}
