% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_germplasm_progeny.R
\name{ba_germplasm_progeny}
\alias{ba_germplasm_progeny}
\title{ba_germplasm_progeny}
\usage{
ba_germplasm_progeny(con = NULL, germplasmDbId = "",
  rclass = c("tibble", "data.frame", "list", "json"))
}
\arguments{
\item{con}{brapi connection object}

\item{germplasmDbId}{character; \strong{REQUIRED ARGUMENT} with default ''}

\item{rclass}{character; default: tibble}
}
\value{
character; default: tibble
}
\description{
Gets germplasm progenies given an id.
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_germplasm_progeny(con, germplasmDbId = "1")

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Germplasm/Germplasm_Progeny_GET.md}{github}
}
\seealso{
Other germplasm: \code{\link{ba_germplasm_breedingmethods}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_studies_germplasm_details}}

Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_commoncropnames}}, \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_attributes}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_studies_germplasm_details}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon
}
\concept{brapicore}
\concept{germplasm}
