% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_germplasm_search.R
\name{ba_germplasm_search}
\alias{ba_germplasm_search}
\title{ba_germplasm_search}
\usage{
ba_germplasm_search(con = NULL, germplasmPUI = "",
  germplasmDbId = "", germplasmName = "", commonCropName = "",
  pageSize = 1000, page = 0, rclass = c("tibble", "data.frame",
  "list", "json"))
}
\arguments{
\item{con}{brapi connection object}

\item{germplasmPUI}{character, search for germplasm information using a
Permanent Unique Identifier (DOI, URI, etc.); default: ""}

\item{germplasmDbId}{character, search for germplasm information using an
internal germplasm database identifier; default: ""}

\item{germplasmName}{character, search for germplasm information using the
name of the germplasm; default: ""}

\item{commonCropName}{character, search for germplasm information using the
common crop name related to the germplasm; default: ""}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "data.frame"/"list"/"json"}
}
\value{
An object of class as defined by rclass containing the germaplasm
        information result of the search.
}
\description{
Retrieve germplasm information as result of a GET search.
}
\note{
Tested against: test-server

BrAPI Version: 1.0, 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_germplasm_search(con, "1")


}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Germplasm/GermplasmSearch_GET.md}{github}
}
\seealso{
Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_commoncropnames}}, \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search_post}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search_post}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search_post}},
  \code{\link{ba_programs_search_post}},
  \code{\link{ba_programs}},
  \code{\link{ba_studies_germplasm_details}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_attributes}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search_post}},
  \code{\link{ba_germplasmattributes_attributes}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_markerprofiles_allelematrices_details}},
  \code{\link{ba_markerprofiles_allelematrices_search_post}},
  \code{\link{ba_markerprofiles_allelematrices_search}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search_post}},
  \code{\link{ba_markers_search}}

Other germplasm: \code{\link{ba_germplasm_breedingmethods}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search_post}},
  \code{\link{ba_studies_germplasm_details}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{brapicore}
\concept{genotyping}
\concept{germplasm}
