% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_germplasmattributes.R
\name{ba_germplasmattributes}
\alias{ba_germplasmattributes}
\title{ba_germplasmattributes}
\usage{
ba_germplasmattributes(con = NULL, attributeCategoryDbId = "0",
  rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{attributeCategoryDbId}{integer; default: 0}

\item{rclass}{character; default: tibble}
}
\value{
rclass as set by parameter
}
\description{
germplasmattributes call.
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  ba_germplasmattributes(con)

}
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/GermplasmAttributes/ListAttributesByAttributeCategoryDbId.md}
}
\seealso{
Other germplasmattributes: \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search}}
}
\author{
Reinhard Simon
}
