% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_germplasmattributes_details.R
\name{ba_germplasmattributes_details}
\alias{ba_germplasmattributes_details}
\title{ba_germplasmattributes_details}
\usage{
ba_germplasmattributes_details(con = NULL, germplasmDbId = "1",
  attributeList = "1", page = 0, pageSize = 10, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{germplasmDbId}{character; default: 1.}

\item{attributeList}{character vector; default: 1.}

\item{page}{integer; default: 0.}

\item{pageSize}{integer; default: 10.}

\item{rclass}{character; default: tibble; or else: json, list, data.frame.}
}
\value{
rclass as set by parameter.
}
\description{
attibutes call.
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  ba_germplasmattributes_details(con)

}
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/GermplasmAttributes/GermplasmAttributeValuesByGermplasmDbId.md}
}
\seealso{
Other germplasmattributes: \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search}}
}
\author{
Reinhard Simon
}
