% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_locations.R
\name{ba_locations}
\alias{ba_locations}
\title{ba_locations}
\usage{
ba_locations(con = NULL, locationType = "all", page = 0,
  pageSize = 1e+06, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{locationType}{string, list of data types}

\item{page}{integer; default 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists locations available on a brapi server
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)
  library(maps)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  # Example 1
  bl <- ba_locations()
  bl \%>\% chart()

  # Example 2
  bl \%>\% chart(chart_type = "map")
}
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Locations/ListLocations.md}(github)
}
\author{
Reinhard Simon
}
