% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_login.R
\name{ba_login}
\alias{ba_login}
\title{ba_login}
\usage{
ba_login(con)
}
\arguments{
\item{con}{list, brapi connection object}
}
\value{
The list object of class 'brapi_con' with the authentication token filled.
}
\description{
Login into a brapi compatible database.
}
\details{
Caches a session token and informs of success or failure.
}
\note{
Tested against: BMS, sweetpotatobase

BrAPI Version: 1.0, 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  #con <- ba_login(con)

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Authentication/Authentication.md}{github}
}
\seealso{
Other brapiutils: \code{\link{as.ba_db}},
  \code{\link{ba_can_internet}}, \code{\link{ba_check}},
  \code{\link{ba_connect}}, \code{\link{ba_db}},
  \code{\link{ba_logout}}, \code{\link{is.ba_con}},
  \code{\link{print.ba_con}},
  \code{\link{print.ba_db_list}}
}
\author{
Reinhard Simon, Maikel Verouden
}
