% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_logout.R
\name{ba_logout}
\alias{ba_logout}
\title{ba_logout}
\usage{
ba_logout(con)
}
\arguments{
\item{con}{A required object of class 'brapi_con'.}
}
\value{
The object of class 'brapi_con' with the authentication token removed.
}
\description{
Logout from a brapi compatible database.
}
\examples{
if (interactive()) {
  library(brapi)
  # Need to connect to a database with genetic data

  # con <- ba_db()$xxxxx

  # ba_logout(con = con)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Authentication/Authentication.md}{github}
}
\seealso{
Other brapiutils: \code{\link{as.ba_db}},
  \code{\link{ba_can_internet}}, \code{\link{ba_check}},
  \code{\link{ba_connect}}, \code{\link{ba_db}},
  \code{\link{ba_login}}, \code{\link{is.ba_con}},
  \code{\link{print.ba_con}},
  \code{\link{print.ba_db_list}}
}
\author{
Reinhard Simon
}
