% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_markerprofiles_allelematrix_search.R
\name{ba_markerprofiles_allelematrix_search}
\alias{ba_markerprofiles_allelematrix_search}
\title{ba_markerprofiles_allelematrix_search}
\usage{
ba_markerprofiles_allelematrix_search(con = NULL,
  markerprofileDbId = "", markerDbId = "", expandHomozygotes = FALSE,
  unknownString = "-", sepPhased = "|", sepUnphased = "/",
  format = "json", page = 0, pageSize = 10000, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{markerprofileDbId}{character vector; default ''}

\item{markerDbId}{character vector; default ''}

\item{expandHomozygotes}{logical; default false}

\item{unknownString}{chaaracter; default: '-'}

\item{sepPhased}{character; default: '|'}

\item{sepUnphased}{character; default: '/'}

\item{format}{character; default: json; other: csv, tsv}

\item{page}{integer; default: 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{character; default: tibble}
}
\value{
data.frame
}
\description{
Gets markers in matrix format. If the format parameter is set to either csv or tsv the returned object
is always a tibble object. If the format parameter is 'json' (default) the rclass parameter can be used
to as in other functions.
}
\note{
The handling of long-running responses via asynch status messages is not yet implemented.
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_markerprofiles_allelematrix_search(con)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/MarkerProfiles/MarkerProfileAlleleMatrix.md}{github}
}
\seealso{
Other markerprofiles: \code{\link{ba_markerprofiles_allelematrices_details}},
  \code{\link{ba_markerprofiles_allelematrices_search_post}},
  \code{\link{ba_markerprofiles_allelematrices_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_attributes}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search_post}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_attributes}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_markerprofiles_allelematrices_details}},
  \code{\link{ba_markerprofiles_allelematrices_search_post}},
  \code{\link{ba_markerprofiles_allelematrices_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search_post}},
  \code{\link{ba_markers_search}}
}
\author{
Reinhard Simon
}
\concept{genotyping}
\concept{markerprofiles}
