% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_markerprofiles_details.R
\name{ba_markerprofiles_details}
\alias{ba_markerprofiles_details}
\title{ba_markerprofiles_details}
\usage{
ba_markerprofiles_details(con = NULL, markerprofileDbId = "",
  expandHomozygotes = FALSE, unknownString = "", sepPhased = "",
  sepUnphased = "", page = 0, pageSize = 10000,
  rclass = c("tibble", "data.frame", "list", "json"))
}
\arguments{
\item{con}{brapi connection object}

\item{markerprofileDbId}{character; default 0}

\item{expandHomozygotes}{logical; default false}

\item{unknownString}{chaaracter; default: ''}

\item{sepPhased}{character; default: ''}

\item{sepUnphased}{character; default: ''}

\item{page}{integer; default: 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{character; default: tibble}
}
\value{
rclass as requested
}
\description{
Lists markers as result of a search.
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_markerprofiles_search(con, germplasmDbId = "1")
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/MarkerProfiles/MarkerProfiles_MarkerProfileDbId_GET.md}{github}
}
\seealso{
Other markerprofiles: \code{\link{ba_markerprofiles_allelematrices_details}},
  \code{\link{ba_markerprofiles_allelematrices_search_post}},
  \code{\link{ba_markerprofiles_allelematrices_search}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_search}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_attributes}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search_post}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_attributes}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_markerprofiles_allelematrices_details}},
  \code{\link{ba_markerprofiles_allelematrices_search_post}},
  \code{\link{ba_markerprofiles_allelematrices_search}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search_post}},
  \code{\link{ba_markers_search}}
}
\author{
Reinhard Simon
}
\concept{genotyping}
\concept{markerprofiles}
