% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_markerprofiles_search.R
\name{ba_markerprofiles_search}
\alias{ba_markerprofiles_search}
\title{ba_markerprofiles_search}
\usage{
ba_markerprofiles_search(con = NULL, germplasmDbId = "", studyDbId = "",
  sampleDbId = "", extractDbId = "", methodDbId = "all", page = 0,
  pageSize = 10000, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{germplasmDbId}{character; default ""}

\item{studyDbId}{character; default ""}

\item{sampleDbId}{character; default: ""}

\item{extractDbId}{character; default: ""}

\item{methodDbId}{string; default: unknown}

\item{page}{integer; default: 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{character; default: tibble}
}
\value{
data.frame
}
\description{
Lists markers as result of a search.
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  ba_markerprofiles_search(con, "3")

}
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/MarkerProfiles/MarkerProfileSearch.md}
}
\seealso{
Other markerprofiles: \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_germplasmattributes}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search}}
}
\author{
Reinhard Simon
}
