% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_markers_search.R
\name{ba_markers_search}
\alias{ba_markers_search}
\title{ba_markers_search}
\usage{
ba_markers_search(con = NULL, markerDbIds = "", name = "",
  matchMethod = "exact", includeSynonyms = NA, type = "",
  pageSize = 1000, page = 0, rclass = c("tibble", "data.frame",
  "list", "json"))
}
\arguments{
\item{con}{list, brapi connection object}

\item{markerDbIds}{character vector, search for specific markers by internal
database identifiers, supplied as a comma separated
character vector of internal marker database identifiers
e.g. c("1185","1186"); default: ""}

\item{name}{character, search for markers by a search pattern of a marker
name or synonym e.g. "11_1002", "11_1%", "11_1*" or "11_10?02";
default: ""}

\item{matchMethod}{character, search for markers using a match method with as
possible values "case_insensitive", "exact" (which is case
sensitive), and "wildcard" (whis is case insensitive and
uses both "*" and "%" for any number of characters adn "?"
for one character matching); default: "exact"}

\item{includeSynonyms}{logical, search for markers including synonyms,
specified as TRUE, or without, specified as FALSE;
default: NA}

\item{type}{character, search for a specific type of marker, e.g. "SNP";
default: ""}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{rclass}{character, class of the object to be returned; default: "tibble"
, possible other values: "json"/"list"/"vector"/"data.frame"}

\item{active}{logical; search studies by active status (by default both
active and inactive studies are shown); default: NA, other
possible values TRUE (show active)/FALSE (show inactive)}
}
\value{
An object of class as defined by rclass containing the markers
        fulfilling the search criteria.
}
\description{
Search for markers on a brapi server via a GET method.
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_markers_search(con)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Markers/MarkerSearch_GET.md}{github}
}
\seealso{
Other markers: \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search_post}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_attributes}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search_post}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_attributes}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_markerprofiles_allelematrices_details}},
  \code{\link{ba_markerprofiles_allelematrices_search_post}},
  \code{\link{ba_markerprofiles_allelematrices_search}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search_post}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{genotyping}
\concept{markers}
