% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_observationvariables_search.R
\name{ba_observationvariables_search}
\alias{ba_observationvariables_search}
\title{ba_observationvariables_search}
\usage{
ba_observationvariables_search(con = NULL, observationVariableDbIds = "any",
  ontologyXrefs = "any", ontologyDbIds = "any", methodDbIds = "any",
  scaleDbIds = "any", names = "any", datatypes = "any",
  traitClasses = "any", rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{observationVariableDbIds}{vector of string; default: any}

\item{ontologyXrefs}{vector of string; default: any}

\item{ontologyDbIds}{vector of string; default: any}

\item{methodDbIds}{vector of string; default: any}

\item{scaleDbIds}{vector of string; default: any}

\item{names}{vector of string; default: any}

\item{datatypes}{vector of string; default: any}

\item{traitClasses}{vector of string; default: any}

\item{rclass}{string; default: tibble}
}
\value{
rclass
}
\description{
lists the breeding observationvariables
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  #TODO

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/ObservationVariables/VariableSearch.md}{github}
}
\seealso{
Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon
}
