% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_observationvariables_search_post.R
\name{ba_observationvariables_search_post}
\alias{ba_observationvariables_search_post}
\title{ba_observationvariables_search_post}
\usage{
ba_observationvariables_search_post(con = NULL, datatypes = "",
  methodDbIds = "", names = "", observationVariableDbIds = "",
  ontologyXrefs = "", ontologyDbIds = "", scaleDbIds = "",
  traitClasses = "", pageSize = 1000, page = 0,
  rclass = c("tibble", "data.frame", "list", "json"))
}
\arguments{
\item{con}{list, brapi connection object}

\item{datatypes}{character vector, filter the search results by a list of
scale data types, supplied as a comma separated character
vector e.g. c("numeric"); default: ""}

\item{methodDbIds}{character vector, filter the search results by a list of
methods, supplied as a comma separated character vector
of internal program database identifiers e.g.
c("method-1", "method-2"); default: ""}

\item{names}{character vector, search for observation variables by human
readable observation variable names, supplied as a comma
separated character vector e.g. c("caro_spectro"); default: ""}

\item{observationVariableDbIds}{character vector, filter search results of
observation variables by specified
observation variable internal database
identifiers, supplied as a comma separated
character vector e.g. c("obs-variable-id1",
"obs-variable-id1"); default: ""}

\item{ontologyXrefs}{character vector, filter search results of observation
variables using ontology cross references, supplied as
a comma separated character vector e.g. c("CO:123",
"CO:456"); default: ""}

\item{ontologyDbIds}{character vector, filter search results of observation
variables by ontology internal database identifiers,
supplied as a comma separated character vector e.g. c("
CO_334:0100632"); default: ""}

\item{scaleDbIds}{chararcter vector, filter search results of observation
variables by specified scale internal database identifiers
, supplied as a comma separated character vector e.g. c("
scale-1", "scale-2"); default: ""}

\item{traitClasses}{character vector, filter search results of observation
variables by trait classes, supplied as a comma separated
character vector e.g. c("Phenological", "Physiological");
default: ""}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"data.frame"}
}
\value{
An object of class as defined by \code{rclass} containing the
        observation variables fulfilling the search criteria.
}
\description{
Search for observation variables on a brapi server via a POST method.
}
\note{
Tested against: test-server

BrAPI Version: 1.0, 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_observationvariables_search_post(con)

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/ObservationVariables/VariableSearch.md}{github}
}
\author{
Reinhard Simon, Maikel Verouden
}
