% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_phenotypes_search.R
\name{ba_phenotypes_search}
\alias{ba_phenotypes_search}
\title{ba_phenotypes_search}
\usage{
ba_phenotypes_search(con = NULL, germplasmDbId = "",
  observationVariableDbId = "", studyDbId = "", locationDbId = "",
  trialDbId = "", programDbId = "", seasonDbId = "",
  observationLevel = "", observationTimeStampRangeStart = "",
  observationTimeStampRangeEnd = "", pageSize = 1000, page = 0,
  rclass = c("tibble", "data.frame", "list", "json"))
}
\arguments{
\item{con}{list, brapi connection object}

\item{germplasmDbId}{character, search for observation units where the
specified germplasm has been used/tested; e.g. "CML123";
default: ""}

\item{observationVariableDbId}{character, search for observation units where
the specified observation variable, supplied as
a character string of the internal observation
variable database identifier e.g. "CO-PH-123",
has been measured; default: ""}

\item{studyDbId}{character, search for observation units by an internal study
database identifier e.g. "1001"; default: ""}

\item{locationDbId}{character, search for observation units with the observed
Phenotypes by an internal location database identifier;
default: ""}

\item{trialDbId}{character, search for observation units with the observed
Phenotypes by an internal trial database identifier;
default: ""}

\item{programDbId}{character, search for observation units with the observed
Phenotypes by an internal program database identifier;
default: ""}

\item{seasonDbId}{character, search for observation units with the observed
Phenotypes by an internal season database identifier;
default: ""}

\item{observationLevel}{character, search for observation units with the
observed Phenotypes by specifying observation level
e.g. "plant"/"plot"; default: ""}

\item{observationTimeStampRangeStart}{chararcter, specifying the time stamp
range start from which to search for the
observation units with the observed
Phenotypes; default: ""}

\item{observationTimeStampRangeEnd}{character, specifying the time stamp range
end to which to search for the observation
units with the observed Phenotypes;
default: ""}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"data.frame"}
}
\value{
An object of class as defined by rclass containing the observation
        units with the observed Phenotypes.
}
\description{
Retrieve a list of observation units with the observed Phenotypes using a GET
method.
}
\note{
Tested against: test-server

BrAPI Version: 1.2

BrAPI Status: active
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Phenotypes/PhenotypesSearch_GET.md}{github}
}
\seealso{
Other studies: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_germplasm_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}}

Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples_search_post}},
  \code{\link{ba_samples_search}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{phenotyping}
\concept{studies}
