% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_programs.R
\name{ba_programs}
\alias{ba_programs}
\title{ba_programs}
\usage{
ba_programs(con = NULL, programName = "any", abbreviation = "any",
  page = 0, pageSize = 10000, rclass = "tibble")
}
\arguments{
\item{con}{list, brapi connection object}

\item{programName}{character; default: any}

\item{abbreviation}{character; default: any}

\item{page}{integer requested page number, default = 0 (1st page)}

\item{pageSize}{integer, items per page (default = 100)}

\item{rclass}{character, default: tibble}
}
\value{
rclass
}
\description{
lists the breeding programs
}
\details{
BRAPI discussion: Should this return also the crop?
}
\examples{
if (interactive()) {
  library(brapi)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  ba_programs(con)

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Programs/ListPrograms.md}{github}
}
\seealso{
Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon
}
