% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_programs_search_post.R
\name{ba_programs_search_post}
\alias{ba_programs_search_post}
\title{ba_programs_search_post}
\usage{
ba_programs_search_post(con = NULL, abbreviation = "",
  leadPerson = "", name = "", objective = "", programDbId = "",
  commonCropName = "", pageSize = 1000, page = 0,
  rclass = c("tibble", "data.frame", "list", "json"))
}
\arguments{
\item{con}{list, brapi connection object}

\item{abbreviation}{character, search programs by program abbreviation;
default: ""}

\item{leadPerson}{character, search programs by program leader name; default:
""}

\item{name}{character, search programs by program name; default: ""}

\item{objective}{character, search programs by program objective; default: ""}

\item{programDbId}{character, search programs by program database identifier;
default: ""}

\item{commonCropName}{character, search programs by commonCropName; default:
""}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"data.frame"}
}
\value{
An object of class as defined by rclass containing the breeding
        programs matching the search criteria.
}
\description{
Search breeding programs via a POST method
}
\details{
Searching by commonCropName has not be tested yet, because none of
         the databases support this feature yet (not even the test-server)!
}
\note{
Tested against: test-server

BrAPI Version: 1.0, 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_programs_search_post(con)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Programs/ProgramSearch.md}{github}
}
\seealso{
Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_commoncropnames}}, \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search_post}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search_post}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search_post}},
  \code{\link{ba_programs}},
  \code{\link{ba_studies_germplasm_details}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{brapicore}
