% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_samples_search_post.R
\name{ba_samples_search}
\alias{ba_samples_search}
\title{ba_samples_search}
\usage{
ba_samples_search(con = NULL, sampleDbId = "",
  observationUnitDbId = "", plateDbId = "", germplasmDbId = "",
  pageSize = 1000, page = 0, rclass = c("tibble", "data.frame",
  "list", "json"))
}
\arguments{
\item{con}{brapi connection object}

\item{sampleDbId}{character, default ''}

\item{observationUnitDbId}{character, default ''}

\item{plateDbId}{character, default ''}

\item{germplasmDbId}{character, default ''}

\item{pageSize}{integer,default 1000}

\item{page}{integer, default 0}

\item{rclass}{character; default: "tibble" possible other values: "json"/"list"/"data.frame"}
}
\value{
rclass as defined
}
\description{
Used to retrieve list of Samples from a Sample Tracking system based on some search criteria.
}
\note{
Tested against: BMS
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_samples(con, sampleDbId = "sam1")
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Samples/SampleSearch_GET.md}{github}
}
\seealso{
Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_get}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{phenotyping}
