% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_expiration.R
\name{ba_show_expiration}
\alias{ba_show_expiration}
\title{ba_show_expiration}
\usage{
ba_show_expiration(con)
}
\arguments{
\item{con}{brapi connection}
}
\description{
ba_show_expiration
}
\examples{

library(brapi)

if (interactive()) {
  con <- ba_db()$testserver
  ba_show_expiration(con)
}
}
