% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_germplasm_details.R
\name{ba_studies_germplasm_details}
\alias{ba_studies_germplasm_details}
\title{ba_studies_germplasm_details}
\usage{
ba_studies_germplasm_details(con = NULL, studyDbId = "", page = 0,
  pageSize = 1000, rclass = "tibble")
}
\arguments{
\item{con}{list, brapi connection object}

\item{studyDbId}{character, the internal database identifier for a study of
which the germplasm details are to be retrieved e.g. "1001";
\strong{REQUIRED ARGUMENT} with default: ""}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"data.frame"}
}
\value{
An object of class as defined by rclass containing the germplasm
        details for a requested study.
}
\description{
Gets germplasm details for a specific study available on a brapi server.
}
\note{
Tested against: sweetpotatobase, test-server

BrAPI Version: 1.0, 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_studies_germplasm_details(con = con, studyDbId = "1001")
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Studies/Studies_Germplasm_GET.md}{github}
}
\seealso{
Other studies: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_get}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}}

Other germplasm: \code{\link{ba_germplasm_details_study-deprecated}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search}}

Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_commoncropnames}},
  \code{\link{ba_crops-deprecated}},
  \code{\link{ba_germplasm_details_study-deprecated}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{brapicore}
\concept{germplasm}
\concept{studies}
