% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_observationlevels.R
\name{ba_studies_observationlevels}
\alias{ba_studies_observationlevels}
\title{ba_studies_observationlevels}
\usage{
ba_studies_observationlevels(con = NULL, rclass = "vector")
}
\arguments{
\item{con}{brapi connection object}

\item{rclass}{character; default is FALSE; whether to display the raw list object or not}
}
\value{
a vector of crop names or NULL
}
\description{
lists studies_observationlevels available in a database
}
\examples{
if (interactive()) {
  library(brapi)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  ba_studies_observationlevels(con)

}
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Studies/ListObservationLevels.md}
}
\seealso{
Other studies: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table}}

Other phenotyping: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon
}
