% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_observationunits_save.R
\name{ba_studies_observationunits_save}
\alias{ba_studies_observationunits_save}
\title{ba_studies_observationunits_save}
\usage{
ba_studies_observationunits_save(con = NULL, studyDbId = "",
  unitData = NULL, observationLevel = "plot", transactionDbId = "1234",
  commit = TRUE)
}
\arguments{
\item{con}{brapi connection object}

\item{studyDbId}{character a unique study ID}

\item{unitData}{data.frame or tibble observation unit data}

\item{observationLevel}{character plot (default) or plant}

\item{transactionDbId}{character string}

\item{commit}{logical TRUE (default)}
}
\value{
character a unique sample ID assigned by the server
}
\description{
Call to invoke for saving the measurements (observations) collected from field for all the observation units.
}
\examples{
if (interactive()) {
  library(brapi)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  out <- ba_login(con)

  df <- as.data.frame(cbind(
    observationUnitDbId = rep(c("abc-123", "abc-456"), each = 2),
    observationDbId = c(1, 2, 3, 4),
    observationVariableId = rep(c(18020, 51496), 2),
    observationVariableName = rep(c("Plant_height", "GW100_g"), 2),
    collector = rep("Mr. Technician", 4),
    observationTimeStamp = rep("2015-06-16T00:53:26Z", 4),
    value = c(11, 111, 22, 222)
  ))

  df

  ba_studies_observationunits_save(
    con,
    studyDbId = "1",
    unitData = df,
    observationLevel = "plot",
    transactionDbId = "1234",
    commit = TRUE
  )

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Studies/SaveOrUpdateObservationUnits.md}{github}
}
\seealso{
Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon
}
