% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_search_get.R
\name{ba_studies_search_get}
\alias{ba_studies_search_get}
\title{ba_studies_search_get}
\usage{
ba_studies_search_get(con = NULL, studyDbId = "", trialDbId = "",
  programDbId = "", commonCropName = "", locationDbId = "",
  seasonDbId = "", studyType = "", germplasmDbIds = "",
  observationVariableDbIds = "", active = TRUE, sortBy = "",
  sortOrder = "", pageSize = 1000, page = 0, rclass = c("tibble",
  "data.frame", "list", "json"))
}
\arguments{
\item{con}{list, brapi connection object}

\item{studyDbId}{character, search for study details by an internal study
database identifier; default: ""}

\item{trialDbId}{character, search for studies by an internal trial database
identifier; default: ""}

\item{programDbId}{character, search studies by an internal program database
identifier; default: ""}

\item{commonCropName}{character, search for studies by a common crop name;
default: ""}

\item{locationDbId}{character, search for studies by an internal location
database identifier; default: ""}

\item{seasonDbId}{character, search for studies by an internal season
database identifier; default: ""}

\item{studyType}{character; search for studies based on a study type e.g.
"Nursery", "Trial" or "Genotype"; default: ""}

\item{germplasmDbIds}{character vector; search for studies where specified
germplasms, supplied as a comma separated character
vector of internal gerplasm database identifiers e.g.
c("CML123","CWL123"), have been used/tested; default:
""}

\item{observationVariableDbIds}{character vector; search for studies where
specified observation variables, supplied as
a comma separated character vector of
internal observation variable database
identifiers e.g. c("CO-PH-123:,"Var-123"),
have been measured; default: ""}

\item{active}{logical; search studies by active status; default: "TRUE",
other possible values TRUE/FALSE}

\item{sortBy}{character; name of the field to sort by; default: ""}

\item{sortOrder}{character; sort order direction; default: "", possible values
"asc"/"desc"}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"data.frame"}
}
\value{
An object of class as defined by rclass containing the studies and
        details fulfilling the search criteria.
}
\description{
Search for study details on a brapi server via a GET method.
}
\note{
Tested against: test-server

BrAPI Version: 1.0, 1.1, 1.2

BrAPI Status: active
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Studies/StudiesSearch_GET.md}{github}
}
\seealso{
Other studies: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_germplasm_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}}

Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples_search}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{phenotyping}
\concept{studies}
