% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_search_post.R
\name{ba_studies_search_post}
\alias{ba_studies_search_post}
\title{ba_studies_search_post}
\usage{
ba_studies_search_post(con = NULL, studyDbIds = "", trialDbIds = "",
  programDbIds = "", locationDbIds = "", seasonDbId = "",
  studyType = "", studyNames = "", studyLocations = "",
  programNames = "", commonCropName = "", germplasmDbIds = "",
  observationVariableDbIds = "", active = "", sortBy = "",
  sortOrder = "", page = 0, pageSize = 1000, rclass = "tibble")
}
\arguments{
\item{con}{list, brapi connection object}

\item{studyDbIds}{character, search for study details of specified studies,
supplied as a comma separated string of internal study
database identifiers (e.g. "1,5,19"); default: ""}

\item{trialDbIds}{character, search for studies and  details of specified
trials, supplied as a comma separated string of internal
trial database identifiers (e.g. "15,12,19"); default: ""}

\item{programDbIds}{character, search for studies and details of specified
programs, supplied as a comma separated string of
internal program database identifiers (e.g. "15,12,19");
default: ""}

\item{locationDbIds}{character, search for studies and detaisl of specified
locations, supplied as a comma separated string of
internal location database identifiers (e.g. "13,23,45");
default:""}

\item{seasonDbId}{character, search for studies by an internal season
database identifier; default: ""}

\item{studyType}{character; search for studies and details based on a study
type e.g. "Nursery", "Trial" or "Genotype"; default: ""}

\item{studyNames}{character, search for studies by study names, supplies as a
comma separated string of study names (e.g. ); default: ""}

\item{studyLocations}{character; search for studies by study locations,
supplied as a comma separated string of study locations
(e.g. ); default: ""}

\item{programNames}{character; search for studies by program names, supplied
as a comma separated string of program names (e.g. );
default: ""}

\item{commonCropName}{character, search for studies by a common crop name;
default: ""}

\item{germplasmDbIds}{character; search for studies where specified
germplasms, supplied as a comma separated string of
internal gerplasm database identifiers (e.g. "1,12,
281"), have been used/tested; default: ""}

\item{observationVariableDbIds}{character; search for studies where specified
observation variables, supplied as a comma
separated string of internal observation
variable database identifiers (e.g. "15,100"
), have been measured; default: ""}

\item{active}{logical; search studies by active status; default: "",
other possible values TRUE/FALSE}

\item{sortBy}{character; name of the field to sort by; default: ""}

\item{sortOrder}{character; sort order direction; default: "", possible values
"asc"/"desc"}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"data.frame"}
}
\value{
An object of class as defined by rclass containing the studies and
        details fulfilling the search criteria.
}
\description{
Search for study details on a brapi server via a POST method
}
\note{
Tested against: sweetpotatobase, test-server

BrAPI Version: 1.0, 1.1, 1.2

BrAPI Status: active
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Studies/StudiesSearch_POST.md}{github}
}
\seealso{
Other studies: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_get}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}}

Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_get}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon, Maikel Verouden
}
