% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_seasons.R
\name{ba_studies_seasons}
\alias{ba_studies_seasons}
\title{ba_studies_seasons}
\usage{
ba_studies_seasons(con = NULL, year = 0, pageSize = 1000, page = 0,
  rclass = c("tibble", "data.frame", "list", "json"))
}
\arguments{
\item{con}{list, brapi connection object}

\item{year}{integer, filter by specified year; default: 0}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"data.frame"}
}
\value{
An object of class as defined by rclass containing all seasons (or
        years).
}
\description{
Retrieve all seasons (or years) on a brapi server.
}
\note{
Tested against: sweetpotato, test-server

BrAPI Version: 1.0, 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_studies_seasons(con)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Studies/Seasons_GET.md}{github}
}
\seealso{
Other studies: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_germplasm_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_get}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}}

Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples_search}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_get}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{phenotyping}
\concept{studies}
