% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_studytypes.R
\name{ba_studies_studytypes}
\alias{ba_studies_studytypes}
\title{ba_studies_studytypes}
\usage{
ba_studies_studytypes(con = NULL, page = 0, pageSize = 1000,
  rclass = "tibble")
}
\arguments{
\item{con}{list, brapi connection object}

\item{page}{integer; default 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{character; default: tibble}
}
\value{
rclass as defined
}
\description{
lists studies_studytypes available on a brapi server
}
\examples{
if (interactive()) {
  library(brapi)

  # Need to connect to a database with genetic data

  # con <- ba_db()$xxxxx

  # ba_studies_studytypes(con = con, ...)
}
}
\author{
Reinhard Simon
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Studies/ListStudyTypes.md}{github}
}
\seealso{
Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}

Other studies: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}}
}

