% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_table.R
\name{ba_studies_table}
\alias{ba_studies_table}
\title{ba_studies_table}
\usage{
ba_studies_table(con = NULL, studyDbId = "1", format = "json",
  rclass = "tibble")
}
\arguments{
\item{con}{list; brapi connection object}

\item{studyDbId}{character; default: 1}

\item{format}{character; one of: json, csv, tsv. Defasult: json}

\item{rclass}{character; default: tibble}
}
\value{
rclass as defined
}
\description{
lists studies_table available on a brapi server
}
\examples{
if (interactive()) {
  library(brapi)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  ba_studies_table(con)

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Studies/StudyObservationUnitsAsTable.md}{github}
}
\seealso{
Other studies: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}}

Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon
}
