% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_traits_details.R
\name{ba_traits_details}
\alias{ba_traits_details}
\title{ba_traits_details}
\usage{
ba_traits_details(con = NULL, traitDbId = "", rclass = c("tibble",
  "data.frame", "list", "json"))
}
\arguments{
\item{con}{list; brapi connection object}

\item{traitDbId}{character; \strong{REQUIRED ARGUMENT} with default ''}

\item{rclass}{character; default: tibble}
}
\value{
rclass as defined
}
\description{
lists brapi_traits_details available on a brapi server
}
\note{
Tested against: sweetpotatobase, testserver

BrAPI Version: 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_traits_details(con, traitDbId = "1")
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Traits/TraitDetails.md}{github}
}
\seealso{
Other traits: \code{\link{ba_traits}}

Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_commoncropnames}}, \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search_post}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search_post}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_studies_germplasm_details}},
  \code{\link{ba_traits}}
}
\author{
Reinhard Simon
}
\concept{brapicore}
\concept{traits}
