% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_trials.R
\name{ba_trials}
\alias{ba_trials}
\title{ba_trials}
\usage{
ba_trials(con = NULL, programDbId = "", locationDbId = "",
  active = "any", sortBy = "", sortOrder = "", pageSize = 1000,
  page = 0, rclass = "tibble")
}
\arguments{
\item{con}{list, brapi connection object}

\item{programDbId}{character; program filter to only return trials associated
with given program database identifier; default: ""}

\item{locationDbId}{character, location filter to only return trails associated
with given location databas identifier; default: ""}

\item{active}{logical; filter active status; default: "any", other possible
values TRUE/FALSE}

\item{sortBy}{character; name of the field to sort by; default: ""}

\item{sortOrder}{character; sort order direction; default: "", possible values
"asc"/"desc"}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"data.frame"}
}
\value{
An object of class as defined by rclass containing tiral summaries.
}
\description{
Lists trial summaries available on a brapi server
}
\note{
Tested against: BMS, sweetpotatobase, test-server

BrAPI Version: 1.0, 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_trials(con)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Trials/ListTrialSummaries.md}{github}
}
\seealso{
Other trials: \code{\link{ba_trials_details}}

Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationVariables}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_get}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{phenotyping}
\concept{trials}
