% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_trials_details.R
\name{ba_trials_details}
\alias{ba_trials_details}
\title{ba_trials_details}
\usage{
ba_trials_details(con = NULL, trialDbId = "any", rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{trialDbId}{character; default: 'any'; otherwise an identifier}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists trials_details available on a brapi server
}
\examples{
if (interactive()) {
  library(brapi)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  ba_trials_details(con)

}
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Trials/GetTrialById.md}(github)
}
\seealso{
Other trials: \code{\link{ba_trials}}

Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon
}
