% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_vendor_plates_details.R
\name{ba_vendor_plates_details}
\alias{ba_vendor_plates_details}
\title{ba_vendor_plates_details}
\usage{
ba_vendor_plates_details(con = NULL, vendorPlateDbId = "",
  rclass = c("list", "json"))
}
\arguments{
\item{con}{list, brapi connection object}

\item{vendorPlateDbId}{character, \strong{REQUIRED ARGUMENT} with default: ""}

\item{rclass}{character; default: "list" possible other values: "json"}
}
\value{
rclass as defined
}
\description{
Details of the plate.
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_vendor_plates_details(con, "pl1")
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/ExternalVendorSamples/Vendor_Plates_GET.md}{github}
}
\author{
Reinhard Simon, Maikel Verouden
}
