% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_vendor_plates_search.R
\name{ba_vendor_plates_search}
\alias{ba_vendor_plates_search}
\title{ba_vendor_plates_search}
\usage{
ba_vendor_plates_search(con = NULL, vendorProjectDbId = "",
  vendorPlateDbId = "", clientPlateDbId = "", sampleInfo = TRUE,
  pageSize = 1000, page = 0, rclass = c("list", "json"))
}
\arguments{
\item{con}{list, brapi connection object}

\item{vendorProjectDbId}{character,  default: ""}

\item{vendorPlateDbId}{character,  default: ""}

\item{clientPlateDbId}{character,  default: ""}

\item{sampleInfo}{logical,  default: TRUE}

\item{pageSize}{integer; default 1000}

\item{page}{integer; default 0}

\item{rclass}{character; default: "list" possible other values: "json"}
}
\value{
rclass as defined
}
\description{
Search for vendor plates
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_vendor_plates_search(con)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/ExternalVendorSamples/Vendor_PlatesSearch_GET.md}{github}
}
\author{
Reinhard Simon, Maikel Verouden
}
