% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_vendor_plates_search_post.R
\name{ba_vendor_plates_search_post}
\alias{ba_vendor_plates_search_post}
\title{ba_vendor_plates_search_post}
\usage{
ba_vendor_plates_search_post(con = NULL, vendorProjectDbIds = "",
  vendorPlateDbIds = "", clientPlateDbIds = "", sampleInfo = TRUE,
  pageSize = 1000, page = 0, rclass = c("list", "json"))
}
\arguments{
\item{con}{list, brapi connection object}

\item{vendorProjectDbIds}{character,  default: ""}

\item{vendorPlateDbIds}{character,  default: ""}

\item{clientPlateDbIds}{character,  default: ""}

\item{sampleInfo}{logical,  default: TRUE}

\item{pageSize}{integer; default 1000}

\item{page}{integer; default 0}

\item{rclass}{character; default: "list" possible other values: "json"}
}
\value{
rclass as defined
}
\description{
Search for vendor plates
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_vendor_plates_search_post(con)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/ExternalVendorSamples/Vendor_PlatesSearch_POST.md}{github}
}
\author{
Reinhard Simon, Maikel Verouden
}
