% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_commoncropnames.R,
%   R/ba_studies_germplasm_details.R, R/brapi-deprecated.R
\name{brapi-deprecated}
\alias{brapi-deprecated}
\alias{ba_commoncropnames}
\alias{ba_germplasm_details_study}
\title{ba_commoncropnames}
\usage{
ba_commoncropnames(con = NULL, rclass = "tibble")

ba_germplasm_details_study(con = NULL, studyDbId = "", page = 0,
  pageSize = 1000, rclass = "tibble")
}
\arguments{
\item{con}{list, brapi connection object}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"vector"/"data.frame"}

\item{studyDbId}{character, the internal database identifier for a study of
which the germplasm details are to be retrieved e.g. "1001";
\strong{REQUIRED ARGUMENT} with default: ""}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{con}{list, brapi connection object}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"data.frame"}
}
\value{
an object of class as specified by rclass with common crop names or
        NULL

An object of class as defined by rclass containing the germplasm
        details for a requested study.
}
\description{
List supported crops in a database.

Gets germplasm details for a specific study available on a brapi server.

These functions still work but will be removed (defunct) in a future version.
}
\details{
\itemize{
 \item \code{\link{ba_crops}}: This function is deprecated, and will
 be removed in a future version of this package. Use \code{\link{ba_commoncropnames}} instead.
}
\itemize{
 \item \code{\link{ba_germplasm_details_study}}: This function is deprecated, and will
 be removed in a future version of this package. Use \code{\link{ba_studies_germplasm_details}} instead.
}
}
\note{
Tested against: test-server

BrAPI Version: 1.2

BrAPI Status: active

Tested against: sweetpotatobase, test-server

BrAPI Version: 1.0, 1.1, 1.2

BrAPI Status: active
}
\section{\code{ba_crops}}{

For \code{ba_crops}, use \code{\link{ba_commoncropnames}}.
}

\section{\code{ba_germplasm_details_study}}{

For \code{ba_germplasm_details_study}, use \code{\link{ba_studies_germplasm_details}}.
}

\examples{
if (interactive()) {
  library(brapi)

  con <-  ba_db()$testserver

  ba_commoncropnames(con = con)
}
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_studies_germplasm_details(con = con, studyDbId = "1001")
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Crops/CommonCropNames_GET.md}{github}

\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Studies/Studies_Germplasm_GET.md}{github}
}
\seealso{
Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_crops-deprecated}},
  \code{\link{ba_germplasm_details_study-deprecated}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}

Other studies: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_get}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}}

Other germplasm: \code{\link{ba_germplasm_details_study-deprecated}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search}}

Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_crops-deprecated}},
  \code{\link{ba_germplasm_details_study-deprecated}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon, Maikel Verouden

Reinhard Simon, Maikel Verouden
}
\concept{brapicore}
\concept{germplasm}
\concept{studies}
