% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi-package.r
\docType{package}
\name{brapi-package}
\alias{brapi-package}
\title{brapi Package.}
\description{
BrAPI (www.brapi.org) is an initiative to facilitate data
             exchange between databases dedicated to support plant breeding.
             The specific objective is to provide an Application Programming
             Interface (API) such that data can be accessed programmatically
             using a RESTful approach. The brapi R package accesses the API
             and transforms the data responses into tables. This project is
             simply meant as a container to organize publications.
}
