% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_auth.R
\name{brapi_auth}
\alias{brapi_auth}
\title{Authenticate against a brapi compatible database}
\usage{
brapi_auth(user, password)
}
\arguments{
\item{user}{characters}

\item{password}{characters}
}
\description{
Caches a session token and informs of success or failure.
}
\author{
Reinhard Simon
}

