% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRAPI_CON.R
\name{brapi_con}
\alias{brapi_con}
\title{brapi_con}
\usage{
brapi_con(crop, db, port, user, pwd, session = "")
}
\arguments{
\item{crop}{character string}

\item{db}{character string (url)}

\item{port}{numeric}

\item{user}{character string}

\item{pwd}{character string}

\item{session}{character string}
}
\value{
object of class 'brapi'
}
\description{
S3 class constructor.
}
\examples{

# construct a sample object
library(brapi)

brapi = brapi_con("potato", "http://sample.com", 8080, "rs", "pwd")

is.brapi(brapi)

brapi
}
\author{
Reinahard Simon
}
\seealso{
Other brapi_con: \code{\link{is.brapi}},
  \code{\link{print.brapi_con}}
}

