% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_studies_table.R
\name{brapi_studies_table}
\alias{brapi_studies_table}
\title{brapi_studies_table}
\usage{
brapi_studies_table(con = NULL, studyDbId = 1, format = "json",
  rclass = "tibble")
}
\arguments{
\item{con}{object; brapi connection object}

\item{studyDbId}{string; default: 1}

\item{format}{string; one of: json, csv, tsv. Defasult: json}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists brapi_studies_table available on a brapi server
}
\author{
Reinhard Simon
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/brapi_studies_table/Listbrapi_studies_table.md}
}
\seealso{
Other brapi_call: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}},
  \code{\link{brapi_traits_details}},
  \code{\link{brapi_traits}}, \code{\link{calls}},
  \code{\link{crops}}, \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{locations}}, \code{\link{maps_details}},
  \code{\link{maps_positions_range}},
  \code{\link{maps_positions}}, \code{\link{maps}},
  \code{\link{marker_search}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}},
  \code{\link{observationLevels}}, \code{\link{programs}},
  \code{\link{seasons}}, \code{\link{studies_germplasm}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationVariables}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_search}}, \code{\link{studies}},
  \code{\link{studyTypes}}, \code{\link{trial_by_id}},
  \code{\link{trials}}

Other phenotyping: \code{\link{brapi_traits_details}},
  \code{\link{brapi_traits}}, \code{\link{locations}},
  \code{\link{observationLevels}},
  \code{\link{studies_germplasm}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationVariables}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_search}}, \code{\link{studyTypes}},
  \code{\link{trial_by_id}}, \code{\link{trials}}
}

