% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapiConnect.R
\name{brapiConnect}
\alias{brapiConnect}
\title{brapiConnect}
\usage{
brapiConnect(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
shiny server side complement
}

