% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapiConnectAddin.R
\name{brapiConnectAddin}
\alias{brapiConnectAddin}
\title{brapiConnectAddin}
\usage{
brapiConnectAddin()
}
\description{
Provides a form to connect to a BrAPI compatible database.
}
\examples{
if (interactive()) {
  brapiConnectAddin()
}
}
\author{
Reinhard Simon
}

