% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapiConnectInput.R
\name{brapiConnectInput}
\alias{brapiConnectInput}
\title{BrAPI Connect}
\usage{
brapiConnectInput(id, label = "Connect to BrAPI database")
}
\arguments{
\item{id}{text}

\item{label}{text}
}
\description{
Creates a shiny based widget to collect connection data.
}
\author{
Reinhard Simon
}

