% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can_internet.R
\name{can_internet}
\alias{can_internet}
\title{can_internet}
\usage{
can_internet(url = "www.google.org")
}
\arguments{
\item{url}{a url to check; default is Google}
}
\value{
logical TRUE or FALSE
}
\description{
Check for internet connection.
}
\references{
stackoverflow
}

