% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-methods.R
\name{chart.ba_genomemaps}
\alias{chart.ba_genomemaps}
\title{chart.ba_genomemaps}
\usage{
\method{chart}{ba_genomemaps}(x, ...)
}
\arguments{
\item{x}{a brapi_locations data.frame/tibble or a numeric vector of longitudes}

\item{...}{other plot parameters}
}
\description{
print method for an object of class brapi_brapi_genomemaps, which will only display the crop,  database address:port and user
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  # example 1
  gm <- ba_genomemaps(con)
  gm \%>\% chart

}
}
\author{
Reinhard Simon
}
