% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-methods.R
\name{chart.ba_locations}
\alias{chart.ba_locations}
\title{chart.ba_locations}
\usage{
\method{chart}{ba_locations}(x, ...)
}
\arguments{
\item{x}{a brapi_locations data.frame/tibble or a numeric vector of longitudes}

\item{...}{other plot parameters}
}
\description{
method for an object of class brapi_locations, which will only display the crop,  database address:port and user
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)
  library(maps)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  # Example 1
  bl <- ba_locations()
  bl \%>\% chart()

  # Example 2
  bl \%>\% chart(chart_type = "map")
}
}
\author{
Reinhard Simon
}
