% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-methods.R
\name{chart.brapi_genomemaps_details}
\alias{chart.brapi_genomemaps_details}
\title{chart.brapi_genomemaps_details}
\usage{
\method{chart}{brapi_genomemaps_details}(x, ...)
}
\arguments{
\item{x}{a brapi_genomemaps_details_data data.frame/tibble or a numeric vector of longitudes}

\item{...}{other plot parameters}
}
\description{
print method for an object of class brapi_brapi_genomemaps_details, which will only display the crop,  database address:port and user
}
\author{
Reinhard Simon
}
