% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.brapi_locations.R
\name{chart.brapi_locations}
\alias{chart.brapi_locations}
\title{chart.brapi_locations}
\usage{
\method{chart}{brapi_locations}(x, ...)
}
\arguments{
\item{x}{a brapi_locations data.frame/tibble or a numeric vector of longitudes}

\item{...}{other plot parameters}
}
\description{
print method for an object of class brapi_con, which will only display the crop,  database address:port and user
}
\author{
Reinhard Simon
}

