% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{chart}
\alias{chart}
\title{chart}
\usage{
chart(x, ...)
}
\arguments{
\item{x}{an object to be charted}

\item{...}{additional parameters}
}
\description{
A generic method to chart brapi objects
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)
  library(maps)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  # example 1
  gm <- ba_genomemaps(con)
  gm \%>\% chart

  # example 2
  gmd <- ba_genomemaps_details(con, mapDbId = 1)
  gmd \%>\% chart

  # example 3
  bl <- ba_locations()
  bl \%>\% chart()

  # example 4
  bl \%>\% chart(chart_type = "map")
}
}
\seealso{
Other brapiutils: \code{\link{describe}},
  \code{\link{is.ba_con}}, \code{\link{print.ba_db_list}}
}
\author{
Reinhard Simon
}
