% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{check}
\usage{
check(verbose = TRUE)
}
\arguments{
\item{verbose}{logical; default TRUE}
}
\value{
logical
}
\description{
Checks if a BrAPI server can be reached given the connection details.
}
\details{
Raises errors.
}

