% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{connect}
\usage{
connect(crop = "sweetpotato", db = "127.0.0.1", port = 2021,
  protocol = "http://", multi = FALSE, user = "user",
  password = "password", session = "")
}
\arguments{
\item{crop}{string; default: sweetpotato}

\item{db}{string; default '127.0.0.1' (localhost)}

\item{port}{integer; default: 2021}

\item{protocol}{string; default: http:// (else: https://)}

\item{multi}{logical; default: FALSE}

\item{user}{string; default: user}

\item{password}{string; default: password}

\item{session}{string; default: '' (empty)}
}
\value{
logical
}
\description{
Sets a global list variable 'brapi' that contains all parameters to connect to
a BrAPI compliant online database.
}
\details{
The default values are chosen to match a local mock test server. The mock server
must be started in a separate R session using 'brapi::mock_server()'.
}
\seealso{
Other access: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{locations}}, \code{\link{maps}}
}

