% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crops.R
\name{crops}
\alias{crops}
\title{crops}
\usage{
crops(format = NULL)
}
\arguments{
\item{format}{logical; default is FALSE; whether to display the raw json object or not}
}
\value{
a vector of crop names or NULL
}
\description{
lists crops available in a database
}
\author{
Reinhard Simon
}

