% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe-methods.R
\name{describe.ba_locations}
\alias{describe.ba_locations}
\title{describe.ba_locations}
\usage{
\method{describe}{ba_locations}(x, ...)
}
\arguments{
\item{x}{a brapi_locations object}

\item{...}{other print parameters}
}
\description{
describe method for an object of class brapi_con, which will only display the crop,  database address:port and user
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  lcs <- ba_locations(con)
  lcs \%>\% describe
}
}
\author{
Reinhard Simon
}
