% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe-methods.R
\name{describe.brapi_locations}
\alias{describe.brapi_locations}
\title{describe.brapi_locations}
\usage{
\method{describe}{brapi_locations}(x, ...)
}
\arguments{
\item{x}{a brapi_locations object}

\item{...}{other print parameters}
}
\description{
describe method for an object of class brapi_con, which will only display the crop,  database address:port and user
}
\author{
Reinhard Simon
}
