% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{describe}
\usage{
describe(x, ...)
}
\arguments{
\item{x}{an object to be described}

\item{...}{additional parameters}
}
\description{
A generic method to sum up brapi objects
}
\author{
Reinhard Simon
}
\seealso{
Other brapiutils: \code{\link{chart}},
  \code{\link{is.brapi_con}}, \code{\link{is.brapi}}
}

