% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldbook_analysisInput.R
\name{fieldbook_analysisInput}
\alias{fieldbook_analysisInput}
\title{Fieldbook analysis input}
\usage{
fieldbook_analysisInput(id, label = "Fieldbook analysis parameters")
}
\arguments{
\item{id}{text}

\item{label}{text}
}
\description{
Creates a shiny based widget to collect field book analysis parameters.
}
\author{
Reinhard Simon
}

