% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomemaps.R
\name{genomemaps}
\alias{genomemaps}
\title{genomemaps}
\usage{
genomemaps(con = NULL, species = "all", type = "all", page = 0,
  pageSize = 30, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{species}{string; default:}

\item{type}{string, default: all}

\item{page}{integer; default 0}

\item{pageSize}{integer; default 30}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists genomemaps available on a brapi server
}
\author{
Reinhard Simon
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/GenomeMaps/ListOfGenomeMaps.md}{github}
}
\seealso{
Other genomemaps: \code{\link{genomemaps_data_range}},
  \code{\link{genomemaps_data}},
  \code{\link{genomemaps_details}}

Other genotyping: \code{\link{genomemaps_data_range}},
  \code{\link{genomemaps_data}},
  \code{\link{genomemaps_details}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_search}},
  \code{\link{germplasmattributes_categories}},
  \code{\link{germplasmattributes_details}},
  \code{\link{germplasmattributes}},
  \code{\link{markerprofiles_allelematrix_search}},
  \code{\link{markerprofiles_details}},
  \code{\link{markerprofiles_search}},
  \code{\link{markers_details}},
  \code{\link{markers_search}}
}

