% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomemaps_data.R
\name{genomemaps_data}
\alias{genomemaps_data}
\title{genomemaps_data}
\usage{
genomemaps_data(con = NULL, mapDbId = 0, linkageGroupId = 0, page = 0,
  pageSize = 30, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{mapDbId}{integer; default 0}

\item{linkageGroupId}{integer; default 0}

\item{page}{integer; default 0}

\item{pageSize}{integer; default 30}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists genomemaps_data available on a brapi server
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/GenomeMaps/GenomeMapData.md}{github}
}
\seealso{
Other genomemaps: \code{\link{genomemaps_data_range}},
  \code{\link{genomemaps_details}},
  \code{\link{genomemaps}}

Other genotyping: \code{\link{genomemaps_data_range}},
  \code{\link{genomemaps_details}},
  \code{\link{genomemaps}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_search}},
  \code{\link{germplasmattributes_categories}},
  \code{\link{germplasmattributes_details}},
  \code{\link{germplasmattributes}},
  \code{\link{markerprofiles_allelematrix_search}},
  \code{\link{markerprofiles_details}},
  \code{\link{markerprofiles_search}},
  \code{\link{markers_details}},
  \code{\link{markers_search}}
}
\author{
Reinhard Simon
}
