% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomemaps_data_range.R
\name{genomemaps_data_range}
\alias{genomemaps_data_range}
\title{genomemaps_data_range}
\usage{
genomemaps_data_range(con = NULL, mapDbId = 1, linkageGroupId = 1,
  min = 1, max = 1000, page = 0, pageSize = 30, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{mapDbId}{integer; default 0}

\item{linkageGroupId}{integer; default 0}

\item{min}{integer; default 0}

\item{max}{integer; default 1000}

\item{page}{integer; default 0}

\item{pageSize}{integer; default 30}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists genomemaps_data_range available on a brapi server
}
\author{
Reinhard Simon
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/GenomeMaps/GenomeMapDataByRangeOnLinkageGroup.md}
}
\seealso{
Other genomemaps: \code{\link{genomemaps_data}},
  \code{\link{genomemaps_details}},
  \code{\link{genomemaps}}

Other genotyping: \code{\link{genomemaps_data}},
  \code{\link{genomemaps_details}},
  \code{\link{genomemaps}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_search}},
  \code{\link{germplasmattributes_categories}},
  \code{\link{germplasmattributes_details}},
  \code{\link{germplasmattributes}},
  \code{\link{markerprofiles_allelematrix_search}},
  \code{\link{markerprofiles_details}},
  \code{\link{markerprofiles_search}},
  \code{\link{markers_details}},
  \code{\link{markers_search}}
}

