% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomemaps_details.R
\name{genomemaps_details}
\alias{genomemaps_details}
\title{genomemaps_details}
\usage{
genomemaps_details(con = NULL, mapDbId = 1, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{mapDbId}{integer; default 0}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists genomemaps_details available on a brapi server
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/GenomeMaps/GenomeMapDetails.md}
}
\seealso{
Other genomemaps: \code{\link{genomemaps_data_range}},
  \code{\link{genomemaps_data}}, \code{\link{genomemaps}}

Other genotyping: \code{\link{genomemaps_data_range}},
  \code{\link{genomemaps_data}}, \code{\link{genomemaps}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_search}},
  \code{\link{germplasmattributes_categories}},
  \code{\link{germplasmattributes_details}},
  \code{\link{germplasmattributes}},
  \code{\link{markerprofiles_allelematrix_search}},
  \code{\link{markerprofiles_details}},
  \code{\link{markerprofiles_search}},
  \code{\link{markers_details}},
  \code{\link{markers_search}}
}
\author{
Reinhard Simon
}
