% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm.R
\name{germplasm}
\alias{germplasm}
\title{germplasm}
\usage{
germplasm(germplasmDbId = 0, rclass = "tibble")
}
\arguments{
\item{germplasmDbId}{string; default 0; an internal ID for a germplasm}

\item{rclass}{character; tibble}
}
\value{
list
}
\description{
Gets germplasm details given an id.
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/germplasm-details-by-germplasmdbid}
}

