% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm_attributes.R
\name{germplasm_attributes}
\alias{germplasm_attributes}
\title{germplasm_germplasm_attributes}
\usage{
germplasm_attributes(germplasmDbId = 1, attributeList = 1, page = 0,
  pageSize = 10, rclass = "tibble")
}
\arguments{
\item{germplasmDbId}{integer; default: 1.}

\item{page}{integer; default: 0.}

\item{pageSize}{integer; default: 10.}

\item{rclass}{character; default: tibble; or else: json, list, data.frame.}

\item{attributeDbId}{integer vector; default: 1.}
}
\value{
rclass as set by parameter.
}
\description{
attibutes call.
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/germplasm-attribute-values-by-germplasmdbid/germplasm-attribute-values}
}
\seealso{
Other attributes: \code{\link{attributes_categories}},
  \code{\link{attributes}}

Other brapi_call: \code{\link{attributes_categories}},
  \code{\link{attributes}}

Other genotyping: \code{\link{attributes_categories}},
  \code{\link{attributes}}
}

