% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm_details.R
\name{germplasm_details}
\alias{germplasm_details}
\title{germplasm details}
\usage{
germplasm_details(id = NULL)
}
\arguments{
\item{id}{integer number}
}
\value{
list of lists
}
\description{
germplasm details
}
\author{
Reinhard Simon
}

