% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm_details_study.R
\name{germplasm_details_study}
\alias{germplasm_details_study}
\title{germplasm_details_study}
\usage{
germplasm_details_study(con = NULL, studyDbId = 123, page = 0,
  pageSize = 1000, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{studyDbId}{character; default: 123}

\item{page}{integer; default 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists studies_germplasm available on a brapi server
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Germplasm/GermplasmDetailsListByStudyDbId.md}
}
\seealso{
Other germplasm: \code{\link{germplasm_details}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}

Other brapicore: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_details}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}},
  \code{\link{observationvariables_datatypes}},
  \code{\link{observationvariables_details}},
  \code{\link{observationvariables_ontologies}},
  \code{\link{observationvariables_search}},
  \code{\link{observationvariables}},
  \code{\link{programs_search}}, \code{\link{programs}},
  \code{\link{traits_details}}, \code{\link{traits}}
}
\author{
Reinhard Simon
}
