% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm_markerprofiles.R
\name{germplasm_markerprofiles}
\alias{germplasm_markerprofiles}
\title{germplasm_markerprofiles}
\usage{
germplasm_markerprofiles(id = NULL)
}
\arguments{
\item{id}{integer}
}
\value{
list of marker profile ids
}
\description{
Gets minimal marker profile data from database using database internal id
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/germplasm/germplasm-markerprofiles/markerprofiles-by-germplasmdbid?console=1}
}

