% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm_mcpd.R
\name{germplasm_mcpd}
\alias{germplasm_mcpd}
\title{germplasm MCPD data}
\usage{
germplasm_mcpd(id = NULL)
}
\arguments{
\item{id}{integer}
}
\value{
list
}
\description{
germplasm MCPD data
}
\author{
Reinhard Simon
}

