% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm_pedigree.R
\name{germplasm_pedigree}
\alias{germplasm_pedigree}
\title{germplasm_pedigree}
\usage{
germplasm_pedigree(germplasmDbId = 0, notation = "purdue",
  rclass = "tibble")
}
\arguments{
\item{germplasmDbId}{integer; default 0}

\item{notation}{string; default: purdue format}

\item{rclass}{string; default: tibble}
}
\value{
list of pedigree data
}
\description{
Gets minimal pedigree data from database using database internal id
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/germplasm-pedigree}
}

