% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm_search.R
\name{germplasm_search}
\alias{germplasm_search}
\title{germplasm search}
\usage{
germplasm_search(germplasmDbId = "none", germplasmName = "none",
  germplasmPUI = "none", page = 0, pageSize = 10)
}
\arguments{
\item{germplasmDbId}{string}

\item{germplasmName}{string}

\item{germplasmPUI}{string}

\item{page}{integer}

\item{pageSize}{integer}
}
\value{
data.frame
}
\description{
Lists germplasmm as result of a search.
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/reference/germplasm/germplasm-search/search-names-to-retrieve-germplasm-records}
}

