% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasmattributes.R
\name{germplasmattributes}
\alias{germplasmattributes}
\title{germplasmattributes}
\usage{
germplasmattributes(con = NULL, attributeCategoryDbId = 0,
  rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{attributeCategoryDbId}{integer; default: 0}

\item{rclass}{character; default: tibble}
}
\value{
rclass as set by parameter
}
\description{
germplasmattributes call.
}
\author{
Reinhard Simon
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/GermplasmAttributes/ListAttributesByAttributeCategoryDbId.md}
}
\seealso{
Other genotyping: \code{\link{genomemaps_data_range}},
  \code{\link{genomemaps_data}},
  \code{\link{genomemaps_details}},
  \code{\link{genomemaps}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_search}},
  \code{\link{germplasmattributes_categories}},
  \code{\link{germplasmattributes_details}},
  \code{\link{markerprofiles_allelematrix_search}},
  \code{\link{markerprofiles_details}},
  \code{\link{markerprofiles_search}},
  \code{\link{markers_details}},
  \code{\link{markers_search}}

Other germplasmattributes: \code{\link{germplasmattributes_categories}},
  \code{\link{germplasmattributes_details}}
}

