% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasmattributes_details.R
\name{germplasmattributes_details}
\alias{germplasmattributes_details}
\title{germplasmattributes_details}
\usage{
germplasmattributes_details(con = NULL, germplasmDbId = 1,
  attributeList = 1, page = 0, pageSize = 10, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{germplasmDbId}{integer; default: 1.}

\item{attributeList}{integer vector; default: 1.}

\item{page}{integer; default: 0.}

\item{pageSize}{integer; default: 10.}

\item{rclass}{character; default: tibble; or else: json, list, data.frame.}
}
\value{
rclass as set by parameter.
}
\description{
attibutes call.
}
\author{
Reinhard Simon
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/GermplasmAttributes/GermplasmAttributeValuesByGermplasmDbId.md}
}
\seealso{
Other genotyping: \code{\link{genomemaps_data_range}},
  \code{\link{genomemaps_data}},
  \code{\link{genomemaps_details}},
  \code{\link{genomemaps}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_search}},
  \code{\link{germplasmattributes_categories}},
  \code{\link{germplasmattributes}},
  \code{\link{markerprofiles_allelematrix_search}},
  \code{\link{markerprofiles_details}},
  \code{\link{markerprofiles_search}},
  \code{\link{markers_details}},
  \code{\link{markers_search}}

Other germplasmattributes: \code{\link{germplasmattributes_categories}},
  \code{\link{germplasmattributes}}
}

