% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_germplasm.R
\name{get_top_germplasm}
\alias{get_top_germplasm}
\title{get_top_germplasm}
\usage{
get_top_germplasm(study = NULL, frac = 0.1, max_g = 20,
  trait = "Harvest index")
}
\arguments{
\item{study}{data.frame}

\item{frac}{numeric value between 0.01 and 1 inclusive}

\item{max_g}{integer maximum number of germplasm}

\item{trait}{character the trait to use for calculation of ranking}
}
\value{
data.frame
}
\description{
Selects a top list of germplasm by calculating ranks on the given trait.
The parameters frac and max_g are both used to calculate a minimum number of germplasm.
That is if the number based on fraction is bigger than max_g, the latter will be used.
}

