% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.ba_con.R
\name{is.ba_con}
\alias{is.ba_con}
\title{is.ba_con}
\usage{
is.ba_con(obj)
}
\arguments{
\item{obj}{object to be tested}
}
\value{
boolean (TRUE of FALSE)
}
\description{
check whether an object is of class brapi_con
}
\seealso{
Other brapiutils: \code{\link{as.ba_db}},
  \code{\link{ba_can_internet}}, \code{\link{ba_check}},
  \code{\link{ba_connect}}, \code{\link{ba_db}},
  \code{\link{ba_login}}, \code{\link{ba_logout}},
  \code{\link{print.ba_con}},
  \code{\link{print.ba_db_list}}
}
\author{
Reinahard Simon, Maikel Verouden
}
\concept{brapiutils}
