% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRAPI_CON.R
\name{is.brapi}
\alias{is.brapi}
\title{is.brapi}
\usage{
is.brapi(obj)
}
\arguments{
\item{obj}{object to be tested}
}
\value{
boolean
}
\description{
check type function
}
\examples{

# construct a sample object
library(brapi)

brapi = brapi_con("potato", "http://sample.com", 8080, "rs", "pwd")

is.brapi(brapi)

brapi
}
\author{
Reinahard Simon
}
\seealso{
Other brapi_con: \code{\link{brapi_auth}},
  \code{\link{brapi_con}}, \code{\link{print.brapi_con}}
}

