% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.brapi_con.R
\name{is.brapi_con}
\alias{is.brapi_con}
\title{is.brapi_con}
\usage{
is.brapi_con(obj)
}
\arguments{
\item{obj}{object to be tested}
}
\value{
boolean (TRUE of FALSE)
}
\description{
check whether an object is of class brapi_con
}
\seealso{
Other brapiutils: \code{\link{chart}},
  \code{\link{describe}}, \code{\link{is.brapi}},
  \code{\link{print.brapi_db_list}}
}
\author{
Reinahard Simon, Maikel Verouden
}
