% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{locations}
\alias{locations}
\title{locations}
\usage{
locations(locationType = "all", rclass = "tibble", page = 0,
  pageSize = 1000)
}
\arguments{
\item{locationType}{string, list of data types}

\item{rclass}{string; default: tibble}

\item{page}{integer; default 0}

\item{pageSize}{integer; default 1000}
}
\value{
rclass as defined
}
\description{
lists locations available on a brapi server
}
\author{
Reinhard Simon
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Locations/ListLocations.md}
}
\seealso{
Other access: \code{\link{calls}}, \code{\link{connect}},
  \code{\link{crops}}

Other brapi_call: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}}, \code{\link{calls}},
  \code{\link{crops}}, \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{marker_search}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}},
  \code{\link{programs}}, \code{\link{seasons}}

Other core: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{marker_search}}, \code{\link{programs}},
  \code{\link{seasons}}
}

