% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations_list.R
\name{locations_list}
\alias{locations_list}
\title{list locations via breeding API.}
\usage{
locations_list()
}
\value{
a data.frame of locations
}
\description{
Sweetpotatobase API uses internally paging; this call uses a pageSize of 10000 for now to retrieve all locations.
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/reference/locations/list/list-locations}
}

