% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_details.R
\name{maps_details}
\alias{maps_details}
\title{maps_details}
\usage{
maps_details(mapDbId = 0, rclass = "tibble")
}
\arguments{
\item{mapDbId}{integer; default 0}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists maps available on a brapi server
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/call-search}
}
\seealso{
Other brapi_call: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}},
  \code{\link{brapi_studies_table}},
  \code{\link{brapi_traits_details}},
  \code{\link{brapi_traits}}, \code{\link{calls}},
  \code{\link{crops}}, \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{locations}},
  \code{\link{maps_positions_range}},
  \code{\link{maps_positions}}, \code{\link{maps}},
  \code{\link{marker_search}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}},
  \code{\link{observationLevels}}, \code{\link{programs}},
  \code{\link{seasons}}, \code{\link{studies_germplasm}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationVariables}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_search}}, \code{\link{studies}},
  \code{\link{studyTypes}}, \code{\link{trial_by_id}},
  \code{\link{trials}}

Other genotyping: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}},
  \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{maps_positions_range}},
  \code{\link{maps_positions}}, \code{\link{maps}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}}
}

