% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerprofiles.R
\name{markerprofiles}
\alias{markerprofiles}
\title{marker search}
\usage{
markerprofiles(germplasmDbId = 0, studyDbId = 0, sampleDbId = 0,
  extractDbId = 0, methodDbId = 0, page = 0, pageSize = 10000,
  rclass = "tibble")
}
\arguments{
\item{germplasmDbId}{integer; default 0}

\item{studyDbId}{integer; default 0}

\item{sampleDbId}{integer; default: 0}

\item{extractDbId}{integer; default: 0}

\item{methodDbId}{string; default: unknown}

\item{page}{integer; default: 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{character; default: tibble}
}
\value{
data.frame
}
\description{
Lists markers as result of a search.
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/markerprofiles}
}
\seealso{
Other brapi_call: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}}, \code{\link{calls}},
  \code{\link{crops}}, \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{locations}}, \code{\link{marker_search}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markers}}, \code{\link{programs}},
  \code{\link{seasons}}

Other genotyping: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}},
  \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markers}}
}

