% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerprofiles_allelematrix_search.R
\name{markerprofiles_allelematrix_search}
\alias{markerprofiles_allelematrix_search}
\title{markerprofiles_allelematrix_search}
\usage{
markerprofiles_allelematrix_search(con = NULL, markerprofileDbId = "",
  markerDbId = "", expandHomozygotes = FALSE, unknownString = "-",
  sepPhased = "|", sepUnphased = "/", format = "json", page = 0,
  pageSize = 10000, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{markerprofileDbId}{character vector; default ""}

\item{markerDbId}{character vector; default ""}

\item{expandHomozygotes}{logical; default false}

\item{unknownString}{chaaracter; default: '-'}

\item{sepPhased}{character; default: '|'}

\item{sepUnphased}{character; default: '/'}

\item{format}{character; default: json; other: csv, tsv}

\item{page}{integer; default: 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{character; default: tibble}
}
\value{
data.frame
}
\description{
Gets markers in matrix format. If the format parameter is set to either csv or tsv the returned object
is always a tibble object. If the format parameter is 'json' (default) the rclass parameter can be used
to as in other functions.
}
\note{
The handling of long-running responses via asynch status messages is not yet implemented.
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/MarkerProfiles/MarkerProfileAlleleMatrix.md}
}
\seealso{
Other markerprofiles: \code{\link{markerprofiles_details}},
  \code{\link{markerprofiles_search}}

Other genotyping: \code{\link{genomemaps_data_range}},
  \code{\link{genomemaps_data}},
  \code{\link{genomemaps_details}},
  \code{\link{genomemaps}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_search}},
  \code{\link{germplasmattributes_categories}},
  \code{\link{germplasmattributes_details}},
  \code{\link{germplasmattributes}},
  \code{\link{markerprofiles_details}},
  \code{\link{markerprofiles_search}},
  \code{\link{markers_details}},
  \code{\link{markers_search}}
}
\author{
Reinhard Simon
}
