% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerprofiles_alleles.R
\name{markerprofiles_alleles}
\alias{markerprofiles_alleles}
\title{marker search}
\usage{
markerprofiles_alleles(markerprofilesDbId = "", expandHomozygotes = FALSE,
  unknownString = "N", sepPhased = "|", sepUnphased = "/", page = 0,
  pageSize = 10000, rclass = "tibble")
}
\arguments{
\item{markerprofilesDbId}{integer; default 0}

\item{expandHomozygotes}{logical; default false}

\item{unknownString}{chaaracter; default: '-'}

\item{sepPhased}{character; default: '|'}

\item{sepUnphased}{character; default: '/'}

\item{page}{integer; default: 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{character; default: tibble}
}
\value{
data.frame
}
\description{
Lists markers as result of a search.
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/markerprofiles_alleles}
}
\seealso{
Other brapi_call: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}}, \code{\link{calls}},
  \code{\link{crops}}, \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{locations}}, \code{\link{maps_details}},
  \code{\link{maps_positions_range}},
  \code{\link{maps_positions}}, \code{\link{maps}},
  \code{\link{marker_search}},
  \code{\link{markerprofiles}}, \code{\link{markers}},
  \code{\link{observationLevels}}, \code{\link{programs}},
  \code{\link{seasons}}, \code{\link{studies_layout}},
  \code{\link{studies_search}}, \code{\link{studies}},
  \code{\link{studyTypes}}, \code{\link{trial_by_id}},
  \code{\link{trials}}

Other genotyping: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}},
  \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{maps_details}},
  \code{\link{maps_positions_range}},
  \code{\link{maps_positions}}, \code{\link{maps}},
  \code{\link{markerprofiles}}, \code{\link{markers}}
}

