% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerprofiles_search.R
\name{markerprofiles_search}
\alias{markerprofiles_search}
\title{markerprofiles_search}
\usage{
markerprofiles_search(con = NULL, germplasmDbId = "", studyDbId = "",
  sampleDbId = "", extractDbId = "", methodDbId = "all", page = 0,
  pageSize = 10000, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{germplasmDbId}{character; default ""}

\item{studyDbId}{character; default ""}

\item{sampleDbId}{character; default: ""}

\item{extractDbId}{character; default: ""}

\item{methodDbId}{string; default: unknown}

\item{page}{integer; default: 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{character; default: tibble}
}
\value{
data.frame
}
\description{
Lists markers as result of a search.
}
\author{
Reinhard Simon
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/MarkerProfiles/MarkerProfileSearch.md}
}
\seealso{
Other genotyping: \code{\link{genomemaps_data_range}},
  \code{\link{genomemaps_data}},
  \code{\link{genomemaps_details}},
  \code{\link{genomemaps}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_search}},
  \code{\link{germplasmattributes_categories}},
  \code{\link{germplasmattributes_details}},
  \code{\link{germplasmattributes}},
  \code{\link{markerprofiles_allelematrix_search}},
  \code{\link{markerprofiles_details}},
  \code{\link{markers_details}},
  \code{\link{markers_search}}

Other markerprofiles: \code{\link{markerprofiles_allelematrix_search}},
  \code{\link{markerprofiles_details}}
}

