% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_details.R
\name{markers_details}
\alias{markers_details}
\title{markers_details}
\usage{
markers_details(con = NULL, markerDbId = 0, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{markerDbId}{integer; marker id; default: 0}

\item{rclass}{character; default: tibble}
}
\value{
data.frame
}
\description{
Lists markers as result of a search.
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Markers/MarkerDetailsByMarkerDbId.md}
}
\seealso{
Other markers: \code{\link{markers_search}}

Other genotyping: \code{\link{genomemaps_data_range}},
  \code{\link{genomemaps_data}},
  \code{\link{genomemaps_details}},
  \code{\link{genomemaps}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_search}},
  \code{\link{germplasmattributes_categories}},
  \code{\link{germplasmattributes_details}},
  \code{\link{germplasmattributes}},
  \code{\link{markerprofiles_allelematrix_search}},
  \code{\link{markerprofiles_details}},
  \code{\link{markerprofiles_search}},
  \code{\link{markers_search}}
}
\author{
Reinhard Simon
}
