% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_search.R
\name{markers_search}
\alias{markers_search}
\title{markers_search}
\usage{
markers_search(con = NULL, name = "*", type = "all",
  matchMethod = "wildcard", include = "synonyms", page = 0,
  pageSize = 1000, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{name}{string; marker name; default: *}

\item{type}{string; default: all; other: SNP}

\item{matchMethod}{string; default: wildcard; other: case_insensitive, exact}

\item{include}{string; default: synonyms}

\item{page}{integer; default: 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{character; default: tibble}
}
\value{
data.frame
}
\description{
Lists markers as result of a search.
}
\author{
Reinhard Simon
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Markers/MarkerSearch.md}{github}
}
\seealso{
Other genotyping: \code{\link{genomemaps_data_range}},
  \code{\link{genomemaps_data}},
  \code{\link{genomemaps_details}},
  \code{\link{genomemaps}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_search}},
  \code{\link{germplasmattributes_categories}},
  \code{\link{germplasmattributes_details}},
  \code{\link{germplasmattributes}},
  \code{\link{markerprofiles_allelematrix_search}},
  \code{\link{markerprofiles_details}},
  \code{\link{markerprofiles_search}},
  \code{\link{markers_details}}

Other markers: \code{\link{markers_details}}
}

