% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{mock_server}
\alias{mock_server}
\title{mock}
\usage{
mock_server(port = 2021)
}
\arguments{
\item{port}{integer; default 2021}
}
\description{
Starts a mock BrAPI server intended for local testing of brapi calls.
}
\seealso{
Other utility: \code{\link{show_info}}
}

