% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observationvariables.R
\name{observationvariables}
\alias{observationvariables}
\title{observationvariables}
\usage{
observationvariables(con = NULL, traitClass = "all", page = 0,
  pageSize = 1000, rclass = "tibble")
}
\arguments{
\item{con}{object; brapi connection object}

\item{traitClass}{string; default 'all'}

\item{page}{integer; default 0}

\item{pageSize}{integer; defautlt 1000}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists brapi_variables available on a brapi server
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/ObservationVariables/VariableList.md}
}
\seealso{
Other observationvariables: \code{\link{observationvariables_datatypes}},
  \code{\link{observationvariables_details}},
  \code{\link{observationvariables_ontologies}}

Other brapicore: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_details_study}},
  \code{\link{germplasm_details}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}},
  \code{\link{observationvariables_datatypes}},
  \code{\link{observationvariables_details}},
  \code{\link{observationvariables_ontologies}},
  \code{\link{observationvariables_search}},
  \code{\link{programs_search}}, \code{\link{programs}},
  \code{\link{traits_details}}, \code{\link{traits}}
}
\author{
Reinhard Simon
}
