% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observationvariables_datatypes.R
\name{observationvariables_datatypes}
\alias{observationvariables_datatypes}
\title{observationvariables_datatypes}
\usage{
observationvariables_datatypes(con = NULL, rclass = "tibble")
}
\arguments{
\item{con}{object; brapi connection object; default = NULL}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists variables_datatypes available on a brapi server
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/ObservationVariables/VariableDataTypeList.md}
}
\seealso{
Other observationvariables: \code{\link{observationvariables_details}},
  \code{\link{observationvariables_ontologies}},
  \code{\link{observationvariables}}

Other brapicore: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_details_study}},
  \code{\link{germplasm_details}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}},
  \code{\link{observationvariables_details}},
  \code{\link{observationvariables_ontologies}},
  \code{\link{observationvariables_search}},
  \code{\link{observationvariables}},
  \code{\link{programs_search}}, \code{\link{programs}},
  \code{\link{traits_details}}, \code{\link{traits}}
}
\author{
Reinhard Simon
}
