% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observationvariables_details.R
\name{observationvariables_details}
\alias{observationvariables_details}
\title{observationvariables_details}
\usage{
observationvariables_details(con = NULL,
  observationVariableDbId = "CO_334:0100622", rclass = "tibble")
}
\arguments{
\item{con}{object; brapi connection object}

\item{observationVariableDbId}{string; default 'CO_334:0100622'; from test server.}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists brapi_variables_details available on a brapi server
}
\author{
Reinhard Simon
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/ObservationVariables/VariableDetails.md}
}
\seealso{
Other brapicore: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_details_study}},
  \code{\link{germplasm_details}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}},
  \code{\link{observationvariables_datatypes}},
  \code{\link{observationvariables_ontologies}},
  \code{\link{observationvariables}},
  \code{\link{programs_search}}, \code{\link{programs}},
  \code{\link{traits_details}}, \code{\link{traits}}

Other observationvariables: \code{\link{observationvariables_datatypes}},
  \code{\link{observationvariables_ontologies}},
  \code{\link{observationvariables}}
}

