% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observationvariables_ontologies.R
\name{observationvariables_ontologies}
\alias{observationvariables_ontologies}
\title{observationvariables_ontologies}
\usage{
observationvariables_ontologies(con = NULL, page = 0, pageSize = 1000,
  rclass = "tibble")
}
\arguments{
\item{con}{object; brapi connection object; default = NULL}

\item{page}{integer; default 0}

\item{pageSize}{integer; defautlt 1000}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists variables_datatypes available on a brapi server
}
\author{
Reinhard Simon
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/ObservationVariables/VariableOntologyList.md}
}
\seealso{
Other brapicore: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_details_study}},
  \code{\link{germplasm_details}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}},
  \code{\link{observationvariables_datatypes}},
  \code{\link{observationvariables_details}},
  \code{\link{observationvariables_search}},
  \code{\link{observationvariables}},
  \code{\link{programs_search}}, \code{\link{programs}},
  \code{\link{traits_details}}, \code{\link{traits}}

Other observationvariables: \code{\link{observationvariables_datatypes}},
  \code{\link{observationvariables_details}},
  \code{\link{observationvariables}}
}

