% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observationvariables_search.R
\name{observationvariables_search}
\alias{observationvariables_search}
\title{observationvariables_search}
\usage{
observationvariables_search(con = NULL, observationVariableDbIds = "any",
  ontologyXrefs = "any", ontologyDbIds = "any", methodDbIds = "any",
  scaleDbIds = "any", names = "any", datatypes = "any",
  traitClasses = "any", rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{observationVariableDbIds}{vector of string; default: any}

\item{ontologyXrefs}{vector of string; default: any}

\item{ontologyDbIds}{vector of string; default: any}

\item{methodDbIds}{vector of string; default: any}

\item{scaleDbIds}{vector of string; default: any}

\item{names}{vector of string; default: any}

\item{datatypes}{vector of string; default: any}

\item{traitClasses}{vector of string; default: any}

\item{rclass}{string; default: tibble}
}
\value{
rclass
}
\description{
lists the breeding observationvariables
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/ObservationVariables/VariableSearch.md}{github}
}
\seealso{
Other brapicore: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_details_study}},
  \code{\link{germplasm_details}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}},
  \code{\link{observationvariables_datatypes}},
  \code{\link{observationvariables_details}},
  \code{\link{observationvariables_ontologies}},
  \code{\link{observationvariables}},
  \code{\link{programs_search}}, \code{\link{programs}},
  \code{\link{traits_details}}, \code{\link{traits}}
}
\author{
Reinhard Simon
}
