% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.locations_list.R
\name{print.brapi}
\alias{print.brapi}
\title{print.brapi}
\usage{
\method{print}{brapi}(x, ...)
}
\arguments{
\item{x}{printable locations_list object}

\item{...}{extended parameters}
}
\value{
data.frame of class brapi
}
\description{
print.brapi
}

