% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.brapi_con.R
\name{print.brapi_con}
\alias{print.brapi_con}
\title{print.brapi_con}
\usage{
\method{print}{brapi_con}(x, ...)
}
\arguments{
\item{x}{a brapi_con object}

\item{...}{other print parameters}
}
\description{
print method for an object of class brapi_con, which will only display the crop,  database address:port and user
}
\author{
Reinhard Simon, Maikel Verouden
}

