% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.brapi_db_list}
\alias{print.brapi_db_list}
\title{print.brapi_db_list}
\usage{
\method{print}{brapi_db_list}(x, ...)
}
\arguments{
\item{x}{a brapi_db_list object}

\item{...}{other print parameters}
}
\description{
print method for an object of class brapi_db_list
}
\author{
Reinhard Simon, Maikel Verouden
}
\seealso{
Other brapiutils: \code{\link{chart}},
  \code{\link{describe}}, \code{\link{is.brapi_con}},
  \code{\link{is.brapi}}
}

