% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program_list.R
\name{program_list}
\alias{program_list}
\title{program list}
\usage{
program_list(pageSize = 20, page = 1)
}
\arguments{
\item{pageSize}{items per page}

\item{page}{integer requested page number}
}
\value{
data.frame
}
\description{
lists the breeding programs
}
\details{
BRAPI discussion: Should this return also the crop?
}
\author{
Reinhard Simon
}

