% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs_search.R
\name{programs_search}
\alias{programs_search}
\title{programs_search}
\usage{
programs_search(con = NULL, programDbId = "any", name = "any",
  abbreviation = "any", objective = "any", leadPerson = "any",
  rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{programDbId}{string; default: any}

\item{name}{string; default: any}

\item{abbreviation}{string; default: any}

\item{objective}{string; default: any}

\item{leadPerson}{string; default: any}

\item{rclass}{string; default: tibble}
}
\value{
rclass
}
\description{
searches the breeding programs
}
\author{
Reinhard Simon
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Programs/ProgramSearch.md}{github}
}
\seealso{
Other brapicore: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_details_study}},
  \code{\link{germplasm_details}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}},
  \code{\link{observationvariables_datatypes}},
  \code{\link{observationvariables_details}},
  \code{\link{observationvariables_ontologies}},
  \code{\link{observationvariables_search}},
  \code{\link{observationvariables}},
  \code{\link{programs}}, \code{\link{traits_details}},
  \code{\link{traits}}
}

