% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{seasons}
\alias{seasons}
\title{seasons}
\usage{
seasons(year = NULL, page = 0, pageSize = 1000, rclass = "list")
}
\arguments{
\item{year}{integer}

\item{page}{integer requested page number}

\item{pageSize}{items per page}

\item{rclass}{character; one of list (default) or data.frame}
}
\value{
data.frame
}
\description{
lists the seasons and years
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/list-seasons/list-seasons-or-years}
}

