% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_auth.R
\name{set_brapi}
\alias{set_brapi}
\title{set BRAPI database}
\usage{
set_brapi(url, port = 3000)
}
\arguments{
\item{url}{string a URL to a local or remote BRAPI compliant database}

\item{port}{integer number}
}
\value{
no returned value
}
\description{
Define which BRAPI database to use for this session.
Leading 'http://' and trailing '/' will be removed.
}
\author{
Reinhard Simon
}

