% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_if_local.R
\name{skip_if_local}
\alias{skip_if_local}
\title{skip_if_local}
\usage{
skip_if_local()
}
\value{
boolean
}
\description{
A helper function to minimize time consuming testing against databases locally.
}
\details{
Only used within formal tests of the package.
}
