% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies_details.R
\name{studies_details}
\alias{studies_details}
\title{studies_details}
\usage{
studies_details(con = NULL, studiesDbId = 0, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{studiesDbId}{string; default 0; an internal ID for a studies}

\item{rclass}{character; tibble}
}
\value{
data.frame
}
\description{
Gets studies details given an id.
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Studies/StudyDetails.md}{github}
}
\seealso{
Other studies: \code{\link{studies_layout}},
  \code{\link{studies_observationlevels}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_observationvariables}},
  \code{\link{studies_search}},
  \code{\link{studies_seasons}},
  \code{\link{studies_studytypes}},
  \code{\link{studies_table}}

Other phenotyping: \code{\link{studies_layout}},
  \code{\link{studies_observationlevels}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_observationvariables}},
  \code{\link{studies_search}},
  \code{\link{studies_seasons}},
  \code{\link{studies_studytypes}},
  \code{\link{studies_table}},
  \code{\link{trials_details}}, \code{\link{trials}}
}
\author{
Reinhard Simon
}
