% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies_layout.R
\name{studies_layout}
\alias{studies_layout}
\title{studies_layout}
\usage{
studies_layout(con = NULL, studyDbId = 1, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{studyDbId}{string; default: 1}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists studies_layout available on a brapi server
}
\author{
Reinhard Simon
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Studies/PlotLayoutDetails.md}{github}
}
\seealso{
Other phenotyping: \code{\link{locations}},
  \code{\link{studies_details}},
  \code{\link{studies_observationlevels}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_observationvariables}},
  \code{\link{studies_search}},
  \code{\link{studies_seasons}},
  \code{\link{studies_studytypes}},
  \code{\link{studies_table}},
  \code{\link{trials_details}}, \code{\link{trials}}

Other studies: \code{\link{studies_details}},
  \code{\link{studies_observationlevels}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_observationvariables}},
  \code{\link{studies_search}},
  \code{\link{studies_seasons}},
  \code{\link{studies_studytypes}},
  \code{\link{studies_table}}
}

