% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies_observationlevels.R
\name{studies_observationlevels}
\alias{studies_observationlevels}
\title{studies_observationlevels}
\usage{
studies_observationlevels(con = NULL, rclass = "vector")
}
\arguments{
\item{con}{brapi connection object}

\item{rclass}{character; default is FALSE; whether to display the raw list object or not}
}
\value{
a vector of crop names or NULL
}
\description{
lists studies_observationlevels available in a database
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Studies/ListObservationLevels.md}
}
\seealso{
Other studies: \code{\link{studies_details}},
  \code{\link{studies_layout}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_observationvariables}},
  \code{\link{studies_search}},
  \code{\link{studies_seasons}},
  \code{\link{studies_studytypes}},
  \code{\link{studies_table}}

Other phenotyping: \code{\link{studies_details}},
  \code{\link{studies_layout}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_observationvariables}},
  \code{\link{studies_search}},
  \code{\link{studies_seasons}},
  \code{\link{studies_studytypes}},
  \code{\link{studies_table}},
  \code{\link{trials_details}}, \code{\link{trials}}
}
\author{
Reinhard Simon
}
