% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies_observationunits.R
\name{studies_observationunits}
\alias{studies_observationunits}
\title{studies_observationunits}
\usage{
studies_observationunits(studyDbId = 1, observationLevel = "plot",
  rclass = "tibble")
}
\arguments{
\item{studyDbId}{character; default: 1}

\item{observationLevel}{string; default: plot; alternative: plant}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists studies_observationunits available on a brapi server
}
\author{
Reinhard Simon
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/studies_observationunits/Liststudies_observationunits.md}
}
\seealso{
Other brapi_call: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}}, \code{\link{calls}},
  \code{\link{crops}}, \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{locations}}, \code{\link{maps_details}},
  \code{\link{maps_positions_range}},
  \code{\link{maps_positions}}, \code{\link{maps}},
  \code{\link{marker_search}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}},
  \code{\link{observationLevels}}, \code{\link{programs}},
  \code{\link{seasons}}, \code{\link{studies_germplasm}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationVariables}},
  \code{\link{studies_observations}},
  \code{\link{studies_search}}, \code{\link{studies}},
  \code{\link{studyTypes}}, \code{\link{trial_by_id}},
  \code{\link{trials}}

Other phenotyping: \code{\link{locations}},
  \code{\link{observationLevels}},
  \code{\link{studies_germplasm}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationVariables}},
  \code{\link{studies_observations}},
  \code{\link{studies_search}}, \code{\link{studyTypes}},
  \code{\link{trial_by_id}}, \code{\link{trials}}
}

