% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies_search.R
\name{studies_search}
\alias{studies_search}
\title{studies_search}
\usage{
studies_search(studyType = "any", programDbId = "any",
  locationDbId = "any", seasonDbId = "any", germplasmDbIds = "any",
  observationVariableDbIds = "any", active = "any", sortBy = "any",
  sortOrder = "any", page = 0, pageSize = 1000, rclass = "tibble")
}
\arguments{
\item{studyType}{string; default: any}

\item{programDbId}{string; default: any}

\item{locationDbId}{string; default: any}

\item{seasonDbId}{string; default: any}

\item{germplasmDbIds}{string; default: any}

\item{observationVariableDbIds}{string; default: any}

\item{active}{string; default: any}

\item{sortBy}{string; default: any}

\item{sortOrder}{string; default: any}

\item{page}{integer; default: 1000}

\item{pageSize}{integer; default: 0}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists trials available on a brapi server
}
\author{
Reinhard Simon
}
\seealso{
Other brapi_call: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}},
  \code{\link{brapi_studies_table}}, \code{\link{calls}},
  \code{\link{crops}}, \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{locations}}, \code{\link{maps_details}},
  \code{\link{maps_positions_range}},
  \code{\link{maps_positions}}, \code{\link{maps}},
  \code{\link{marker_search}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}},
  \code{\link{observationLevels}}, \code{\link{programs}},
  \code{\link{seasons}}, \code{\link{studies_germplasm}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationVariables}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies}}, \code{\link{studyTypes}},
  \code{\link{trial_by_id}}, \code{\link{trials}}

Other phenotyping: \code{\link{brapi_studies_table}},
  \code{\link{locations}}, \code{\link{observationLevels}},
  \code{\link{studies_germplasm}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationVariables}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studyTypes}}, \code{\link{trial_by_id}},
  \code{\link{trials}}
}

