% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies_seasons.R
\name{studies_seasons}
\alias{studies_seasons}
\title{studies_seasons}
\usage{
studies_seasons(con = NULL, year = NULL, page = 0, pageSize = 1000,
  rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{year}{integer}

\item{page}{integer requested page number}

\item{pageSize}{items per page}

\item{rclass}{character; one of tibble (default), json, list or data.frame}
}
\value{
data.frame
}
\description{
lists the studies seasons and years
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Studies/ListSeasons.md}
}
\seealso{
Other studies: \code{\link{studies_details}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationlevels}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_observationvariables}},
  \code{\link{studies_search}},
  \code{\link{studies_studytypes}},
  \code{\link{studies_table}}

Other phenotyping: \code{\link{studies_details}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationlevels}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_observationvariables}},
  \code{\link{studies_search}},
  \code{\link{studies_studytypes}},
  \code{\link{studies_table}},
  \code{\link{trials_details}}, \code{\link{trials}}
}
\author{
Reinhard Simon
}
