% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies_studytypes.R
\name{studies_studytypes}
\alias{studies_studytypes}
\title{studies_studytypes}
\usage{
studies_studytypes(con = NULL, page = 0, pageSize = 1000,
  rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{page}{integer; default 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists studies_studytypes available on a brapi server
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Studies/ListStudyTypes.md}(github)
}
\seealso{
Other studies: \code{\link{studies_details}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationlevels}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_observationvariables}},
  \code{\link{studies_search}},
  \code{\link{studies_seasons}},
  \code{\link{studies_table}}

Other phenotyping: \code{\link{studies_details}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationlevels}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_observationvariables}},
  \code{\link{studies_search}},
  \code{\link{studies_seasons}},
  \code{\link{studies_table}},
  \code{\link{trials_details}}, \code{\link{trials}}
}
\author{
Reinhard Simon
}
