% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies_table.R
\name{studies_table}
\alias{studies_table}
\title{studies_table}
\usage{
studies_table(con = NULL, studyDbId = 1, format = "json",
  rclass = "tibble")
}
\arguments{
\item{con}{object; brapi connection object}

\item{studyDbId}{string; default: 1}

\item{format}{string; one of: json, csv, tsv. Defasult: json}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists studies_table available on a brapi server
}
\author{
Reinhard Simon
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Studies/StudyObservationUnitsAsTable.md}{github}
}
\seealso{
Other phenotyping: \code{\link{studies_details}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationlevels}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_observationvariables}},
  \code{\link{studies_search}},
  \code{\link{studies_seasons}},
  \code{\link{studies_studytypes}},
  \code{\link{trials_details}}, \code{\link{trials}}

Other studies: \code{\link{studies_details}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationlevels}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_observationvariables}},
  \code{\link{studies_search}},
  \code{\link{studies_seasons}},
  \code{\link{studies_studytypes}}
}

