% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{study}
\alias{study}
\title{study description}
\usage{
study(studyId = NULL)
}
\arguments{
\item{studyId}{integer}
}
\value{
list of study attributes
}
\description{
Gets minimal metadata about a study
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/study/study-details/retrieve-study-details?console=1}
}

