% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_details.R
\name{study_details}
\alias{study_details}
\title{study details}
\usage{
study_details(studyId = NULL)
}
\arguments{
\item{studyId}{integer}
}
\value{
list of study attributes
}
\description{
Gets additional metadata about a study
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/study/study-details/retrieve-study-details?console=1}
}

