% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_instances.R
\name{study_instances}
\alias{study_instances}
\title{study instances}
\usage{
study_instances(studyId = NULL, page = 1, pageSize = 100)
}
\arguments{
\item{studyId}{integer}

\item{page}{integer requested page number}

\item{pageSize}{items per page}
}
\value{
list of marker profile ids
}
\description{
Gets list of study instances for a study
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/study/list-studies/list-of-study-summaries?console=1}
}

