% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_layout.R
\name{study_layout}
\alias{study_layout}
\title{study layout}
\usage{
study_layout(studyId = NULL)
}
\arguments{
\item{studyId}{integer}
}
\value{
list of study attributes
}
\description{
Gets additional metadata about a study
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/study/layout/retrieve-study-details?console=1}
}

