% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_plot.R
\name{study_plot}
\alias{study_plot}
\title{study plot}
\usage{
study_plot(studyId = NULL, observationVariableId = NULL)
}
\arguments{
\item{studyId}{integer}

\item{observationVariableId}{integer}
}
\value{
list
}
\description{
Gets metadata about a study plot
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/study/plot/retrieve-plot-details?console=1}
}

