% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies_table.R
\name{study_table}
\alias{study_table}
\title{studies table}
\usage{
study_table(studyId = NULL)
}
\arguments{
\item{studyId}{integer}
}
\value{
fieldbook with attributes
}
\description{
Gets a simple fieldbook book view back in form of a table
}
\details{
The meta attribute has information on year, studyDbId, studyName, locationDbId, list of unique germplasm names, plotNames
variable names.
}
\author{
Reinhard Simon
}

