% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_studyTypes.R
\name{studyTypes}
\alias{studyTypes}
\title{studyTypes}
\usage{
studyTypes(page = 1, pageSize = 1000)
}
\arguments{
\item{page}{integer requested page number}

\item{pageSize}{items per page}
}
\value{
list of marker profile ids
}
\description{
Gets minimal marker profile data from database using database internal id
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/germplasm/germplasm-markerprofiles/markerprofiles-by-germplasmdbid?console=1}
}

