% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{traits}
\alias{traits}
\title{traits}
\usage{
traits(con = NULL, page = 0, pageSize = 1000, rclass = "tibble")
}
\arguments{
\item{con}{object; brapi connection object}

\item{page}{integer; default 0}

\item{pageSize}{integer; defautlt 1000}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists brapi_traits available on a brapi server
}
\author{
Reinhard Simon
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Traits/ListAllTraits.md}{github}
}
\seealso{
Other brapicore: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_details_study}},
  \code{\link{germplasm_details}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}},
  \code{\link{observationvariables_datatypes}},
  \code{\link{observationvariables_details}},
  \code{\link{observationvariables_ontologies}},
  \code{\link{observationvariables}},
  \code{\link{programs_search}}, \code{\link{programs}},
  \code{\link{traits_details}}

Other traits: \code{\link{traits_details}}
}

