% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits_details.R
\name{traits_details}
\alias{traits_details}
\title{traits_details}
\usage{
traits_details(con = NULL, traitDbId = 1, rclass = "tibble")
}
\arguments{
\item{con}{object; brapi connection object}

\item{traitDbId}{alphanumeric; default 1}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists brapi_traits_details available on a brapi server
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Traits/TraitDetails.md}(github)
}
\seealso{
Other traits: \code{\link{traits}}

Other brapicore: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_details_study}},
  \code{\link{germplasm_details}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}},
  \code{\link{observationvariables_datatypes}},
  \code{\link{observationvariables_details}},
  \code{\link{observationvariables_ontologies}},
  \code{\link{observationvariables_search}},
  \code{\link{observationvariables}},
  \code{\link{programs_search}}, \code{\link{programs}},
  \code{\link{traits}}
}
\author{
Reinhard Simon
}
