% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_by_id.R
\name{trial_by_id}
\alias{trial_by_id}
\title{trial_by_id}
\usage{
trial_by_id(trialDbId = "any", rclass = "tibble")
}
\arguments{
\item{trialDbId}{character; default: 'any'; otherwise an identifier}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists trials available on a brapi server
}
\author{
Reinhard Simon
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Trials/ListTrialSummaries.md}
}
\seealso{
Other brapi_call: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}},
  \code{\link{brapi_studies_table}},
  \code{\link{brapi_variables_datatypes}},
  \code{\link{brapi_variables_ontologies}},
  \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{locations}}, \code{\link{maps_details}},
  \code{\link{maps_positions_range}},
  \code{\link{maps_positions}}, \code{\link{maps}},
  \code{\link{marker_search}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}},
  \code{\link{observationLevels}}, \code{\link{programs}},
  \code{\link{seasons}}, \code{\link{studies_germplasm}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationVariables}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_search}}, \code{\link{studies}},
  \code{\link{studyTypes}}, \code{\link{trials}}

Other phenotyping: \code{\link{brapi_studies_table}},
  \code{\link{brapi_traits_details}},
  \code{\link{brapi_traits}},
  \code{\link{brapi_variables_details}},
  \code{\link{brapi_variables}}, \code{\link{locations}},
  \code{\link{observationLevels}},
  \code{\link{studies_germplasm}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationVariables}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_search}}, \code{\link{studyTypes}},
  \code{\link{trials}}
}

