% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trials.R
\name{trials}
\alias{trials}
\title{trials}
\usage{
trials(con = NULL, programDbId = "any", locationDbId = "any",
  active = "any", sortBy = "none", sortOrder = "asc", page = 0,
  pageSize = 1000, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{programDbId}{character; default: 'any'; otherwise an identifier}

\item{locationDbId}{character; default: 'any'; otherwise an identifier}

\item{active}{character; default: 'any'; otherwise TRUE or FALSE}

\item{sortBy}{character; default: 'none'; otherwise a name from the first level entry names of the return object.}

\item{sortOrder}{character; default: 'asc'; otherwise 'desc'}

\item{page}{integer; default 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists trials available on a brapi server
}
\author{
Reinhard Simon
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Trials/ListTrialSummaries.md}
}
\seealso{
Other phenotyping: \code{\link{locations}},
  \code{\link{studies_details}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationlevels}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_observationvariables}},
  \code{\link{studies_search}},
  \code{\link{studies_seasons}},
  \code{\link{studies_studytypes}},
  \code{\link{studies_table}}, \code{\link{trials_details}}

Other trials: \code{\link{trials_details}}
}

