% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trials_details.R
\name{trials_details}
\alias{trials_details}
\title{trials_details}
\usage{
trials_details(con = NULL, trialDbId = "any", rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{trialDbId}{character; default: 'any'; otherwise an identifier}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists trials_details available on a brapi server
}
\author{
Reinhard Simon
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Trials/GetTrialById.md}(github)
}
\seealso{
Other phenotyping: \code{\link{studies_details}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationlevels}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_observationvariables}},
  \code{\link{studies_search}},
  \code{\link{studies_seasons}},
  \code{\link{studies_studytypes}},
  \code{\link{studies_table}}, \code{\link{trials}}

Other trials: \code{\link{trials}}
}

