% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelematrix_search.R
\name{allelematrix_search}
\alias{allelematrix_search}
\title{marker search}
\usage{
allelematrix_search(markerprofileDbId = 0, markerDbId = 0,
  expandHomozygotes = FALSE, unknownString = "-", sepPhased = "|",
  sepUnphased = "/", format = "json", page = 0, pageSize = 10000,
  method = "GET", rclass = "tibble")
}
\arguments{
\item{markerprofileDbId}{integer; default 0}

\item{markerDbId}{integer; default 0}

\item{expandHomozygotes}{logical; default false}

\item{unknownString}{chaaracter; default: '-'}

\item{sepPhased}{character; default: '|'}

\item{sepUnphased}{character; default: '/'}

\item{format}{character; default: json; other: csv, tsv}

\item{page}{integer; default: 0}

\item{pageSize}{integer; default 1000}

\item{method}{character; web access method. Default: GET; other: POST}

\item{rclass}{character; default: tibble}
}
\value{
data.frame
}
\description{
Lists markers as result of a search.
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/allelematrix_search}
}
\seealso{
Other brapi_call: \code{\link{attributes_categories}},
  \code{\link{attributes}}, \code{\link{calls}},
  \code{\link{crops}}, \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{locations}}, \code{\link{maps}},
  \code{\link{marker_search}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}},
  \code{\link{programs}}, \code{\link{seasons}}

Other genotyping: \code{\link{attributes_categories}},
  \code{\link{attributes}},
  \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{maps}}, \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}}
}

