% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.ba_db.R
\name{as.ba_db}
\alias{as.ba_db}
\title{as.ba_db}
\usage{
as.ba_db(secure = FALSE, protocol = "http://", db = "127.0.0.1",
  port = 80, apipath = NULL, multicrop = FALSE, crop = "", user = "",
  password = "", token = "", granttype = "password",
  clientid = "rbrapi", bms = FALSE)
}
\arguments{
\item{secure}{logical; default: false}

\item{protocol}{character}

\item{db}{character}

\item{port}{integer; default: 80}

\item{apipath}{character; default: ''}

\item{multicrop}{logical; default: false}

\item{crop}{character}

\item{user}{character; default: ''}

\item{password}{character; default: ''}

\item{token}{character; default: ''}

\item{granttype}{character; default: 'password'}

\item{clientid}{character; default: 'rbrapi'}

\item{bms}{logical; default: false}
}
\value{
ba_db class
}
\description{
as.ba_db
}
\examples{
if (interactive()) {

  con <- as.ba_db(
    crop = "potato",
    user = "myself"
    )

  print(con)
}
}
\seealso{
Other brapiutils: \code{\link{ba_can_internet}},
  \code{\link{ba_chart.ba_genomemaps_details}},
  \code{\link{ba_chart.ba_genomemaps}},
  \code{\link{ba_chart.ba_locations}},
  \code{\link{ba_chart}}, \code{\link{ba_check}},
  \code{\link{ba_connect}}, \code{\link{ba_db}},
  \code{\link{ba_describe.ba_locations}},
  \code{\link{ba_describe}}, \code{\link{ba_login}},
  \code{\link{ba_logout}}, \code{\link{is.ba_con}},
  \code{\link{print.ba_con}},
  \code{\link{print.ba_db_list}}
}
\author{
Reinhard Simon, Maikel Verouden
}
