% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.brapi_db.R
\name{as.brapi_db}
\alias{as.brapi_db}
\title{Tas.brapi_db}
\usage{
as.brapi_db(crop = "any", protocol = "http://", db = "127.0.0.1",
  port = 80, multi = FALSE)
}
\arguments{
\item{crop}{character}

\item{protocol}{character}

\item{db}{character}

\item{port}{integer}

\item{multi}{logical}
}
\value{
brapi_db class
}
\description{
Tas.brapi_db
}

