% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_calls.R
\name{ba_calls}
\alias{ba_calls}
\title{ba_calls}
\usage{
ba_calls(con = NULL, datatypes = "csv", pageSize = 50, page = 0,
  rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{datatypes}{string, list of data types}

\item{pageSize}{integer; default = 1000}

\item{page}{integer; default = 0}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists calls available on a brapi server
}
\examples{
if (interactive()) {
  library(brapi)

  # Need to connect to a database with genetic data

  # con <- ba_db()$xxxxx

  # ba_calls(con = con, ...)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Calls/Calls.md}{github}
}
\seealso{
Other brapicore: \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon
}
