% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_can_internet.R
\name{ba_can_internet}
\alias{ba_can_internet}
\title{ba_can_internet}
\usage{
ba_can_internet(url = "www.google.org")
}
\arguments{
\item{url}{a url to check (default \href{https://www.google.org}{www.google.org})}
}
\value{
logical TRUE when internet access is available or FALSE when internet access is not available
}
\description{
Check for internet connection.
}
\examples{
if (interactive()) {
  library(brapi)

  # make sure brapiTS::mock_server() is running in a separate process

  # Checks against a well known and high uptime web site. Default: google

  ba_can_internet()

  # or use an alternative

  ba_can_internet("www.amazon.com")

}
}
\references{
\href{http://stackoverflow.com/questions/5076593/how-to-determine-if-you-have-an-internet-connection-in-r}{Stack Overflow} and  \code{\link[curl]{has_internet}}
}
