% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ba_chart}
\alias{ba_chart}
\title{ba_chart}
\usage{
ba_chart(x, ...)
}
\arguments{
\item{x}{an object to be charted}

\item{...}{additional parameters}
}
\description{
A generic method to chart brapi objects
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)
  library(maps)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  # example 1
  gm <- ba_genomemaps(con)
  gm \%>\% ba_chart

  # example 2
  gmd <- ba_genomemaps_details(con, mapDbId = "1")
  gmd \%>\% ba_chart

  # example 3
  bl <- ba_locations(con)
  bl \%>\% ba_chart()

  # example 4
  bl \%>\% ba_chart(chart_type = "map")
}
}
\seealso{
Other brapiutils: \code{\link{as.ba_db}},
  \code{\link{ba_can_internet}},
  \code{\link{ba_chart.ba_genomemaps_details}},
  \code{\link{ba_chart.ba_genomemaps}},
  \code{\link{ba_chart.ba_locations}},
  \code{\link{ba_check}}, \code{\link{ba_connect}},
  \code{\link{ba_db}},
  \code{\link{ba_describe.ba_locations}},
  \code{\link{ba_describe}}, \code{\link{ba_login}},
  \code{\link{ba_logout}}, \code{\link{is.ba_con}},
  \code{\link{print.ba_con}},
  \code{\link{print.ba_db_list}}
}
\author{
Reinhard Simon
}
