% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_check.R
\name{ba_check}
\alias{ba_check}
\title{ba_check}
\usage{
ba_check(con = NULL, verbose = TRUE, brapi_calls = "any")
}
\arguments{
\item{con}{brapi_connection object}

\item{verbose}{logical; default TRUE}

\item{brapi_calls}{character vector; default: any. Use to check if one or more calls are implemented by the server according to the calls url.}
}
\value{
logical
}
\description{
Checks if a BrAPI server can be reached given the connection details.
}
\details{
Raises errors.
}
\examples{
if (interactive()) {
  library(brapi)

  # make sure you are connected to the internet

  con <- ba_db()$testserver

  ba_check(con)
}
}
\seealso{
Other brapiutils: \code{\link{as.ba_db}},
  \code{\link{ba_can_internet}}, \code{\link{ba_connect}},
  \code{\link{ba_db}}, \code{\link{ba_login}},
  \code{\link{ba_logout}}, \code{\link{is.ba_con}},
  \code{\link{print.ba_con}},
  \code{\link{print.ba_db_list}}
}
\author{
Reinhard Simon
}
