% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_commoncropnames.R
\name{ba_commoncropnames}
\alias{ba_commoncropnames}
\title{ba_commoncropnames}
\usage{
ba_commoncropnames(con = NULL, rclass = c("tibble", "data.frame",
  "list", "json"))
}
\arguments{
\item{con}{list, brapi connection object}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"vector"/"data.frame"}
}
\value{
an object of class as specified by rclass with common crop names or
        NULL
}
\description{
List supported crops in a database.
}
\note{
Tested against: test-server

BrAPI Version: 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <-  ba_db()$testserver

  ba_commoncropnames(con = con)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Crops/CommonCropNames_GET.md}{github}
}
\seealso{
Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search_post}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search_post}},
  \code{\link{ba_programs_search_post}},
  \code{\link{ba_programs}},
  \code{\link{ba_studies_germplasm_details}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{brapicore}
