% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_db.R
\name{ba_db}
\alias{ba_db}
\title{ba_db}
\usage{
ba_db()
}
\value{
list
}
\description{
A white list of known brapi databases.
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)

  # make sure brapiTS::mock_server() is running in a separate process

  bdb <- ba_db()

  print(bdb)

  bdb$mockbase \%>\% print

}
}
\seealso{
Other brapiutils: \code{\link{as.ba_db}},
  \code{\link{ba_can_internet}},
  \code{\link{ba_chart.ba_genomemaps_details}},
  \code{\link{ba_chart.ba_genomemaps}},
  \code{\link{ba_chart.ba_locations}},
  \code{\link{ba_chart}}, \code{\link{ba_check}},
  \code{\link{ba_connect}},
  \code{\link{ba_describe.ba_locations}},
  \code{\link{ba_describe}}, \code{\link{ba_login}},
  \code{\link{ba_logout}}, \code{\link{is.ba_con}},
  \code{\link{print.ba_con}},
  \code{\link{print.ba_db_list}}
}
\author{
Reinhard Simon
}
