% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_describe-methods.R
\name{ba_describe.ba_locations}
\alias{ba_describe.ba_locations}
\title{ba_describe.ba_locations}
\usage{
\method{ba_describe}{ba_locations}(x, ...)
}
\arguments{
\item{x}{a brapi_locations object}

\item{...}{other print parameters}
}
\description{
describe method for an object of class brapi_con, which will only display the crop, database address:port and user
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)

  ba_can_internet()

  con <- ba_db()$sweetpotatobase

  lcs <- ba_locations(con)
  lcs \%>\% ba_describe
}
}
\author{
Reinhard Simon
}
\seealso{
Other brapiutils: \code{\link{as.ba_db}},
  \code{\link{ba_can_internet}},
  \code{\link{ba_chart.ba_genomemaps_details}},
  \code{\link{ba_chart.ba_genomemaps}},
  \code{\link{ba_chart.ba_locations}},
  \code{\link{ba_chart}}, \code{\link{ba_check}},
  \code{\link{ba_connect}}, \code{\link{ba_db}},
  \code{\link{ba_describe}}, \code{\link{ba_login}},
  \code{\link{ba_logout}}, \code{\link{is.ba_con}},
  \code{\link{print.ba_con}},
  \code{\link{print.ba_db_list}}
}

