% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_genomemaps.R
\name{ba_genomemaps}
\alias{ba_genomemaps}
\title{ba_genomemaps}
\usage{
ba_genomemaps(con = NULL, species = "", type = "", page = 0,
  pageSize = 30, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{species}{character; default:}

\item{type}{character, default: ''}

\item{page}{integer; default 0}

\item{pageSize}{integer; default 30}

\item{rclass}{character; default: tibble}
}
\value{
rclass as defined
}
\description{
Get list of maps
}
\note{
This call must have set a specific identifier. The default is an empty string.
     If not changed to an identifier present in the database this will result in an error.
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_genomemaps(con)

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/GenomeMaps/ListOfGenomeMaps.md}{github}
}
\seealso{
Other genomemaps: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_germplasm_attributes}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_attributes}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_markerprofiles_allelematrices_search}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search}}
}
\author{
Reinhard Simon
}
\concept{genomemaps}
\concept{genotyping}
