% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_genomemaps_data.R
\name{ba_genomemaps_data}
\alias{ba_genomemaps_data}
\title{ba_genomemaps_data}
\usage{
ba_genomemaps_data(con = NULL, mapDbId = "", linkageGroupId = NULL,
  linkageGroupName = "", pageSize = 1000, page = 0,
  rclass = c("tibble", "data.frame", "list", "json"))
}
\arguments{
\item{con}{brapi connection object}

\item{mapDbId}{character; default ''}

\item{linkageGroupId}{character; default: NULL}

\item{linkageGroupName}{character; default ''}

\item{pageSize}{integer; default 1000}

\item{page}{integer; default 0}

\item{rclass}{character; default: tibble}
}
\value{
rclass as defined
}
\description{
markers ordered by linkageGroup and position
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_genomemaps_data(con, "gm1")

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/GenomeMaps/GenomeMapData.md}{github}
}
\seealso{
Other genomemaps: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_attributes}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search_post}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_attributes}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_markerprofiles_allelematrices_details}},
  \code{\link{ba_markerprofiles_allelematrices_search}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search}}
}
\author{
Reinhard Simon
}
\concept{genomemaps}
\concept{genotyping}
