% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_genomemaps_data.R
\name{ba_genomemaps_data}
\alias{ba_genomemaps_data}
\title{ba_genomemaps_data}
\usage{
ba_genomemaps_data(con = NULL, mapDbId = "0", linkageGroupId = "0",
  page = 0, pageSize = 30, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{mapDbId}{character; default 0}

\item{linkageGroupId}{character; default 0}

\item{page}{integer; default 0}

\item{pageSize}{integer; default 30}

\item{rclass}{character; default: tibble}
}
\value{
rclass as defined
}
\description{
markers ordered by linkageGroup and position
}
\examples{
if (interactive()) {
  library(brapi)

  # Need to connect to a database with genetic data

  # con <- ba_db()$xxxxx

  # ba_genomemaps(con = con, ...)

}
}
\author{
Reinhard Simon
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/GenomeMaps/GenomeMapData.md}{github}
}
\seealso{
Other genomemaps: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_germplasmattributes}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search}}
}

