% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_genomemaps_data_range.R
\name{ba_genomemaps_data_range}
\alias{ba_genomemaps_data_range}
\title{ba_genomemaps_data_range}
\usage{
ba_genomemaps_data_range(con = NULL, mapDbId = 1, linkageGroupId = 1,
  min = 1, max = 1000, page = 0, pageSize = 30, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{mapDbId}{integer; default 0}

\item{linkageGroupId}{integer; default 0}

\item{min}{integer; default 0}

\item{max}{integer; default 1000}

\item{page}{integer; default 0}

\item{pageSize}{integer; default 30}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists genomemaps_data_range available on a brapi server
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  #TODO

}
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/GenomeMaps/GenomeMapDataByRangeOnLinkageGroup.md}
}
\seealso{
Other genomemaps: \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}}

Other genotyping: \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_germplasmattributes}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search}}
}
\author{
Reinhard Simon
}
