% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_germplasm_attributes.R
\name{ba_germplasm_attributes}
\alias{ba_germplasm_attributes}
\title{ba_germplasm_attributes}
\usage{
ba_germplasm_attributes(con = NULL, germplasmDbId = "",
  attributeList = NULL, attributeDbIds = "", pageSize = 1000,
  page = 0, rclass = c("tibble", "data.frame", "list", "json"))
}
\arguments{
\item{con}{brapi connection object}

\item{germplasmDbId}{character, the internal database identifier for a
germplasm of which the germplasm attribute values are to
be retrieved e.g. "9932"; \strong{REQUIRED ARGUMENT}
with default: ""}

\item{attributeList}{character vector; \strong{DEPRECATED ARGUMENT} with
default: NULL}

\item{attributeDbIds}{character vector, restrict the response to only the
listed internal attribute database identifiers, supplied
as a character vector, e.g. c("123", "234", "345");
default: ""}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "data.frame"/"list"/"json"}
}
\value{
An object of class as defined by rclass containing the germplasm
        attribute values.
}
\description{
Retrieve germplasm attibute values by a specified gerplasm database
identifier.
}
\note{
Tested against: test-server

BrAPI Version: 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_germplasm_attributes(con, germplasmDbId = "1")

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/GermplasmAttributes/Germplasm_Attributes_GET.md}{github}
}
\seealso{
Other germplasmattributes: \code{\link{ba_germplasmattributes_attributes}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search_post}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_attributes}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_markerprofiles_allelematrices_details}},
  \code{\link{ba_markerprofiles_allelematrices_search_post}},
  \code{\link{ba_markerprofiles_allelematrices_search}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search_post}},
  \code{\link{ba_markers_search}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{genotyping}
\concept{germplasmattributes}
