% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_germplasm_attributes.R
\name{ba_germplasm_attributes}
\alias{ba_germplasm_attributes}
\title{ba_germplasm_attributes}
\usage{
ba_germplasm_attributes(con = NULL, germplasmDbId = "",
  attributeList = NULL, attributeDbIds = "", pageSize = 1000,
  page = 0, rclass = c("tibble", "data.frame", "list", "json"))
}
\arguments{
\item{con}{brapi connection object}

\item{germplasmDbId}{character; \strong{REQUIRED ARGUMENT} with default: ''.}

\item{attributeList}{character vector;  \strong{DEPRECATED ARGUMENT} with default: NULL.}

\item{attributeDbIds}{character vector; default: ''}

\item{pageSize}{integer; default: 1000.}

\item{page}{integer; default: 0.}

\item{rclass}{character; default: tibble; or else: json, list, data.frame.}
}
\value{
rclass as defined
}
\description{
germplasm_attibutes call.
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_germplasmattributes_details(con, germplasmDbId = "1")

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/GermplasmAttributes/Germplasm_Attributes_GET.md}{github}
}
\seealso{
Other germplasmattributes: \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_germplasmattributes}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_germplasmattributes}},
  \code{\link{ba_markerprofiles_allelematrices_search}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search}}
}
\author{
Reinhard Simon
}
\concept{genotyping}
\concept{germplasmattributes}
