% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_germplasm_breedingmethods_details.R
\name{ba_germplasm_breedingmethods_details}
\alias{ba_germplasm_breedingmethods_details}
\title{ba_germplasm_breedingmethods_details}
\usage{
ba_germplasm_breedingmethods_details(con = NULL,
  breedingMethodDbId = "", rclass = c("tibble", "data.frame", "list",
  "json"))
}
\arguments{
\item{con}{list, brapi connection object}

\item{breedingMethodDbId}{character, the internal database identifier for a
 breeding method of which the details are to be
retrieved e.g. "bm1"; \strong{REQUIRED ARGUMENT} with
default: ""}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "data.frame"/"list"/"json"}
}
\value{
An object of class as defined by rclass containing the details of the
        requestedd breeding method.
}
\description{
Retrieve details of a breeding method available on a BrAPI compliant database
server.
}
\note{
Tested against: test-server

BrAPI Version: 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_germplasm_breedingmethods_details(con, "bm1")


}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Germplasm/BreedingMethods_BreedingMethodDbId_GET.md}{github}
}
\author{
Reinhard Simon, Maikel Verouden
}
