% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_germplasm_details.R
\name{ba_germplasm_details}
\alias{ba_germplasm_details}
\title{ba_germplasm_details}
\usage{
ba_germplasm_details(con = NULL, germplasmDbId = "",
  rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{germplasmDbId}{string; default 0; an internal ID for a germplasm}

\item{rclass}{character; tibble}
}
\value{
list
}
\description{
Gets germplasm details given an id.
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_germplasm_details(con, "1")
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Germplasm/GermplasmDetailsByGermplasmDbId.md}{github}
}
\seealso{
Other germplasm: \code{\link{ba_germplasm_details_study-deprecated}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search}},
  \code{\link{brapi-deprecated}}

Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_crops-deprecated}},
  \code{\link{ba_germplasm_details_study-deprecated}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}},
  \code{\link{brapi-deprecated}}
}
\author{
Reinhard Simon
}
\concept{brapicore}
\concept{germplasm}
