% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_germplasm_details_study.R
\name{ba_germplasm_details_study}
\alias{ba_germplasm_details_study}
\title{ba_germplasm_details_study}
\usage{
ba_germplasm_details_study(con = NULL, studyDbId = 123, page = 0,
  pageSize = 1000, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{studyDbId}{character; default: 123}

\item{page}{integer; default 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists studies_germplasm available on a brapi server
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  ba_germplasm_details_study(con)

}
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Germplasm/GermplasmDetailsListByStudyDbId.md}
}
\seealso{
Other germplasm: \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_search}}

Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon
}
