% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_germplasm_details_study.R
\name{ba_germplasm_details_study}
\alias{ba_germplasm_details_study}
\title{ba_germplasm_details_study}
\usage{
ba_germplasm_details_study(con = NULL, studyDbId = "",
  pageSize = 1000, page = 0, rclass = "tibble")
}
\arguments{
\item{con}{list, brapi connection object}

\item{studyDbId}{character, the internal database identifier for a study of
which the germplasm details are to be retrieved e.g. "1001";
\strong{REQUIRED ARGUMENT} with default: ""}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"data.frame"}
}
\value{
An object of class as defined by rclass containing the germplasm
        details for a requested study.
}
\description{
Gets germplasm details for a specific study available on a brapi server.
}
\note{
Tested against: sweetpotatobase, test-server

BrAPI Version: 1.0

BrAPI Status: deprecated
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_germplasm_details_study(con, "1001")
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.0/Specification/Germplasm/GermplasmDetailsListByStudyDbId.md}{github}
}
\seealso{
Other germplasm: \code{\link{ba_germplasm_breedingmethods}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_studies_germplasm_details}}

Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_commoncropnames}}, \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_studies_germplasm_details}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon
}
\concept{brapicore}
\concept{germplasm}
