% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_germplasm_markerprofiles.R
\name{ba_germplasm_markerprofiles}
\alias{ba_germplasm_markerprofiles}
\title{ba_germplasm_markerprofiles}
\usage{
ba_germplasm_markerprofiles(con = NULL, germplasmDbId = 3,
  rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{germplasmDbId}{integer}

\item{rclass}{character, default: list; alternative: vector}
}
\value{
list of marker profile ids
}
\description{
Gets minimal marker profile data from database using database internal id
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  #TODO

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Germplasm/GermplasmMarkerprofiles.md}{github}
}
\seealso{
Other germplasm: \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_search}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_germplasmattributes}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search}}
}
\author{
Reinhard Simon
}
