% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_germplasm_pedigree.R
\name{ba_germplasm_pedigree}
\alias{ba_germplasm_pedigree}
\title{ba_germplasm_pedigree}
\usage{
ba_germplasm_pedigree(con = NULL, germplasmDbId = 0, notation = "purdue",
  rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{germplasmDbId}{integer; default 0}

\item{notation}{string; default: purdue format}

\item{rclass}{string; default: tibble}
}
\value{
list of pedigree data
}
\description{
Gets minimal pedigree data from database using database internal id
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  #TODO

}
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Germplasm/GermplasmPedigree.md}(github)
}
\seealso{
Other germplasm: \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search}}

Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon
}
