% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_germplasm_pedigree.R
\name{ba_germplasm_pedigree}
\alias{ba_germplasm_pedigree}
\title{ba_germplasm_pedigree}
\usage{
ba_germplasm_pedigree(con = NULL, germplasmDbId = "", notation = "",
  rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{germplasmDbId}{character; default ''}

\item{notation}{character; optional, recommended value: purdue format. Default: ''}

\item{rclass}{character; default: tibble}
}
\value{
list of pedigree data
}
\description{
Gets minimal pedigree data from database using database internal id
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_germplasm_pedigree(con, germplasmDbId = "1")

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Germplasm/GermplasmPedigree.md}{github}
}
\seealso{
Other germplasm: \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search}}

Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_commoncropnames}}, \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon
}
\concept{brapicore}
\concept{germplasm}
