% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_germplasm_progeny.R
\name{ba_germplasm_progeny}
\alias{ba_germplasm_progeny}
\title{ba_germplasm_progeny}
\usage{
ba_germplasm_progeny(con = NULL, germplasmDbId = "",
  rclass = c("tibble", "data.frame", "list", "json"))
}
\arguments{
\item{con}{list, brapi connection object}

\item{germplasmDbId}{character, the internal database identifier for a
germplasm of which the internal germaplasm database
identifiers of all the progeny are to be retrieved e.g.
"9932"; \strong{REQUIRED ARGUMENT} with default: ""}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "data.frame"/"list"/"json"}
}
\value{
An object of class as defined by rclass containing the internal
        germplasm database identifiers for all the progeny.
}
\description{
Retrieve the internal germplasm database identifiers for all the progeny of
a particular internal germplasm database identifier.
}
\note{
Tested against: test-server

BrAPI Version: 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_germplasm_progeny(con, germplasmDbId = "1")

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Germplasm/Germplasm_Progeny_GET.md}{github}
}
\seealso{
Other germplasm: \code{\link{ba_germplasm_breedingmethods}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_search_post}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_studies_germplasm_details}}

Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_commoncropnames}}, \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_search_post}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search_post}},
  \code{\link{ba_programs_search_post}},
  \code{\link{ba_programs}},
  \code{\link{ba_studies_germplasm_details}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{brapicore}
\concept{germplasm}
