% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_germplasmattributes_attributes.R
\name{ba_germplasmattributes_attributes}
\alias{ba_germplasmattributes_attributes}
\title{ba_germplasmattributes_attributes}
\usage{
ba_germplasmattributes_attributes(con = NULL,
  attributeCategoryDbId = "", pageSize = 1000, page = 0,
  rclass = c("tibble", "data.frame", "list", "json"))
}
\arguments{
\item{con}{list, brapi connection object}

\item{attributeCategoryDbId}{character, optional, default: ''}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"vector"/"data.frame"}
}
\value{
as defined by rclass
}
\description{
List supported crops in a database.
}
\note{
Tested against: test-server, sweetpotatobsse

BrAPI Version: 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_germplasmattributes_attributes(con, "1")
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/GermplasmAttributes/Attributes_GET.md}{github}
}
\seealso{
Other germplasmattributes: \code{\link{ba_germplasm_attributes}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_attributes}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search_post}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_markerprofiles_allelematrices_details}},
  \code{\link{ba_markerprofiles_allelematrices_search}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search_post}},
  \code{\link{ba_markers_search}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{genotyping}
\concept{germplasmattributes}
