% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_germplasmattributes_details.R
\name{ba_germplasmattributes_details}
\alias{ba_germplasmattributes_details}
\title{ba_germplasmattributes_details}
\usage{
ba_germplasmattributes_details(con = NULL, germplasmDbId = "1",
  attributeList = "", page = 0, pageSize = 10, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{germplasmDbId}{character; default: 1.}

\item{attributeList}{character vector; default: ''.}

\item{page}{integer; default: 0.}

\item{pageSize}{integer; default: 10.}

\item{rclass}{character; default: tibble; or else: json, list, data.frame.}
}
\value{
rclass as set by parameter.
}
\description{
attibutes call.
}
\examples{
if (interactive()) {
  library(brapi)
  # Need to connect to a database with genetic data

  # con <- ba_db()$xxxxx

  # ba_germplasmattributes_details(con = con, ...)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/GermplasmAttributes/GermplasmAttributeValuesByGermplasmDbId.md}{github}
}
\seealso{
Other germplasmattributes: \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search}}
}
\author{
Reinhard Simon
}
