% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_locations.R
\name{ba_locations}
\alias{ba_locations}
\title{ba_locations}
\usage{
ba_locations(con = NULL, locationType = "all", pageSize = 1000,
  page = 0, rclass = "tibble")
}
\arguments{
\item{con}{list, brapi connection object}

\item{locationType}{character, filter locotion type specified.}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"data.frame"}
}
\value{
An object of class as defined by rclass containing locations.
}
\description{
Lists locations available on a brapi server.
}
\details{
All standard attributes are always listed. However, attributes in the additionalInfo
only when at least one record has data.
}
\note{
Tested against: BMS, sweetpotatobase, test-server

BrAPI Version: 1.0, 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)
  # Need to connect to a database with genetic data

  # con <- ba_db()$xxxxx

  # ba_locations(con = con, ...)

}
}
\references{
\href{https://github.com/plantbreeding/API/tree/master/Specification/Locations}{github}
}
\author{
Reinhard Simon, Maikel Verouden
}
