% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_locations.R
\name{ba_locations}
\alias{ba_locations}
\title{ba_locations}
\usage{
ba_locations(con = NULL, locationType = "all", page = 0,
  pageSize = 1000, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{locationType}{character, list of data types}

\item{page}{integer; default 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{character; default: tibble}
}
\value{
rclass as defined
}
\description{
Lists locations available on a brapi server.
}
\note{
All standard attributes are always listed. However, attributes in the additionalInfo only when at least one record has data.
}
\examples{
if (interactive()) {
  library(brapi)
  # Need to connect to a database with genetic data

  # con <- ba_db()$xxxxx

  # ba_locations(con = con, ...)

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Locations/ListLocations.md}{github}
}
\author{
Reinhard Simon
}
