% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_markerprofiles_allelematrices_details.R
\name{ba_markerprofiles_allelematrices_details}
\alias{ba_markerprofiles_allelematrices_details}
\title{ba_markerprofiles_allelematrices_details}
\usage{
ba_markerprofiles_allelematrices_details(con = NULL, studyDbId = "",
  pageSize = 1000, page = 0, rclass = c("tibble", "data.frame",
  "list", "json"))
}
\arguments{
\item{con}{brapi connection object}

\item{studyDbId}{character, \strong{REQUIRED ARGUMENT} with default: ""}

\item{pageSize}{integer; default 1000}

\item{page}{integer; default: 0}

\item{rclass}{character; default: tibble}
}
\value{
rclass as requested
}
\description{
This resource is used for reading and writing genomic matrices
}
\note{
The handling of long-running responses via asynch status messages is not yet implemented.
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_markerprofiles_allelematrices_details(con, studyDbId = "1001")
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/MarkerProfiles/AlleleMatrices_GET.md}{github}
}
\seealso{
Other markerprofiles: \code{\link{ba_markerprofiles_allelematrices_search}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_attributes}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_attributes}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_markerprofiles_allelematrices_search}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search}}
}
\author{
Reinhard Simon
}
\concept{genotyping}
\concept{markerprofiles}
