% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_markerprofiles_details.R
\name{ba_markerprofiles_details}
\alias{ba_markerprofiles_details}
\title{ba_markerprofiles_details}
\usage{
ba_markerprofiles_details(con = NULL, markerprofilesDbId = "",
  expandHomozygotes = FALSE, unknownString = "N", sepPhased = "|",
  sepUnphased = "/", page = 0, pageSize = 10000, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{markerprofilesDbId}{character; default 0}

\item{expandHomozygotes}{logical; default false}

\item{unknownString}{chaaracter; default: '-'}

\item{sepPhased}{character; default: '|'}

\item{sepUnphased}{character; default: '/'}

\item{page}{integer; default: 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{character; default: tibble}
}
\value{
data.frame
}
\description{
Lists markers as result of a search.
}
\examples{
if (interactive()) {
  library(brapi)

  # Need to connect to a database with genetic data

  # con <- ba_db()$xxxxx

  # ba_markerprofiles_details(con = con, ...)
}
}
\author{
Reinhard Simon
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/MarkerProfiles/MarkerProfileData.md}{github}
}
\seealso{
Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_germplasmattributes}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}},
  \code{\link{ba_markers_search}}

Other markerprofiles: \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_search}}
}

