% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_markers_details.R
\name{ba_markers_details}
\alias{ba_markers_details}
\title{ba_markers_details}
\usage{
ba_markers_details(con = NULL, markerDbId = 0, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{markerDbId}{integer; marker id; default: 0}

\item{rclass}{character; default: tibble}
}
\value{
data.frame
}
\description{
Lists markers as result of a search.
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  ba_markers_details(con, 3)

}
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Markers/MarkerDetailsByMarkerDbId.md}
}
\seealso{
Other markers: \code{\link{ba_markers_search}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_germplasmattributes}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_search}}
}
\author{
Reinhard Simon
}
