% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_markers_search.R
\name{ba_markers_search}
\alias{ba_markers_search}
\title{ba_markers_search}
\usage{
ba_markers_search(con = NULL, markerDbIds = "", name = "",
  matchMethod = "", includeSynonyms = TRUE, type = "",
  pageSize = 1000, page = 0, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{markerDbIds}{character, optional, default: ''}

\item{name}{character; marker name; default: ''}

\item{matchMethod}{character; default: ''; other: case_insensitive, exact, wildcard}

\item{includeSynonyms}{logical, default: TRUE}

\item{type}{character; default: all; other: SNP}

\item{pageSize}{integer; default 1000}

\item{page}{integer; default: 0}

\item{rclass}{character; default: tibble}
}
\value{
rclass as requested
}
\description{
Lists markers as result of a search.
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_markers_search(con)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Markers/MarkerSearch_GET.md}{github}
}
\seealso{
Other markers: \code{\link{ba_markers_details}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_attributes}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_attributes}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_markerprofiles_allelematrices_search}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}}
}
\author{
Reinhard Simon
}
\concept{genotyping}
\concept{markers}
