% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_markers_search.R
\name{ba_markers_search}
\alias{ba_markers_search}
\title{ba_markers_search}
\usage{
ba_markers_search(con = NULL, name = "*", type = "all",
  matchMethod = "wildcard", include = "synonyms", page = 0,
  pageSize = 1000, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{name}{character; marker name; default: *}

\item{type}{character; default: all; other: SNP}

\item{matchMethod}{character; default: wildcard; other: case_insensitive, exact}

\item{include}{character; default: synonyms}

\item{page}{integer; default: 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{character; default: tibble}
}
\value{
data.frame
}
\description{
Lists markers as result of a search.
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_markers_search(con)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Markers/MarkerSearch.md}{github}
}
\seealso{
Other markers: \code{\link{ba_markers_details}}

Other genotyping: \code{\link{ba_genomemaps_data_range}},
  \code{\link{ba_genomemaps_data}},
  \code{\link{ba_genomemaps_details}},
  \code{\link{ba_genomemaps}},
  \code{\link{ba_germplasm_attributes}},
  \code{\link{ba_germplasm_markerprofiles}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_germplasmattributes_categories}},
  \code{\link{ba_germplasmattributes_details}},
  \code{\link{ba_germplasmattributes}},
  \code{\link{ba_markerprofiles_allelematrices_search}},
  \code{\link{ba_markerprofiles_allelematrix_search}},
  \code{\link{ba_markerprofiles_details}},
  \code{\link{ba_markerprofiles_search}},
  \code{\link{ba_markers_details}}
}
\author{
Reinhard Simon
}
