% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_observationvariables.R
\name{ba_observationvariables}
\alias{ba_observationvariables}
\title{ba_observationvariables}
\usage{
ba_observationvariables(con = NULL, traitClass = "all", page = 0,
  pageSize = 1000, rclass = "tibble")
}
\arguments{
\item{con}{object; brapi connection object}

\item{traitClass}{string; default 'all'}

\item{page}{integer; default 0}

\item{pageSize}{integer; defautlt 1000}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists brapi_variables available on a brapi server
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  ba_observationvariables(con)

}
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/ObservationVariables/VariableList.md}
}
\seealso{
Other observationvariables: \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}}

Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon
}
