% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_observationvariables_details.R
\name{ba_observationvariables_details}
\alias{ba_observationvariables_details}
\title{ba_observationvariables_details}
\usage{
ba_observationvariables_details(con = NULL,
  observationVariableDbId = "MO_123:0100621", rclass = "tibble")
}
\arguments{
\item{con}{object; brapi connection object}

\item{observationVariableDbId}{string; default 'CO_334:0100622'; from test server.}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists brapi_variables_details available on a brapi server
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  #TODO

}
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/ObservationVariables/VariableDetails.md}
}
\seealso{
Other observationvariables: \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables}}

Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon
}
