% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_observationvariables_ontologies.R
\name{ba_observationvariables_ontologies}
\alias{ba_observationvariables_ontologies}
\title{ba_observationvariables_ontologies}
\usage{
ba_observationvariables_ontologies(con = NULL, page = 0, pageSize = 1000,
  rclass = "tibble")
}
\arguments{
\item{con}{list; brapi connection object; default = NULL}

\item{page}{integer; default 0}

\item{pageSize}{integer; defautlt 1000}

\item{rclass}{character; default: tibble}
}
\value{
rclass as defined
}
\description{
lists variables_datatypes available on a brapi server
}
\examples{
if (interactive()) {
  library(brapi)

  # Need to connect to a database with genetic data

  # con <- ba_db()$xxxxx

  # ba_observationvariables_details(con = con, ...)
}
}
\author{
Reinhard Simon
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/ObservationVariables/VariableOntologyList.md}{github}
}
\seealso{
Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}

Other observationvariables: \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables}}
}

