% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_observationvariables_search.R
\name{ba_observationvariables_search}
\alias{ba_observationvariables_search}
\title{ba_observationvariables_search}
\usage{
ba_observationvariables_search(con = NULL, datatypes = "",
  methodDbIds = "", names = "", observationVariableDbIds = "",
  ontologyXrefs = "", ontologyDbIds = "", page = 0,
  pageSize = 1000, scaleDbIds = "", traitClasses = "",
  rclass = "tibble")
}
\arguments{
\item{con}{brapi list; connection object}

\item{datatypes}{vector of character; default: ''}

\item{methodDbIds}{vector of character; default: ''}

\item{names}{vector of character; default: ''}

\item{observationVariableDbIds}{vector of character; default: ''}

\item{ontologyXrefs}{vector of character; default: ''}

\item{ontologyDbIds}{vector of character; default: ''}

\item{page}{integer; default: 0}

\item{pageSize}{integer, default: 1000}

\item{scaleDbIds}{vector of character; default: ''}

\item{traitClasses}{vector of character; default: ''}

\item{rclass}{character; default: tibble}
}
\value{
rclass
}
\description{
lists the breeding observationvariables
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_observationvariables_search(con)

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/ObservationVariables/VariableSearch.md}{github}
}
\seealso{
Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_commoncropnames}},
  \code{\link{ba_crops-deprecated}},
  \code{\link{ba_germplasm_details_study-deprecated}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon
}
\concept{brapicore}
