% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_programs.R
\name{ba_programs}
\alias{ba_programs}
\title{ba_programs}
\usage{
ba_programs(con = NULL, programName = "", abbreviation = "",
  commonCropName = "", pageSize = 1000, page = 0,
  rclass = "tibble")
}
\arguments{
\item{con}{list, brapi connection object}

\item{programName}{character, filter programs by program name; default: ""}

\item{abbreviation}{character, filter programs by program abbreviation;
default: ""}

\item{commonCropName}{character, filter programs by a common crop name;
default: ""}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"data.frame"}
}
\value{
An object of class as defined by rclass containing the breeding
        programs.
}
\description{
Retrieve a list of the breeding programs.
}
\details{
Filtering by commonCropName has not be tested yet, because none of
         the databases support this feature yet (not even the test-server)!
}
\note{
Tested against: BMS, sweetpotatobase, test-server

BrAPI Version: 1.0, 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_programs(con)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Programs/ListPrograms.md}{github}
}
\seealso{
Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_commoncropnames}}, \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_studies_germplasm_details}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{brapicore}
