% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_programs.R
\name{ba_programs}
\alias{ba_programs}
\title{ba_programs}
\usage{
ba_programs(con = NULL, programName = "any", abbreviation = "any",
  page = 0, pageSize = 1000, rclass = "tibble")
}
\arguments{
\item{con}{list, brapi connection object}

\item{programName}{character; default: any}

\item{abbreviation}{character; default: any}

\item{page}{integer requested page number, default = 0 (1st page)}

\item{pageSize}{integer, items per page (default = 1000)}

\item{rclass}{character; default: "tibble" possible other values: "json"/"list"/"data.frame"}
}
\value{
rclass
}
\description{
lists the breeding programs
}
\details{
V1.2: same
}
\note{
Tested against: sweetpotatobase, BMS
}
\examples{
if (interactive()) {
  library(brapi)

  # Need to connect to a database with genetic data

  # con <- ba_db()$xxxxx

  # ba_programs(con = con, ...)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Programs/ListPrograms.md}{github}
}
\seealso{
Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon, Maikel Verouden
}
