% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_programs_search.R
\name{ba_programs_search}
\alias{ba_programs_search}
\title{ba_programs_search}
\usage{
ba_programs_search(con = NULL, abbreviation = "", leadPerson = "",
  name = "", objective = "", page = 0, pageSize = 1000,
  programDbId = "", rclass = "tibble")
}
\arguments{
\item{con}{list, brapi connection object}

\item{abbreviation}{character; default: ''}

\item{leadPerson}{character; default: ''}

\item{name}{character; default: ''}

\item{objective}{character; default: ''}

\item{page}{integer; default: 0}

\item{pageSize}{integer, default: 1000}

\item{programDbId}{character; default: ''}

\item{rclass}{character; default: tibble}
}
\value{
rclass
}
\description{
searches the breeding programs
}
\examples{
if (interactive()) {
  library(brapi)

  # Need to connect to a database with genetic data

  # con <- ba_db()$xxxxx

  # ba_programs_search(con = con, ...)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Programs/ProgramSearch.md}{github}
}
\seealso{
Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon
}
