% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_programs_search.R
\name{ba_programs_search}
\alias{ba_programs_search}
\title{ba_programs_search}
\usage{
ba_programs_search(con = NULL, abbreviation = "", leadPerson = "",
  name = "", objective = "", programDbId = "", commonCropName = "",
  pageSize = 1000, page = 0, rclass = "tibble")
}
\arguments{
\item{con}{list, brapi connection object}

\item{abbreviation}{character, search programs by program abbreviation;
default: ""}

\item{leadPerson}{character, search programs by program leader name; default:
""}

\item{name}{character, search programs by program name; default: ""}

\item{objective}{character, search programs by program objective, default: ""}

\item{programDbId}{character, search programs by program database identifier
(programDbId), default: ""}

\item{commonCropName}{character, search programs by commonCropName; default:
""}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"data.frame"}
}
\value{
An object of class as defined by rclass containing the breeding
        programs matching the search criteria.
}
\description{
Search breeding programs via a POST method
}
\details{
Testing against test-server always returns all programs in the test
         server example. One would expect info on only one program, when
         programDbId = "1" is specified as function argument. Could be an
         implementation error of the BrAPI call on the database side.
}
\note{
Tested against: test-server (unexpected results, see details section)

BrAPI Version: 1.0, 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  # Need to connect to a database with genetic data

  # con <- ba_db()$xxxxx

  # ba_programs_search(con = con, ...)
}
}
\references{
\href{https://github.com/plantbreeding/API/tree/master/Specification/Programs}{github}
}
\seealso{
Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_commoncropnames}}, \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_traits_details}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon, Maikel Verouden
}
