% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_samples.R
\name{ba_samples}
\alias{ba_samples}
\title{ba_samples}
\usage{
ba_samples(con = NULL, sampleDbId = "", rclass = c("tibble",
  "data.frame", "list", "json"))
}
\arguments{
\item{con}{list, brapi connection object}

\item{sampleDbId}{character, mandatory argument}

\item{rclass}{character; default: "tibble" possible other values: "json"/"list"/"data.frame"}
}
\value{
rclass as defined
}
\description{
Get sample metadata available on a brapi server
}
\note{
Tested against: BMS
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_samples(con, sampleDbId = "sam1")
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Samples/Samples_GET.md}{github}
}
\seealso{
Other phenotyping: \code{\link{ba_phenotypes_search}},
  \code{\link{ba_samples_save}},
  \code{\link{ba_samples_search_post}},
  \code{\link{ba_samples_search}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{phenotyping}
