% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_samples.R
\name{ba_samples}
\alias{ba_samples}
\title{ba_samples}
\usage{
ba_samples(con = NULL, sampleId = "", rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{sampleId}{character}

\item{rclass}{character; default: tibble}
}
\value{
rclass as defined
}
\description{
Get sample metadata available on a brapi server
}
\examples{
if (interactive()) {
  library(brapi)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  sampleId <- "Unique-Plant-SampleId-1234567890"

  # Example 1
  bl <- ba_samples(con, sampleId)

}
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Samples/RetrieveSampleMetadata.md}(github)
}
\seealso{
Other phenotyping: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon
}
