% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_samples_save.R
\name{ba_samples_save}
\alias{ba_samples_save}
\title{ba_samples_save}
\usage{
ba_samples_save(con = NULL, sampleData = list())
}
\arguments{
\item{con}{list, brapi connection object}

\item{sampleData}{list of sample details.}
}
\value{
character a unique sample ID assigned by the server
}
\description{
Save sample data available on a brapi server.
}
\examples{
if (interactive()) {
  library(brapi)

  # Need to connect to a database with genetic data

  # con <- ba_db()$xxxxx

  # ba_samples_save(con = con, ...)

}
}
\author{
Reinhard Simon
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Samples/TakeASample.md}{github}
}
\seealso{
Other phenotyping: \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}

