% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_details.R
\name{ba_studies_details}
\alias{ba_studies_details}
\title{ba_studies_details}
\usage{
ba_studies_details(con = NULL, studyDbId = "", rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{studyDbId}{character; default ''; an internal ID for a study}

\item{rclass}{character; tibble}
}
\value{
data.frame
}
\description{
Gets studies details given an id.
}
\note{
This call must have set a specific identifier. The default is an empty string.
     If not changed to an identifier present in the database this will result in an error.
}
\examples{
if (interactive()) {
  library(brapi)

  # Need to connect to a database with genetic data

  # con <- ba_db()$xxxxx

  # ba_studies_details(con = con, ...)

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Studies/StudyDetails.md}{github}
}
\seealso{
Other studies: \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}}

Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon, Maikel Verouden
}
