% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_details.R
\name{ba_studies_details}
\alias{ba_studies_details}
\title{ba_studies_details}
\usage{
ba_studies_details(con = NULL, studiesDbId = "0", rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{studiesDbId}{character; default 0; an internal ID for a studiess}

\item{rclass}{character; tibble}
}
\value{
data.frame
}
\description{
Gets studies details given an id.
}
\examples{
if (interactive()) {
  library(brapi)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  ba_studies_details(con, studiesDbId = "1")

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Studies/StudyDetails.md}{github}
}
\seealso{
Other studies: \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}}

Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon
}
