% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_layout.R
\name{ba_studies_layout}
\alias{ba_studies_layout}
\title{ba_studies_layout}
\usage{
ba_studies_layout(con = NULL, studyDbId = "1", rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{studyDbId}{string; default: 1}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists studies_layout available on a brapi server
}
\examples{
if (interactive()) {
  library(brapi)


  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  ba_studies_layout(con, studyDbId = "1")

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Studies/PlotLayoutDetails.md}{github}
}
\seealso{
Other studies: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table}}

Other phenotyping: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon
}
