% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_observationlevels.R
\name{ba_studies_observationlevels}
\alias{ba_studies_observationlevels}
\title{ba_studies_observationlevels}
\usage{
ba_studies_observationlevels(con = NULL, pageSize = 1000, page = 0,
  rclass = c("tibble", "data.frame", "list", "vector", "json"))
}
\arguments{
\item{con}{list, brapi connection object}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{rclass}{character, class of the object to be returned; default:
"tibble", possible other values: "data.frame"/"list"/"vector"/
"json"}
}
\value{
An object of class as defined by rclass containing the details of the
        measured observation variables for a requested study.
}
\description{
Retrieve the list of supported observation levels.
}
\details{
Observation levels indicate the granularity level at which the
         measurements are taken. The values are used to supply the
         observationLevel parameter in the observation unit details call.
}
\note{
Tested against: sweetpotatobase, test-server

BrAPI Version: 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_studies_observationlevels(con)

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Studies/ObservationLevels_GET.md}{github}
}
\seealso{
Other studies: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_germplasm_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_get}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}}

Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_get}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{phenotyping}
\concept{studies}
