% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_observationunits.R
\name{ba_studies_observationunits}
\alias{ba_studies_observationunits}
\title{ba_studies_observationunits}
\usage{
ba_studies_observationunits(con = NULL, studyDbId = "",
  observationLevel = "plot", rclass = "tibble")
}
\arguments{
\item{con}{list, brapi connection object}

\item{studyDbId}{character; default: ''}

\item{observationLevel}{character; default: plot; alternative: plant}

\item{rclass}{character; default: tibble}
}
\value{
rclass as defined
}
\description{
lists studies_observationunits available on a brapi server
}
\note{
This call must have set a specific identifier. The default is an empty string.
     If not changed to an identifier present in the database this will result in an error.
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_studies_observationunits(con, studyDbId = "1001")
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Studies/ObservationUnitDetails.md}{github}
}
\seealso{
Other studies: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_germplasm_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationVariables}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_get}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}}

Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationVariables}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_get}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon
}
\concept{phenotyping}
\concept{studies}
