% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_observationunits_save.R
\name{ba_studies_observationunits_save}
\alias{ba_studies_observationunits_save}
\title{ba_studies_observationunits_save}
\usage{
ba_studies_observationunits_save(con = NULL, studyDbId = "",
  unitData = NULL, observationLevel = "plot", transactionDbId = "",
  commit = TRUE)
}
\arguments{
\item{con}{brapi connection object}

\item{studyDbId}{character a unique study ID}

\item{unitData}{data.frame or tibble observation unit data}

\item{observationLevel}{character plot (default) or plant}

\item{transactionDbId}{character string}

\item{commit}{logical TRUE (default)}
}
\value{
character a unique sample ID assigned by the server
}
\description{
Call to invoke for saving the measurements (observations) collected from field for all the observation units.
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  # ba_studies_observationunits_save(con, ...)

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Studies/SaveOrUpdateObservationUnits.md}{github}
}
\seealso{
Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_search_get}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon
}
\concept{phenotyping}
