% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_search.R
\name{ba_studies_search}
\alias{ba_studies_search}
\title{ba_studies_search}
\usage{
ba_studies_search(con = NULL, studyType = "", programDbId = "",
  locationDbId = "", seasonDbId = "", germplasmDbIds = "",
  observationVariableDbIds = "", active = "", sortBy = "",
  sortOrder = "", page = 0, pageSize = 1000, verb = "GET",
  rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{studyType}{character; default: ''}

\item{programDbId}{character; default: ''}

\item{locationDbId}{character; default: ''}

\item{seasonDbId}{character; default: ''}

\item{germplasmDbIds}{character; default: ''}

\item{observationVariableDbIds}{character; default: ''}

\item{active}{character; default: ''}

\item{sortBy}{character; default: ''}

\item{sortOrder}{character; default: ''}

\item{page}{integer; default: 1000}

\item{pageSize}{integer; default: 0}

\item{verb}{character; default: 'GET' (or 'POST')}

\item{rclass}{character; default: tibble}
}
\value{
rclass as defined
}
\description{
lists trials available on a brapi server
}
\note{
This call must have set a specific identifier. The default is an empty string.
     If not changed to an identifier present in the database this will result in an error.

Tested against: sweetpotatobase
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_studies_search(con)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Studies/ListStudySummaries.md}{github}
}
\seealso{
Other studies: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}}

Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon
}
