% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_search.R
\name{ba_studies_search}
\alias{ba_studies_search}
\title{ba_studies_search}
\usage{
ba_studies_search(con = NULL, studyType = "any", programDbId = "any",
  locationDbId = "any", seasonDbId = "any", germplasmDbIds = "any",
  observationVariableDbIds = "any", active = "any", sortBy = "any",
  sortOrder = "any", page = 0, pageSize = 1000, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{studyType}{string; default: any}

\item{programDbId}{string; default: any}

\item{locationDbId}{string; default: any}

\item{seasonDbId}{string; default: any}

\item{germplasmDbIds}{string; default: any}

\item{observationVariableDbIds}{string; default: any}

\item{active}{string; default: any}

\item{sortBy}{string; default: any}

\item{sortOrder}{string; default: any}

\item{page}{integer; default: 1000}

\item{pageSize}{integer; default: 0}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists trials available on a brapi server
}
\examples{
if (interactive()) {
  library(brapi)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  ba_studies_search(con)

}
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Studies/ListStudySummaries.md}(github)
}
\seealso{
Other studies: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}}

Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon
}
