% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_search_post.R
\name{ba_studies_search_post}
\alias{ba_studies_search_post}
\title{ba_studies_search_post}
\usage{
ba_studies_search_post(con = NULL, studyDbIds = "", trialDbIds = "",
  programDbIds = "", locationDbIds = "", seasonDbId = "",
  studyType = "", studyNames = "", studyLocations = "",
  programNames = "", commonCropName = "", germplasmDbIds = "",
  observationVariableDbIds = "", active = NA, sortBy = "",
  sortOrder = "", page = 0, pageSize = 1000, rclass = c("tibble",
  "data.frame", "list", "json"))
}
\arguments{
\item{con}{list, brapi connection object}

\item{studyDbIds}{character vector, search for study details of specified
studies, supplied as a comma separated character vector of
internal study database identifiers e.g. c("35", "345");
default: ""}

\item{trialDbIds}{character vector, search for studies and  details of
specified trials, supplied as a comma separated character
vector of internal trial database identifiers e.g.
c("7", "8"); default: ""}

\item{programDbIds}{character vector, search for studies and details of
specified programs, supplied as a comma separated
character vector of internal program database identifiers
e.g. c("27", "58"); default: ""}

\item{locationDbIds}{character vector, search for studies and details of
specified locations, supplied as a comma separated
character vector of internal location database
identifiers e.g. c("23", "33"); default:""}

\item{seasonDbId}{character vector, search for studies and details by
specified seasons, supplied as a comma separated character
vector of internal season database identifiers e.g.
c("237", "238"); default: ""}

\item{studyType}{character; search for studies and details based on a study
type e.g. "Nursery", or "Yield Trial"; default: ""}

\item{studyNames}{character vector, search for studies by study names,
supplied as a comma separated character vector of study
names e.g. c("Study A", "Study B"); default: ""}

\item{studyLocations}{character vector; search for studies by study
locations, supplied as a comma separated character
vector of study locations e.g. c("Kenya", "Zimbabwe");
default: ""}

\item{programNames}{character vector; search for studies by program names,
 supplied as a comma separated character vector of program
 names e.g. c("Test Program", "Program2");
default: ""}

\item{commonCropName}{character, search for studies by a common crop name
e.g. "wheat"; default: ""}

\item{germplasmDbIds}{character vector; search for studies where specified
germplasms, supplied as a comma separated character
vector of internal gerplasm database identifiers e.g.
c("CML123", "CWL123"), have been used/tested; default:
""}

\item{observationVariableDbIds}{character vector; search for studies where
specified observation variables, supplied as
a comma separated character vector of
internal observation variable database
identifiers e.g. c("CO-PH-123", "Var-123"),
have been measured; default: ""}

\item{active}{logical; search studies by active status (by default both
active and inactive studies are shown); default: NA,other
possible values TRUE (show active)/FALSE (show inactive)}

\item{sortBy}{character; name of the field to sort by e.g. "studyDbId";
default: ""}

\item{sortOrder}{character; sort order direction; default: "", possible values
"asc"/"desc"}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"data.frame"}
}
\value{
An object of class as defined by rclass containing the studies and
        details fulfilling the search criteria.
}
\description{
Search for study details on a brapi server via a POST method.
}
\note{
Tested against: test-server

BrAPI Version: 1.0, 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_studies_search_post(con)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Studies/StudiesSearch_POST.md}{github}
}
\seealso{
Other studies: \code{\link{ba_phenotypes_search}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_germplasm_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}}

Other phenotyping: \code{\link{ba_phenotypes_search}},
  \code{\link{ba_samples_save}},
  \code{\link{ba_samples_search_post}},
  \code{\link{ba_samples_search}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{phenotyping}
\concept{studies}
