% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_seasons.R
\name{ba_studies_seasons}
\alias{ba_studies_seasons}
\title{ba_studies_seasons}
\usage{
ba_studies_seasons(con = NULL, year = 0, page = 0, pageSize = 1000,
  rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{year}{integer}

\item{page}{integer requested page number}

\item{pageSize}{items per page}

\item{rclass}{character; one of tibble (default), json, list or data.frame}
}
\value{
data.frame
}
\description{
lists the studies seasons and years
}
\examples{
if (interactive()) {
  library(brapi)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  ba_studies_seasons(con)

}
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Studies/ListSeasons.md}
}
\seealso{
Other studies: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}}

Other phenotyping: \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon
}
