% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_studytypes.R
\name{ba_studies_studytypes}
\alias{ba_studies_studytypes}
\title{ba_studies_studytypes}
\usage{
ba_studies_studytypes(con = NULL, page = 0, pageSize = 1000,
  rclass = c("tibble", "data.frame", "list", "json"))
}
\arguments{
\item{con}{list, brapi connection object}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{rclass}{character, class of the object to be returned; default:
"tibble", possible other values: "data.frame"/"list"/"json"}
}
\value{
An object of class as defined by rclass containing the study types on
        the BrAPI compliant database server.
}
\description{
Retrieve study types available on a BrAPI compliant database server.
}
\note{
Tested against: test-server, sweetpotatobase

BrAPI Version: 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_studies_studytypes(con)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Studies/StudyTypes_GET.md}{github}
}
\seealso{
Other studies: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_germplasm_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}}

Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples_search}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{phenotyping}
\concept{studies}
