% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_table.R
\name{ba_studies_table}
\alias{ba_studies_table}
\title{ba_studies_table}
\usage{
ba_studies_table(con = NULL, studyDbId = "", format = "csv",
  rclass = "tibble")
}
\arguments{
\item{con}{list; brapi connection object}

\item{studyDbId}{character; default: ''}

\item{format}{character; one of: json, csv, tsv. Default: csv}

\item{rclass}{character; default: "tibble" possible other values: "json"/"list"/"data.frame"}
}
\value{
rclass as defined
}
\description{
lists studies_table available on a brapi server
}
\details{
This function must have set a specific study identifier. The default is an empty
string. If not changed to an study identifier present in the database this will
result in an error.
}
\note{
Tested against: sweetpotatobase, BMS
}
\examples{
if (interactive()) {
  library(brapi)

  # Need to connect to a database with genetic data

  # con <- ba_db()$xxxxx

  # ba_studies_table(con = con, ...)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Studies/StudyObservationUnitsAsTable.md}{github}
}
\seealso{
Other studies: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}}

Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon, Maikel Verouden
}
