% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_studies_table.R
\name{ba_studies_table}
\alias{ba_studies_table}
\title{ba_studies_table}
\usage{
ba_studies_table(con = NULL, studyDbId = "", format = c("csv", "tsv",
  "json"), rclass = c("tibble", "data.frame", "json"))
}
\arguments{
\item{con}{list, brapi connection object}

\item{studyDbId}{character, the internal database identifier for a study of
which the observation units, including actual trait data,
are to be retrieved e.g. "1001"; \strong{REQUIRED ARGUMENT}
with default: ""}

\item{format}{character, the format parameter will cause the data to be
dumped to a file in the specified format; default: "csv",
possible other values: "tsv"/"json"}

\item{rclass}{character, class of the object to be returned; default:
"tibble", possible other values: "data.frame"/"json"}
}
\value{
An object of class as defined by rclass containing the observation
        units, including trait data, for a requested study.
}
\description{
Retrieve observation units, including actual trait data, as table for a
specific study.
}
\details{
This function must have set a specific study identifier. The default
         is an empty string. If not changed to an study identifier present in
         the database this will result in an error.

When specifying the format argument, be aware that format = "csv" or
         "tsv" can only be combined with rclass = "tibble" or "data.frame".
         Using format = "json" allows for specifying the rclass = "json",
         "tibble, or "data.frame".

Choosing format = "json" could give unexpected results, because
         pagenation  (default: pageSize = 1000 and page = 0) is not
         implemented correctly on all databases and additionally the table
         for the specific study could have more than 1000 rows.
}
\note{
Tested against: BMS, test-server, sweetpotatobase

BrAPI Version: 1.0, 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$sweetpotatobase

  ba_studies_table(con, studyDbId = "1207")
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Studies/Studies_Table_GET.md}{github}
}
\seealso{
Other studies: \code{\link{ba_studies_details}},
  \code{\link{ba_studies_germplasm_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}}

Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples_search}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_trials_details}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{phenotyping}
\concept{studies}
