% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_traits.R
\name{ba_traits}
\alias{ba_traits}
\title{ba_traits}
\usage{
ba_traits(con = NULL, pageSize = 1000, page = 0,
  rclass = c("tibble", "data.frame", "list", "json"))
}
\arguments{
\item{con}{list, brapi connection object}

\item{pageSize}{integer, items per page to be returned; default: 1000}

\item{page}{integer, the requested page to be returned; default: 0 (1st page)}

\item{rclass}{character, class of the object to be returned; default:
"tibble", possible other values: "data.frame"/"list"/"json"}
}
\value{
An object of class as defined by rclass containing the traits and
        their associated variables on the BrAPI compliant database server.
}
\description{
Retrieve a list of traits available on a BrAPI compliant database server and
their associated variables.
}
\note{
Tested against: sweetpotatobase, test-server

BrAPI Version: 1.0, 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_traits(con)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Traits/ListAllTraits.md}{github}
}
\seealso{
Other traits: \code{\link{ba_traits_details}}

Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_commoncropnames}}, \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search_post}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search_post}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_studies_germplasm_details}},
  \code{\link{ba_traits_details}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{brapicore}
\concept{traits}
