% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_traits.R
\name{ba_traits}
\alias{ba_traits}
\title{ba_traits}
\usage{
ba_traits(con = NULL, page = 0, pageSize = 1000, rclass = "tibble")
}
\arguments{
\item{con}{list; brapi connection object}

\item{page}{integer; default 0}

\item{pageSize}{integer; defautlt 1000}

\item{rclass}{character; default: tibble}
}
\value{
rclass as defined
}
\description{
lists brapi_traits available on a brapi server
}
\note{
Tested against: sweetpotatobase
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_traits(con)
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Traits/ListAllTraits.md}{github}
}
\seealso{
Other traits: \code{\link{ba_traits_details}}

Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_commoncropnames}},
  \code{\link{ba_crops-deprecated}},
  \code{\link{ba_germplasm_details_study-deprecated}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}},
  \code{\link{ba_traits_details}}
}
\author{
Reinhard Simon
}
\concept{brapicore}
\concept{traits}
