% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_traits_details.R
\name{ba_traits_details}
\alias{ba_traits_details}
\title{ba_traits_details}
\usage{
ba_traits_details(con = NULL, traitDbId = "", rclass = c("tibble",
  "data.frame", "list", "json"))
}
\arguments{
\item{con}{list, brapi connection object}

\item{traitDbId}{character, the internal database identifier for a trait of
which the details and associated variables are to be retrieved
e.g. "1"; \strong{REQUIRED ARGUMENT} with default: ""}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "data.frame"/"list"/"json"}
}
\value{
An object of class as defined by rclass containing the details and
        associated variables of the requested trait identifier.
}
\description{
Retrieve the details and associated variables of specific trait identifier.
}
\note{
Tested against: sweetpotatobase, testserver

BrAPI Version: 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_traits_details(con, traitDbId = "1")
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Traits/TraitDetails.md}{github}
}
\seealso{
Other traits: \code{\link{ba_traits}}

Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_commoncropnames}}, \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_progeny}},
  \code{\link{ba_germplasm_search_post}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_phenotypes_search_post}},
  \code{\link{ba_programs_search_post}},
  \code{\link{ba_programs}},
  \code{\link{ba_studies_germplasm_details}},
  \code{\link{ba_traits}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{brapicore}
\concept{traits}
