% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_traits_details.R
\name{ba_traits_details}
\alias{ba_traits_details}
\title{ba_traits_details}
\usage{
ba_traits_details(con = NULL, traitDbId = "1", rclass = "tibble")
}
\arguments{
\item{con}{object; brapi connection object}

\item{traitDbId}{character; default 1}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists brapi_traits_details available on a brapi server
}
\examples{
if (interactive()) {
  library(brapi)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  ba_traits_details(con)

}
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Traits/TraitDetails.md}(github)
}
\seealso{
Other traits: \code{\link{ba_traits}}

Other brapicore: \code{\link{ba_calls}},
  \code{\link{ba_crops}},
  \code{\link{ba_germplasm_details_study}},
  \code{\link{ba_germplasm_details}},
  \code{\link{ba_germplasm_pedigree}},
  \code{\link{ba_germplasm_search}},
  \code{\link{ba_observationvariables_datatypes}},
  \code{\link{ba_observationvariables_details}},
  \code{\link{ba_observationvariables_ontologies}},
  \code{\link{ba_observationvariables_search}},
  \code{\link{ba_observationvariables}},
  \code{\link{ba_programs_search}},
  \code{\link{ba_programs}}, \code{\link{ba_traits}}
}
\author{
Reinhard Simon
}
