% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_trials.R
\name{ba_trials}
\alias{ba_trials}
\title{ba_trials}
\usage{
ba_trials(con = NULL, programDbId = "any", locationDbId = "any",
  active = TRUE, sortBy = "none", sortOrder = "asc", page = 0,
  pageSize = 10000, rclass = "tibble")
}
\arguments{
\item{con}{list, brapi connection object}

\item{programDbId}{character; default: 'any'; otherwise an identifier}

\item{locationDbId}{character; default: 'any'; otherwise an identifier}

\item{active}{logical; default: 'TRUE'; otherwise TRUE or FALSE}

\item{sortBy}{character; default: 'none'; otherwise a name from the first level entry names of the return object.}

\item{sortOrder}{character; default: 'asc'; otherwise 'desc'}

\item{page}{integer; default 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{character; default: tibble}
}
\value{
rclass as defined
}
\description{
lists trials available on a brapi server
}
\examples{
if (interactive()) {
  library(brapi)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  ba_trials(con)

}
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Trials/ListTrialSummaries.md}{github}
}
\seealso{
Other trials: \code{\link{ba_trials_details}}

Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}},
  \code{\link{ba_trials_details}}
}
\author{
Reinhard Simon
}
