% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_trials_details.R
\name{ba_trials_details}
\alias{ba_trials_details}
\title{ba_trials_details}
\usage{
ba_trials_details(con = NULL, trialDbId = "", rclass = "tibble")
}
\arguments{
\item{con}{list, brapi connection object}

\item{trialDbId}{character, the internal database identifier for a trial of
which the summary details are to be retreived; \strong{REQUIRED
ARGUMENT} with default: ""}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "json"/"list"/"data.frame"}
}
\value{
An object of class as defined by rclass containing the trial summary
        details.
}
\description{
List trial summary details for a given trial database identifier.
}
\note{
Tested against: sweetpotatobase, test-server

BrAPI Version: 1.0, 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_trials_details(con, trialDbId = "101")
}
}
\references{
\url{https://github.com/plantbreeding/API/blob/V1.2/Specification/Trials/GetTrialById.md}(github)
}
\seealso{
Other trials: \code{\link{ba_trials}}

Other phenotyping: \code{\link{ba_samples_save}},
  \code{\link{ba_samples_search}},
  \code{\link{ba_samples}},
  \code{\link{ba_studies_details}},
  \code{\link{ba_studies_layout}},
  \code{\link{ba_studies_observationlevels}},
  \code{\link{ba_studies_observations}},
  \code{\link{ba_studies_observationunits_save}},
  \code{\link{ba_studies_observationunits}},
  \code{\link{ba_studies_observationvariables}},
  \code{\link{ba_studies_search_get}},
  \code{\link{ba_studies_search_post}},
  \code{\link{ba_studies_search}},
  \code{\link{ba_studies_seasons}},
  \code{\link{ba_studies_studytypes}},
  \code{\link{ba_studies_table_save}},
  \code{\link{ba_studies_table}}, \code{\link{ba_trials}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{phenotyping}
\concept{trials}
