% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi-deprecated.R
\name{brapi-deprecated}
\alias{brapi-deprecated}
\title{Deprecated functions in brapi}
\description{
These functions still work but will be removed (defunct) in a future version.
}
\details{
\itemize{
 \item \code{\link{ba_crops}}: This function is deprecated, and will
 be removed in a future version of this package. Use \code{\link{ba_commoncropnames}} instead.
}
}
