% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_crops.R, R/ba_germplasm_details_study.R,
%   R/brapi-deprecated.R
\name{brapi-deprecated}
\alias{brapi-deprecated}
\alias{ba_crops}
\alias{ba_germplasm_details_study}
\title{Deprecated functions in brapi}
\usage{
ba_crops(con = NULL, rclass = "tibble")

ba_germplasm_details_study(con = NULL, studyDbId = "", page = 0,
  pageSize = 1000, rclass = "tibble")
}
\description{
These functions still work but will be removed (defunct) in a future version.
}
\details{
\itemize{
 \item \code{\link{ba_crops}}: This function is deprecated, and will
 be removed in a future version of this package. Use \code{\link{ba_commoncropnames}} instead.
}
\itemize{
 \item \code{\link{ba_germplasm_details_study}}: This function is deprecated, and will
 be removed in a future version of this package. Use \code{\link{ba_studies_germplasm_details}} instead.
}
}
\section{\code{ba_crops}}{

For \code{ba_crops}, use \code{\link{ba_commoncropnames}}.
}

\section{\code{ba_germplasm_details_study}}{

For \code{ba_germplasm_details_study}, use \code{\link{ba_studies_germplasm_details}}.
}

