% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRAPI_CON.R
\name{brapi_con}
\alias{brapi_con}
\title{brapi_con}
\usage{
brapi_con(crop, db, port = 80, user = "", pwd = "", session = "")
}
\arguments{
\item{crop}{character string}

\item{db}{character string (url)}

\item{port}{numeric (optional, default 80)}

\item{user}{character string (optional, default '')}

\item{pwd}{character string (optional, default '')}

\item{session}{character string (optional)}
}
\value{
object of class 'brapi'
}
\description{
S3 class constructor. It also places an object named 'brapi' in the parent environment.
This can be used later to refresh credentials using a 'brapi_auth' function call.
}
\examples{

# construct a sample object
library(brapi)
}
\author{
Reinahard Simon
}
\seealso{
Other brapi_con: \code{\link{brapi_auth}},
  \code{\link{is.brapi}}, \code{\link{print.brapi_con}}
}

