% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{calls}
\alias{calls}
\title{calls}
\usage{
calls(datatypes = "all")
}
\arguments{
\item{datatypes}{string, list of data types}
}
\value{
a data.frame
}
\description{
lists calls available on a brapi server
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/call-search}
}

