% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{calls}
\alias{calls}
\title{calls}
\usage{
calls(raw = FALSE)
}
\arguments{
\item{raw}{logical; default is FALSE; whether to display the raw json object or not}
}
\value{
a data.frame
}
\description{
lists calls available on a brapi server
}
\author{
Reinhard Simon
}

