% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{calls}
\alias{calls}
\title{calls}
\usage{
calls(con = NULL, datatypes = "all", pageSize = 1000, page = 0,
  rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{datatypes}{string, list of data types}

\item{pageSize}{integer; default = 1000}

\item{page}{integer; default = 0}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists calls available on a brapi server
}
\author{
Reinhard Simon
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Calls/Calls.md}{github}
}
\seealso{
Other brapi_core: \code{\link{crops}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{programs}}
}

