% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can_internet.R
\name{can_internet}
\alias{can_internet}
\title{can_internet}
\usage{
can_internet(url = "www.google.org")
}
\arguments{
\item{url}{a url to check (default \href{https://www.google.org}{www.google.org})}
}
\value{
logical TRUE when internet access is available or FALSE when internet access is not available
}
\description{
Check for internet connection.
}
\references{
\href{http://stackoverflow.com/questions/5076593/how-to-determine-if-you-have-an-internet-connection-in-r}{Stack Overflow} and  \code{\link[curl]{has_internet}}
}
