% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{chart}
\alias{chart}
\title{chart}
\usage{
chart(x, ...)
}
\arguments{
\item{x}{an object to be charted}

\item{...}{additional parameters}
}
\description{
A generic method to chart brapi objects
}
\author{
Reinhard Simon
}
\seealso{
Other brapiutils: \code{\link{describe}},
  \code{\link{is.brapi_con}}, \code{\link{is.brapi}},
  \code{\link{print.brapi_db_list}}
}

