% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{check}
\usage{
check(verbose = TRUE, brapi_calls = "any")
}
\arguments{
\item{verbose}{logical; default TRUE}

\item{brapi_calls}{character vector; default: any. Use to check if one or more calls are implemented by the server according to the calls url.}
}
\value{
logical
}
\description{
Checks if a BrAPI server can be reached given the connection details.
}
\details{
Raises errors.
}

