% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{connect}
\usage{
connect(brapi_db = NULL, crop = "sweetpotato", db = "127.0.0.1",
  port = 2021, protocol = "http://", multi = FALSE, user = "user",
  password = "password", session = "")
}
\arguments{
\item{brapi_db}{brapi_db R object for brapi databases. Created by as.brapi_db}

\item{crop}{string; default: sweetpotato}

\item{db}{string; default '127.0.0.1' (localhost)}

\item{port}{integer; default: 80}

\item{protocol}{string; default: http:// (else: https://)}

\item{multi}{logical; default: FALSE}

\item{user}{string; default: user}

\item{password}{string; default: password}

\item{session}{string; default: '' (empty)}
}
\value{
logical
}
\description{
Sets a global list variable 'brapi' that contains all parameters to connect to
a BrAPI compliant online database.
}
\details{
The default values are chosen to match a local mock test server. The mock server
must be started in a separate R session using 'brapi::mock_server()'.

If the brapi_db parameter is used it will override the other parameters; except user, password and session.
}
\examples{
if(interactive()){
  if(can_internet("http://127.0.0.1:2021/brapi/v1")){

    connect()

    calls()

    crops()

  }
}
}
\seealso{
Other access: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{studyTypes}}
}

