% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{connect}
\usage{
connect(brapiDb = NULL, secure = FALSE, protocol = "http://",
  db = "127.0.0.1", port = 2021, apipath = NULL, multicrop = FALSE,
  crop = "sweetpotato", user = "user", password = "password",
  token = "", granttype = "password", clientid = "rbrapi", bms = FALSE)
}
\arguments{
\item{brapiDb}{brapiDb R object for brapi databases. Created by as.brapi_db}

\item{secure}{logical TRUE/FALSE}

\item{protocol}{string; default: http:// (else: https://)}

\item{db}{string; default '127.0.0.1' (localhost)}

\item{port}{integer; default: 80}

\item{apipath}{character; default NULL}

\item{multicrop}{logical; default FALSE}

\item{crop}{string; default: ''}

\item{user}{string; default: user}

\item{password}{string; default: password}

\item{token}{string; default: '' (empty)}

\item{granttype}{string}

\item{clientid}{string: default: rbrapi}

\item{bms}{logical; default: FALSE}
}
\value{
connection object
}
\description{
Sets a global list variable 'brapi' that contains all parameters to connect to
a BrAPI compliant online database.
}
\details{
The default values are chosen to match a local mock test server. The mock server
must be started in a separate R session using 'brapi::mock_server()'.

If the brapi_db parameter is used it will override the other parameters; except user, password and session.
}
\examples{
if(interactive()){
  if(can_internet("http://127.0.0.1:2021/brapi/v1")){

    connect()

    calls()

    crops()

  }
}
}

