% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crops.R
\name{crops}
\alias{crops}
\title{crops}
\usage{
crops(con = NULL, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{rclass}{logical; default is FALSE; whether to display the raw list object or not}
}
\value{
a vector of crop names or NULL
}
\description{
lists crops available in a database
}
\author{
Reinhard Simon
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Crops/ListCrops.md}(github)
}
\seealso{
Other core: \code{\link{calls}},
  \code{\link{germplasm_details_study}},
  \code{\link{germplasm_details}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}},
  \code{\link{observationvariables_datatypes}},
  \code{\link{observationvariables_details}},
  \code{\link{observationvariables_ontologies}},
  \code{\link{observationvariables}},
  \code{\link{programs}}, \code{\link{traits_details}},
  \code{\link{traits}}
}

