% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crops.R
\name{crops}
\alias{crops}
\title{crops}
\usage{
crops(con = NULL, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{rclass}{logical; default is FALSE; whether to display the raw list object or not}
}
\value{
a vector of crop names or NULL
}
\description{
lists crops available in a database
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Crops/ListCrops.md}{github}
}
\seealso{
Other brapi_core: \code{\link{calls}},
  \code{\link{programs}}
}
\author{
Reinhard Simon
}
