% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{describe}
\alias{describe}
\title{describe}
\usage{
describe(x, ...)
}
\arguments{
\item{x}{an object to be described}

\item{...}{additional parameters}
}
\description{
A generic method to sum up brapi objects
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)

  # make sure brapiTS::mock_server() is running in a separate process

  con <- ba_connect()

  lcs <- ba_locations(con)
  lcs \%>\% describe
}
}
\seealso{
Other brapiutils: \code{\link{chart}},
  \code{\link{is.ba_con}}, \code{\link{print.ba_db_list}}
}
\author{
Reinhard Simon
}
