% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm.R
\name{germplasm}
\alias{germplasm}
\title{germplasm}
\usage{
germplasm(germplasmDbId = 0)
}
\arguments{
\item{germplasmDbId}{string; default 0; an internal ID for a germplasm}
}
\value{
list
}
\description{
Gets germplasm details given an id.
}
\author{
Reinhard Simon
}

