% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm.R
\name{germplasm}
\alias{germplasm}
\title{germplasm}
\usage{
germplasm(germplasmDbId = 0, rclass = "tibble")
}
\arguments{
\item{germplasmDbId}{string; default 0; an internal ID for a germplasm}

\item{rclass}{character; tibble}
}
\value{
list
}
\description{
Gets germplasm details given an id.
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/germplasm-details-by-germplasmdbid}
}
\seealso{
Other brapi_call: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}}, \code{\link{calls}},
  \code{\link{crops}}, \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{locations}},
  \code{\link{maps_details}},
  \code{\link{maps_positions_range}},
  \code{\link{maps_positions}}, \code{\link{maps}},
  \code{\link{marker_search}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}},
  \code{\link{observationLevels}}, \code{\link{programs}},
  \code{\link{seasons}}, \code{\link{studies_search}},
  \code{\link{studyTypes}}, \code{\link{trial_by_id}},
  \code{\link{trials}}

Other core: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}},
  \code{\link{marker_search}}, \code{\link{programs}},
  \code{\link{seasons}}

Other germplasm: \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}},
  \code{\link{marker_search}}
}

