% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm_attributes.R
\name{germplasm_attributes}
\alias{germplasm_attributes}
\title{germplasm_germplasm_attributes}
\usage{
germplasm_attributes(germplasmDbId = 1, attributeList = 1, page = 0,
  pageSize = 10, rclass = "tibble")
}
\arguments{
\item{germplasmDbId}{integer; default: 1.}

\item{attributeList}{integer vector; default: 1.}

\item{page}{integer; default: 0.}

\item{pageSize}{integer; default: 10.}

\item{rclass}{character; default: tibble; or else: json, list, data.frame.}
}
\value{
rclass as set by parameter.
}
\description{
attibutes call.
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/germplasm-attribute-values-by-germplasmdbid/germplasm-attribute-values}
}
\seealso{
Other attributes: \code{\link{attributes_categories}},
  \code{\link{attributes}},
  \code{\link{germplasm_markerprofiles}}

Other brapi_call: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}}, \code{\link{calls}},
  \code{\link{crops}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{locations}}, \code{\link{maps}},
  \code{\link{marker_search}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}},
  \code{\link{programs}}, \code{\link{seasons}}

Other genotyping: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{maps}}, \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}}
}

