% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm_details.R
\name{germplasm_details}
\alias{germplasm_details}
\title{germplasm_details}
\usage{
germplasm_details(con = NULL, germplasmDbId = 0, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{germplasmDbId}{string; default 0; an internal ID for a germplasm}

\item{rclass}{character; tibble}
}
\value{
list
}
\description{
Gets germplasm details given an id.
}
\author{
Reinhard Simon
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Germplasm/GermplasmDetailsByGermplasmDbId.md}
}
\seealso{
Other brapicore: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_details_study}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}},
  \code{\link{observationvariables_datatypes}},
  \code{\link{observationvariables_details}},
  \code{\link{observationvariables_ontologies}},
  \code{\link{observationvariables}},
  \code{\link{programs}}, \code{\link{traits_details}},
  \code{\link{traits}}

Other germplasm: \code{\link{germplasm_details_study}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}
}

