% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm_markerprofiles.R
\name{germplasm_markerprofiles}
\alias{germplasm_markerprofiles}
\title{germplasm_markerprofiles}
\usage{
germplasm_markerprofiles(con = NULL, germplasmDbId = 0, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{germplasmDbId}{integer}

\item{rclass}{character, default: list; alternative: vector}
}
\value{
list of marker profile ids
}
\description{
Gets minimal marker profile data from database using database internal id
}
\author{
Reinhard Simon
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Germplasm/GermplasmMarkerprofiles.md}
}
\seealso{
Other genotyping: \code{\link{genomemaps_data_range}},
  \code{\link{genomemaps_data}},
  \code{\link{genomemaps_details}},
  \code{\link{genomemaps}},
  \code{\link{germplasmattributes_categories}},
  \code{\link{germplasmattributes_details}},
  \code{\link{germplasmattributes}},
  \code{\link{markerprofiles_allelematrix_search}},
  \code{\link{markerprofiles_details}},
  \code{\link{markerprofiles_search}},
  \code{\link{markers_details}},
  \code{\link{markers_search}}

Other germplasm: \code{\link{germplasm_details_study}},
  \code{\link{germplasm_details}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}
}

