% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm_markerprofiles.R
\name{germplasm_markerprofiles}
\alias{germplasm_markerprofiles}
\title{germplasm_markerprofiles}
\usage{
germplasm_markerprofiles(germplasmDbId = 0, rclass = "tibble")
}
\arguments{
\item{germplasmDbId}{integer}

\item{rclass}{character, default: list; alternative: vector}
}
\value{
list of marker profile ids
}
\description{
Gets minimal marker profile data from database using database internal id
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/germplasm-markerprofiles}
}
\seealso{
Other attributes: \code{\link{attributes_categories}},
  \code{\link{attributes}},
  \code{\link{germplasm_attributes}}

Other brapi_call: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}}, \code{\link{calls}},
  \code{\link{crops}}, \code{\link{germplasm_attributes}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{locations}}, \code{\link{maps}},
  \code{\link{marker_search}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}},
  \code{\link{programs}}, \code{\link{seasons}}

Other genotyping: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}},
  \code{\link{germplasm_attributes}}, \code{\link{maps}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}}
}

