% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm_pedigree.R
\name{germplasm_pedigree}
\alias{germplasm_pedigree}
\title{germplasm_pedigree}
\usage{
germplasm_pedigree(con = NULL, germplasmDbId = 0, notation = "purdue",
  rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{germplasmDbId}{integer; default 0}

\item{notation}{string; default: purdue format}

\item{rclass}{string; default: tibble}
}
\value{
list of pedigree data
}
\description{
Gets minimal pedigree data from database using database internal id
}
\author{
Reinhard Simon
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Germplasm/GermplasmPedigree.md}(github)
}
\seealso{
Other brapi_core: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_details_study}},
  \code{\link{germplasm_details}},
  \code{\link{germplasm_search}},
  \code{\link{observationvariables_datatypes}},
  \code{\link{observationvariables_details}},
  \code{\link{observationvariables_ontologies}},
  \code{\link{observationvariables}},
  \code{\link{programs}}, \code{\link{traits_details}},
  \code{\link{traits}}

Other germplasm: \code{\link{germplasm_details_study}},
  \code{\link{germplasm_details}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_search}}
}

