% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm_pedigree.R
\name{germplasm_pedigree}
\alias{germplasm_pedigree}
\title{germplasm_pedigree}
\usage{
germplasm_pedigree(id = NULL)
}
\arguments{
\item{id}{integer}
}
\value{
list of pedigree data
}
\description{
Gets minimal pedigree data from database using database internal id
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/germplasm/germplasm-details-list-by-studydbid/germplasm-pedigree-by-id?console=1}
}

