% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm_search.R
\name{germplasm_search}
\alias{germplasm_search}
\title{germplasm search}
\usage{
germplasm_search(name = "*", matchMethod = "wildcard",
  include = "synonyms", pageSize = 10, page = 1)
}
\arguments{
\item{name}{string}

\item{matchMethod}{defaults to 'wildcard'.}

\item{include}{synonyms of germplasm}

\item{pageSize}{integer}

\item{page}{integer}
}
\value{
data.frame
}
\description{
Lists germplasmm as result of a search.
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/reference/germplasm/germplasm-search/search-names-to-retrieve-germplasm-records}
}

