% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm_search.R
\name{germplasm_search}
\alias{germplasm_search}
\title{germplasm search}
\usage{
germplasm_search(germplasmDbId = 0, germplasmName = "none",
  germplasmPUI = "none", page = 0, pageSize = 1000, method = "GET",
  rclass = "tibble")
}
\arguments{
\item{germplasmDbId}{string}

\item{germplasmName}{string}

\item{germplasmPUI}{string}

\item{page}{integer}

\item{pageSize}{integer}

\item{method}{string; default 'GET'; alternative 'POST'}

\item{rclass}{character; default: tibble}
}
\value{
data.frame
}
\description{
Lists germplasmm as result of a search.
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/germplasm-search}
}

