% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm_search.R
\name{germplasm_search}
\alias{germplasm_search}
\title{germplasm_search}
\usage{
germplasm_search(con = NULL, germplasmDbId = 0, germplasmName = "none",
  germplasmPUI = "none", page = 0, pageSize = 1000, method = "GET",
  rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{germplasmDbId}{string}

\item{germplasmName}{string}

\item{germplasmPUI}{string}

\item{page}{integer}

\item{pageSize}{integer}

\item{method}{string; default 'GET'; alternative 'POST'}

\item{rclass}{character; default: tibble}
}
\value{
tibble
}
\description{
Lists germplasmm as result of a search.
}
\author{
Reinhard Simon
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Germplasm/GermplasmSearchGET.md}
   \url{https://github.com/plantbreeding/API/blob/master/Specification/Germplasm/GermplasmSearchPOST.md}
}
\seealso{
Other core: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_details_study}},
  \code{\link{germplasm_details}},
  \code{\link{germplasm_pedigree}},
  \code{\link{observationvariables_datatypes}},
  \code{\link{observationvariables_details}},
  \code{\link{observationvariables_ontologies}},
  \code{\link{observationvariables}},
  \code{\link{programs}}, \code{\link{traits_details}},
  \code{\link{traits}}

Other germplasm: \code{\link{germplasm_details_study}},
  \code{\link{germplasm_details}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}}
}

