% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm_search.R
\name{germplasm_search}
\alias{germplasm_search}
\title{germplasm search}
\usage{
germplasm_search(germplasmDbId = 0, germplasmName = "none",
  germplasmPUI = "none", page = 0, pageSize = 1000, method = "GET",
  rclass = "tibble")
}
\arguments{
\item{germplasmDbId}{string}

\item{germplasmName}{string}

\item{germplasmPUI}{string}

\item{page}{integer}

\item{pageSize}{integer}

\item{method}{string; default 'GET'; alternative 'POST'}

\item{rclass}{character; default: tibble}
}
\value{
data.frame
}
\description{
Lists germplasmm as result of a search.
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/germplasm-search}
}
\seealso{
Other brapi_call: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}}, \code{\link{calls}},
  \code{\link{crops}}, \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm}}, \code{\link{locations}},
  \code{\link{maps_details}},
  \code{\link{maps_positions_range}},
  \code{\link{maps_positions}}, \code{\link{maps}},
  \code{\link{marker_search}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}},
  \code{\link{observationLevels}}, \code{\link{programs}},
  \code{\link{seasons}}, \code{\link{studyTypes}},
  \code{\link{trial_by_id}}, \code{\link{trials}}

Other core: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm}}, \code{\link{marker_search}},
  \code{\link{programs}}, \code{\link{seasons}}

Other germplasm: \code{\link{germplasm_pedigree}},
  \code{\link{germplasm}}, \code{\link{marker_search}}
}

