% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.brapi.R
\name{is.brapi}
\alias{is.brapi}
\title{is.brapi}
\usage{
is.brapi(con)
}
\arguments{
\item{con}{brapi connection object}
}
\value{
logical
}
\description{
check if the connection object is correct
}
\seealso{
Other brapiutils: \code{\link{chart}},
  \code{\link{describe}}, \code{\link{is.brapi_con}},
  \code{\link{print.brapi_db_list}}
}
