% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{locations}
\alias{locations}
\title{locations}
\usage{
locations(con = NULL, locationType = "all", page = 0, pageSize = 1e+06,
  rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{locationType}{string, list of data types}

\item{page}{integer; default 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists locations available on a brapi server
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Locations/ListLocations.md}(github)
}
\author{
Reinhard Simon
}
