% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{locations}
\alias{locations}
\title{locations}
\usage{
locations(input, output, session)
}
\arguments{
\item{input}{shiny}

\item{output}{shiyn}

\item{session}{shiny}
}
\value{
data.frame
}
\description{
locations
}
\author{
Reinhard Simon
}

