% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_search.R
\name{marker_search}
\alias{marker_search}
\title{marker search}
\usage{
marker_search(name = "none", type = "all", matchMethod = "exact",
  include = "synonyms", page = 0, pageSize = 1000, rclass = "tibble")
}
\arguments{
\item{name}{string; marker name; default: none}

\item{type}{string; default: all; other: SNP}

\item{matchMethod}{string; default: exact; other: case_insensitive, wildcard}

\item{include}{string; default: synonyms}

\item{page}{integer; default: 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{character; default: tibble}
}
\value{
data.frame
}
\description{
Lists markers as result of a search.
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/germplasm-search}
}
\seealso{
Other brapi_call: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}}, \code{\link{calls}},
  \code{\link{crops}}, \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{locations}}, \code{\link{maps_positions}},
  \code{\link{maps}}, \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}},
  \code{\link{programs}}, \code{\link{seasons}}

Other core: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{programs}}, \code{\link{seasons}}

Other germplasm: \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}}
}

