% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observationLevels.R
\name{observationLevels}
\alias{observationLevels}
\title{observationLevels}
\usage{
observationLevels(rclass = "vector")
}
\arguments{
\item{rclass}{character; default is FALSE; whether to display the raw list object or not}
}
\value{
a vector of crop names or NULL
}
\description{
lists crops available in a database
}
\author{
Reinhard Simon
}
\seealso{
Other access: \code{\link{calls}}, \code{\link{connect}},
  \code{\link{crops}}

Other brapi_call: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}}, \code{\link{calls}},
  \code{\link{crops}}, \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{locations}}, \code{\link{maps_details}},
  \code{\link{maps_positions_range}},
  \code{\link{maps_positions}}, \code{\link{maps}},
  \code{\link{marker_search}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}},
  \code{\link{programs}}, \code{\link{seasons}}

Other core: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{marker_search}}, \code{\link{programs}},
  \code{\link{seasons}}
}

