% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.ba_con}
\alias{print.ba_con}
\title{print.ba_con}
\usage{
\method{print}{ba_con}(x, ...)
}
\arguments{
\item{x}{a brapi_con object}

\item{...}{other print parameters}
}
\description{
print method for an object of class brapi_con, which will only display the crop,  database address:port and user
}
\seealso{
Other brapiutils: \code{\link{as.ba_db}},
  \code{\link{ba_can_internet}}, \code{\link{ba_check}},
  \code{\link{ba_connect}}, \code{\link{ba_db}},
  \code{\link{ba_login}}, \code{\link{ba_logout}},
  \code{\link{is.ba_con}}, \code{\link{print.ba_db_list}}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{brapiutils}
