% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.ba_db_list}
\alias{print.ba_db_list}
\title{print.ba_db_list}
\usage{
\method{print}{ba_db_list}(x, ...)
}
\arguments{
\item{x}{a ba_db_list object}

\item{...}{other print parameters}
}
\description{
print method for an object of class ba_db_list
}
\seealso{
Other brapiutils: \code{\link{as.ba_db}},
  \code{\link{ba_authenticate}},
  \code{\link{ba_can_internet}},
  \code{\link{ba_chart.ba_genomemaps_details}},
  \code{\link{ba_chart.ba_genomemaps}},
  \code{\link{ba_chart.ba_locations}},
  \code{\link{ba_chart}}, \code{\link{ba_check}},
  \code{\link{ba_connect}}, \code{\link{ba_db}},
  \code{\link{ba_describe.ba_locations}},
  \code{\link{ba_describe}}, \code{\link{is.ba_con}},
  \code{\link{print.ba_con}}
}
\author{
Reinhard Simon, Maikel Verouden
}
