% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRAPI_CON.R
\name{print.brapi_con}
\alias{print.brapi_con}
\title{print.brapi_con}
\usage{
\method{print}{brapi_con}(x, ...)
}
\arguments{
\item{x}{brapi object}

\item{...}{other print parameters}
}
\description{
print method
}
\examples{

# construct a sample object
library(brapi)

brapi = brapi_con("potato", "http://sample.com", 8080, "rs", "pwd")

is.brapi(brapi)

brapi
}
\author{
Reinahard Simon
}
\seealso{
Other brapi_con: \code{\link{brapi_auth}},
  \code{\link{brapi_con}}, \code{\link{is.brapi}}
}

