% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{programs}
\alias{programs}
\title{programs}
\usage{
programs(page = 0, pageSize = 100)
}
\arguments{
\item{page}{integer requested page number, default = 0 (1st page)}

\item{pageSize}{items per page (default = 100)}
}
\value{
data.frame
}
\description{
lists the breeding programs
}
\details{
BRAPI discussion: Should this return also the crop?
}
\author{
Reinhard Simon
}

