% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{programs}
\alias{programs}
\title{programs}
\usage{
programs(page = 0, pageSize = 100, rclass = "tibble")
}
\arguments{
\item{page}{integer requested page number, default = 0 (1st page)}

\item{pageSize}{items per page (default = 100)}

\item{rclass}{string; default: tibble}
}
\value{
rclass
}
\description{
lists the breeding programs
}
\details{
BRAPI discussion: Should this return also the crop?
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/program-list/list-programs}
}

