% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{programs}
\alias{programs}
\title{programs}
\usage{
programs(con = NULL, page = 0, pageSize = 100, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{page}{integer requested page number, default = 0 (1st page)}

\item{pageSize}{items per page (default = 100)}

\item{rclass}{string; default: tibble}
}
\value{
rclass
}
\description{
lists the breeding programs
}
\details{
BRAPI discussion: Should this return also the crop?
}
\author{
Reinhard Simon
}
\references{
\href{https://github.com/plantbreeding/API/blob/master/Specification/Programs/ListPrograms.md}{github}
}
\seealso{
Other brapi_core: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_details_study}},
  \code{\link{germplasm_details}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}},
  \code{\link{observationvariables_datatypes}},
  \code{\link{observationvariables_details}},
  \code{\link{observationvariables_ontologies}},
  \code{\link{observationvariables}},
  \code{\link{traits_details}}, \code{\link{traits}}
}

