% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{programs}
\alias{programs}
\title{programs}
\usage{
programs(page = 0, pageSize = 100, rclass = "tibble")
}
\arguments{
\item{page}{integer requested page number, default = 0 (1st page)}

\item{pageSize}{items per page (default = 100)}

\item{rclass}{string; default: tibble}
}
\value{
rclass
}
\description{
lists the breeding programs
}
\details{
BRAPI discussion: Should this return also the crop?
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/program-list/list-programs}
}
\seealso{
Other brapi_call: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}}, \code{\link{calls}},
  \code{\link{crops}}, \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{locations}}, \code{\link{maps_details}},
  \code{\link{maps_positions_range}},
  \code{\link{maps_positions}}, \code{\link{maps}},
  \code{\link{marker_search}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}},
  \code{\link{observationLevels}}, \code{\link{seasons}},
  \code{\link{studies_germplasm}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationVariables}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_search}}, \code{\link{studies}},
  \code{\link{studyTypes}}, \code{\link{trial_by_id}},
  \code{\link{trials}}

Other core: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{marker_search}}, \code{\link{seasons}},
  \code{\link{studies}}

Other experiments: \code{\link{seasons}}
}

