% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{seasons}
\alias{seasons}
\title{seasons}
\usage{
seasons(year = NULL, page = 0, pageSize = 5, format = "json",
  progress = FALSE)
}
\arguments{
\item{year}{integer}

\item{page}{integer requested page number}

\item{pageSize}{items per page}

\item{format}{character; one of json (default) or data.frame}

\item{progress}{logical default is FALSE}
}
\value{
data.frame
}
\description{
lists the seasons and years
}
\author{
Reinhard Simon
}

