% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{seasons}
\alias{seasons}
\title{seasons}
\usage{
seasons(year = NULL, page = 0, pageSize = 1000, rclass = "tibble")
}
\arguments{
\item{year}{integer}

\item{page}{integer requested page number}

\item{pageSize}{items per page}

\item{rclass}{character; one of tibble (default), json, list or data.frame}
}
\value{
data.frame
}
\description{
lists the seasons and years
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/list-seasons/list-seasons-or-years}
}
\seealso{
Other brapi_call: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}}, \code{\link{calls}},
  \code{\link{crops}}, \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{locations}}, \code{\link{maps_details}},
  \code{\link{maps_positions_range}},
  \code{\link{maps_positions}}, \code{\link{maps}},
  \code{\link{marker_search}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}},
  \code{\link{programs}}

Other core: \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{marker_search}}, \code{\link{programs}}

Other experiments: \code{\link{programs}}
}

