% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\name{studies}
\alias{studies}
\title{studies}
\usage{
studies(programId = NULL, page = 1, pageSize = 100)
}
\arguments{
\item{programId}{integer}

\item{page}{integer requested page number}

\item{pageSize}{items per page}
}
\value{
list of marker profile ids
}
\description{
Gets list of studies in a breeding program
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/study/list-studies/list-of-study-summaries?console=1}
}

