% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies_search.R
\name{studies_search}
\alias{studies_search}
\title{studies_search}
\usage{
studies_search(con = NULL, studyType = "any", programDbId = "any",
  locationDbId = "any", seasonDbId = "any", germplasmDbIds = "any",
  observationVariableDbIds = "any", active = "any", sortBy = "any",
  sortOrder = "any", page = 0, pageSize = 1000, rclass = "tibble")
}
\arguments{
\item{con}{brapi connection object}

\item{studyType}{string; default: any}

\item{programDbId}{string; default: any}

\item{locationDbId}{string; default: any}

\item{seasonDbId}{string; default: any}

\item{germplasmDbIds}{string; default: any}

\item{observationVariableDbIds}{string; default: any}

\item{active}{string; default: any}

\item{sortBy}{string; default: any}

\item{sortOrder}{string; default: any}

\item{page}{integer; default: 1000}

\item{pageSize}{integer; default: 0}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists trials available on a brapi server
}
\author{
Reinhard Simon
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Studies/ListStudySummaries.md}(github)
}
\seealso{
Other phenotyping: \code{\link{studies_details}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationlevels}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_observationvariables}},
  \code{\link{studies_seasons}},
  \code{\link{studies_studytypes}},
  \code{\link{studies_table}},
  \code{\link{trials_details}}, \code{\link{trials}}

Other studies: \code{\link{studies_details}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationlevels}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_observationvariables}},
  \code{\link{studies_seasons}},
  \code{\link{studies_studytypes}},
  \code{\link{studies_table}}
}

