% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyTypes.R
\name{studyTypes}
\alias{studyTypes}
\title{studyTypes}
\usage{
studyTypes(page = 0, pageSize = 1000, rclass = "tibble")
}
\arguments{
\item{page}{integer; default 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists studyTypes available on a brapi server
}
\author{
Reinhard Simon
}
\references{
\url{http://docs.brapi.apiary.io/#reference/0/studyTypes}
}
\seealso{
Other access: \code{\link{calls}}, \code{\link{connect}},
  \code{\link{crops}}

Other brapi_call: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}}, \code{\link{calls}},
  \code{\link{crops}}, \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{locations}}, \code{\link{maps_details}},
  \code{\link{maps_positions_range}},
  \code{\link{maps_positions}}, \code{\link{maps}},
  \code{\link{marker_search}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}},
  \code{\link{observationLevels}}, \code{\link{programs}},
  \code{\link{seasons}}, \code{\link{studies_layout}},
  \code{\link{studies_search}}, \code{\link{studies}},
  \code{\link{trial_by_id}}, \code{\link{trials}}

Other phenotyping: \code{\link{locations}},
  \code{\link{observationLevels}},
  \code{\link{studies_layout}},
  \code{\link{studies_search}}, \code{\link{trial_by_id}},
  \code{\link{trials}}
}

