% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trials.R
\name{trials}
\alias{trials}
\title{trials}
\usage{
trials(programDbId = "any", locationDbId = "any", active = "any",
  sortBy = "none", sortOrder = "asc", page = 0, pageSize = 1000,
  rclass = "tibble")
}
\arguments{
\item{programDbId}{character; default: 'any'; otherwise an identifier}

\item{locationDbId}{character; default: 'any'; otherwise an identifier}

\item{active}{character; default: 'any'; otherwise TRUE or FALSE}

\item{sortBy}{character; default: 'none'; otherwise a name from the first level entry names of the return object.}

\item{sortOrder}{character; default: 'asc'; otherwise 'desc'}

\item{page}{integer; default 0}

\item{pageSize}{integer; default 1000}

\item{rclass}{string; default: tibble}
}
\value{
rclass as defined
}
\description{
lists trials available on a brapi server
}
\author{
Reinhard Simon
}
\references{
\url{https://github.com/plantbreeding/API/blob/master/Specification/Trials/ListTrialSummaries.md}
}
\seealso{
Other brapi_call: \code{\link{allelematrix_search}},
  \code{\link{attributes_categories}},
  \code{\link{attributes}},
  \code{\link{brapi_studies_table}},
  \code{\link{brapi_variables_datatypes}},
  \code{\link{brapi_variables_ontologies}},
  \code{\link{calls}}, \code{\link{crops}},
  \code{\link{germplasm_attributes}},
  \code{\link{germplasm_markerprofiles}},
  \code{\link{germplasm_pedigree}},
  \code{\link{germplasm_search}}, \code{\link{germplasm}},
  \code{\link{locations}}, \code{\link{maps_details}},
  \code{\link{maps_positions_range}},
  \code{\link{maps_positions}}, \code{\link{maps}},
  \code{\link{marker_search}},
  \code{\link{markerprofiles_alleles}},
  \code{\link{markerprofiles}}, \code{\link{markers}},
  \code{\link{observationLevels}}, \code{\link{programs}},
  \code{\link{seasons}}, \code{\link{studies_germplasm}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationVariables}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_search}}, \code{\link{studies}},
  \code{\link{studyTypes}}, \code{\link{trial_by_id}}

Other phenotyping: \code{\link{brapi_studies_table}},
  \code{\link{brapi_traits_details}},
  \code{\link{brapi_traits}}, \code{\link{locations}},
  \code{\link{observationLevels}},
  \code{\link{studies_germplasm}},
  \code{\link{studies_layout}},
  \code{\link{studies_observationVariables}},
  \code{\link{studies_observations}},
  \code{\link{studies_observationunits}},
  \code{\link{studies_search}}, \code{\link{studyTypes}},
  \code{\link{trial_by_id}}
}

